Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1 -0.819178457
#> 2    X2  0.284406089
#> 3    X3  0.420325403
#> 4    X4 -0.131875619
#> 5    X5  0.289195352
#> 6    X1 -0.482357549
#> 7    X2 -1.555643898
#> 8    X3  0.422594129
#> 9    X4 -0.947345487
#> 10   X5  0.731988304
#> 11   X1 -0.668413155
#> 12   X2  0.666548960
#> 13   X3 -0.995032427
#> 14   X4  0.304981467
#> 15   X5 -0.323455543
#> 16   X1 -0.347084772
#> 17   X2 -2.906394314
#> 18   X3 -1.660008612
#> 19   X4  0.602928034
#> 20   X5  0.784139539
#> 21   X1 -0.443199672
#> 22   X2 -0.420348757
#> 23   X3 -1.121277573
#> 24   X4 -1.128065345
#> 25   X5  0.137003217
#> 26   X1 -0.251807694
#> 27   X2 -0.050477342
#> 28   X3  1.413138867
#> 29   X4 -0.279570667
#> 30   X5  0.428257113
#> 31   X1  0.273598044
#> 32   X2 -0.838354273
#> 33   X3 -2.664566691
#> 34   X4 -0.242345287
#> 35   X5  0.003608266
#> 36   X1  0.179367839
#> 37   X2 -1.808835129
#> 38   X3  0.750797898
#> 39   X4 -0.878507774
#> 40   X5  2.442091670
#> 41   X1 -0.362022970
#> 42   X2  0.653394427
#> 43   X3  1.354317828
#> 44   X4  0.748391223
#> 45   X5  0.227829440
#> 46   X1  0.211322899
#> 47   X2  0.591293794
#> 48   X3  1.290788652
#> 49   X4 -0.156000088
#> 50   X5 -1.379864037

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4           X5 Row Column     Numbers
#> 1   0.4203254 -0.1318756  0.289195352   1     X1 -0.81917846
#> 2   0.4203254 -0.1318756  0.289195352   1     X2  0.28440609
#> 3   0.4225941 -0.9473455  0.731988304   2     X1 -0.48235755
#> 4   0.4225941 -0.9473455  0.731988304   2     X2 -1.55564390
#> 5  -0.9950324  0.3049815 -0.323455543   3     X1 -0.66841315
#> 6  -0.9950324  0.3049815 -0.323455543   3     X2  0.66654896
#> 7  -1.6600086  0.6029280  0.784139539   4     X1 -0.34708477
#> 8  -1.6600086  0.6029280  0.784139539   4     X2 -2.90639431
#> 9  -1.1212776 -1.1280653  0.137003217   5     X1 -0.44319967
#> 10 -1.1212776 -1.1280653  0.137003217   5     X2 -0.42034876
#> 11  1.4131389 -0.2795707  0.428257113   6     X1 -0.25180769
#> 12  1.4131389 -0.2795707  0.428257113   6     X2 -0.05047734
#> 13 -2.6645667 -0.2423453  0.003608266   7     X1  0.27359804
#> 14 -2.6645667 -0.2423453  0.003608266   7     X2 -0.83835427
#> 15  0.7507979 -0.8785078  2.442091670   8     X1  0.17936784
#> 16  0.7507979 -0.8785078  2.442091670   8     X2 -1.80883513
#> 17  1.3543178  0.7483912  0.227829440   9     X1 -0.36202297
#> 18  1.3543178  0.7483912  0.227829440   9     X2  0.65339443
#> 19  1.2907887 -0.1560001 -1.379864037  10     X1  0.21132290
#> 20  1.2907887 -0.1560001 -1.379864037  10     X2  0.59129379

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2   Value_X3   Value_X4     Value_X5
#> 1       1 -0.8191785  0.28440609  0.4203254 -0.1318756  0.289195352
#> 2       2 -0.4823575 -1.55564390  0.4225941 -0.9473455  0.731988304
#> 3       3 -0.6684132  0.66654896 -0.9950324  0.3049815 -0.323455543
#> 4       4 -0.3470848 -2.90639431 -1.6600086  0.6029280  0.784139539
#> 5       5 -0.4431997 -0.42034876 -1.1212776 -1.1280653  0.137003217
#> 6       6 -0.2518077 -0.05047734  1.4131389 -0.2795707  0.428257113
#> 7       7  0.2735980 -0.83835427 -2.6645667 -0.2423453  0.003608266
#> 8       8  0.1793678 -1.80883513  0.7507979 -0.8785078  2.442091670
#> 9       9 -0.3620230  0.65339443  1.3543178  0.7483912  0.227829440
#> 10     10  0.2113229  0.59129379  1.2907887 -0.1560001 -1.379864037

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5