download_model {insight}	R Documentation

Download circus models

Description

Downloads pre-compiled models from the circus-repository. The circus-repository contains a variety of fitted models to help the systematic testing of other packages

Usage

download_model(name, url = NULL)

Arguments

name
Model name.

url
String with the URL from where to download the model data. Optional, and should only be used in case the repository-URL is changing. By default, models are downloaded from https://raw.github.com/easystats/circus/master/data/.

Details

The code that generated the model is available at the https://easystats.github.io/circus/reference/index.html.

Value

A model from the circus-repository.

References

https://easystats.github.io/circus/

[Package insight version 0.14.0 Index]