Get model parameters from estimated marginal means objects

Returns the coefficients from a model.

Usage

# S3 method for emmGrid
get_parameters(x, summary = FALSE, merge_parameters = FALSE, ...)

# S3 method for emm_list
get_parameters(x, summary = FALSE, ...)

Arguments

x

A fitted model.

summary

Logical, indicates whether the full posterior samples (summary = FALSE)) or the summarized centrality indices of the posterior samples (summary = TRUE)) should be returned as estimates.

merge_parameters

Logical, if TRUE and x has multiple columns for parameter names (like emmGrid objects may have), these are merged into a single parameter column, with parameters names and values as values.

...

Currently not used.

Value

A data frame with two columns: the parameter names and the related point estimates.

Note

Note that emmGrid or emm_list objects returned by functions from emmeans have a different structure compared to usual regression models. Hence, the Parameter column does not always contain names of variables, but may rather contain values, e.g. for contrasts. See an example for pairwise comparisons below.

Examples

data(mtcars)
model <- lm(mpg ~ wt * factor(cyl), data = mtcars)
if (require("emmeans", quietly = TRUE)) {
  emm <- emmeans(model, "cyl")
  get_parameters(emm)

  emm <- emmeans(model, pairwise ~ cyl)
  get_parameters(emm)
}
#> NOTE: Results may be misleading due to involvement in interactions
#> Warning: Following columns were not found or were no character vectors:
#>   Parameter
#> NOTE: Results may be misleading due to involvement in interactions
#> Warning: Following columns were not found or were no character vectors:
#>   Parameter
#> Warning: Following columns were not found or were no character vectors:
#>   Parameter
#>   Parameter  Estimate Component
#> 1         4 21.403304   emmeans
#> 2         6 19.464549   emmeans
#> 3         8 16.814408   emmeans
#> 4     4 - 6  1.938755 contrasts
#> 5     4 - 8  4.588896 contrasts
#> 6     6 - 8  2.650141 contrasts