Checks if a regression model object is supported by the insight package

Small helper that checks if a model is a supported (regression) model object. supported_models() prints a list of currently supported model classes.

Usage

is_model_supported(x)

supported_models()

Arguments

x

An object.

Value

A logical, TRUE if x is a (supported) model object.

Details

This function returns TRUE if x is a model object that works with the package's functions. A list of supported models can also be found here: https://github.com/easystats/insight.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)

is_model_supported(m)
#> [1] TRUE
is_model_supported(mtcars)
#> [1] FALSE

# to see all supported models
supported_models()
#>   [1] "AKP"                     "Anova.mlm"              
#>   [3] "Arima"                   "BBmm"                   
#>   [5] "BBreg"                   "BFBayesFactor"          
#>   [7] "BGGM"                    "DirichletRegModel"      
#>   [9] "Gam"                     "Glm"                    
#>  [11] "HLfit"                   "LORgee"                 
#>  [13] "MANOVA"                  "MCMCglmm"               
#>  [15] "MixMod"                  "PMCMR"                  
#>  [17] "RM"                      "Rchoice"                
#>  [19] "Sarlm"                   "SemiParBIV"             
#>  [21] "aareg"                   "afex_aov"               
#>  [23] "anova.rms"               "aov"                    
#>  [25] "aovlist"                 "averaging"              
#>  [27] "bamlss"                  "bamlss.frame"           
#>  [29] "bayesQR"                 "bayesx"                 
#>  [31] "bcplm"                   "betamfx"                
#>  [33] "betaor"                  "betareg"                
#>  [35] "bfsl"                    "bife"                   
#>  [37] "bifeAPEs"                "bigglm"                 
#>  [39] "biglm"                   "blavaan"                
#>  [41] "blrm"                    "bracl"                  
#>  [43] "brglm"                   "brmsfit"                
#>  [45] "brmultinom"              "btergm"                 
#>  [47] "censReg"                 "cgam"                   
#>  [49] "cgamm"                   "cglm"                   
#>  [51] "clm"                     "clm2"                   
#>  [53] "clmm"                    "clmm2"                  
#>  [55] "clogit"                  "coeftest"               
#>  [57] "complmrob"               "confusionMatrix"        
#>  [59] "coxme"                   "coxph"                  
#>  [61] "coxph.penal"             "coxr"                   
#>  [63] "cpglm"                   "cpglmm"                 
#>  [65] "crch"                    "crq"                    
#>  [67] "crqs"                    "crr"                    
#>  [69] "dep.effect"              "draws"                  
#>  [71] "drc"                     "eglm"                   
#>  [73] "elm"                     "emmGrid"                
#>  [75] "epi.2by2"                "ergm"                   
#>  [77] "feglm"                   "feis"                   
#>  [79] "felm"                    "fitdistr"               
#>  [81] "fixest"                  "flac"                   
#>  [83] "flexsurvreg"             "flic"                   
#>  [85] "gam"                     "gamlss"                 
#>  [87] "gamm"                    "gamm4"                  
#>  [89] "garch"                   "gbm"                    
#>  [91] "gee"                     "geeglm"                 
#>  [93] "ggcomparisons"           "glht"                   
#>  [95] "glimML"                  "glm"                    
#>  [97] "glmRob"                  "glm_weighit"            
#>  [99] "glmerMod"                "glmgee"                 
#> [101] "glmm"                    "glmmPQL"                
#> [103] "glmmTMB"                 "glmmadmb"               
#> [105] "glmrob"                  "glmx"                   
#> [107] "gls"                     "gmnl"                   
#> [109] "hglm"                    "htest"                  
#> [111] "hurdle"                  "ivFixed"                
#> [113] "iv_robust"               "ivprobit"               
#> [115] "ivreg"                   "lavaan"                 
#> [117] "lm"                      "lmRob"                  
#> [119] "lm_robust"               "lme"                    
#> [121] "lmerMod"                 "lmerModLmerTest"        
#> [123] "lmodel2"                 "lmrob"                  
#> [125] "logistf"                 "logitmfx"               
#> [127] "logitor"                 "logitr"                 
#> [129] "lqm"                     "lqmm"                   
#> [131] "lrm"                     "manova"                 
#> [133] "marginaleffects"         "marginaleffects.summary"
#> [135] "margins"                 "maxLik"                 
#> [137] "mblogit"                 "mclogit"                
#> [139] "mcmc"                    "mcmc.list"              
#> [141] "mcp1"                    "mcp12"                  
#> [143] "mcp2"                    "med1way"                
#> [145] "mediate"                 "merMod"                 
#> [147] "merModList"              "meta_bma"               
#> [149] "meta_fixed"              "meta_random"            
#> [151] "metaplus"                "mhurdle"                
#> [153] "mipo"                    "mira"                   
#> [155] "mixed"                   "mixor"                  
#> [157] "mjoint"                  "mle"                    
#> [159] "mle2"                    "mlm"                    
#> [161] "mlogit"                  "mmclogit"               
#> [163] "mmlogit"                 "mmrm"                   
#> [165] "mmrm_fit"                "mmrm_tmb"               
#> [167] "model_fit"               "multinom"               
#> [169] "multinom_weightit"       "mvord"                  
#> [171] "negbinirr"               "negbinmfx"              
#> [173] "nestedLogit"             "ols"                    
#> [175] "onesampb"                "ordinal_weighit"        
#> [177] "orm"                     "pgmm"                   
#> [179] "phyloglm"                "phylolm"                
#> [181] "plm"                     "poissonirr"             
#> [183] "poissonmfx"              "polr"                   
#> [185] "probitmfx"               "psm"                    
#> [187] "ridgelm"                 "riskRegression"         
#> [189] "rjags"                   "rlm"                    
#> [191] "rlmerMod"                "rma"                    
#> [193] "rma.uni"                 "robmixglm"              
#> [195] "robtab"                  "rq"                     
#> [197] "rqs"                     "rqss"                   
#> [199] "rvar"                    "scam"                   
#> [201] "selection"               "sem"                    
#> [203] "semLm"                   "semLme"                 
#> [205] "serp"                    "slm"                    
#> [207] "speedglm"                "speedlm"                
#> [209] "stanfit"                 "stanmvreg"              
#> [211] "stanreg"                 "summary.lm"             
#> [213] "survfit"                 "survreg"                
#> [215] "svy2lme"                 "svy_vglm"               
#> [217] "svychisq"                "svyglm"                 
#> [219] "svyolr"                  "t1way"                  
#> [221] "tobit"                   "trimcibt"               
#> [223] "truncreg"                "vgam"                   
#> [225] "vglm"                    "wbgee"                  
#> [227] "wblm"                    "wbm"                    
#> [229] "wmcpAKP"                 "yuen"                   
#> [231] "yuend"                   "zcpglm"                 
#> [233] "zeroinfl"                "zerotrunc"