n_obs {insight}	R Documentation

Get number of observations from a model

Description

This method returns the number of observation that were used to fit the model, as numeric value.

Usage

n_obs(x, ...)

## S3 method for class 'svyolr'
n_obs(x, weighted = FALSE, ...)

## S3 method for class 'stanmvreg'
n_obs(x, select = NULL, ...)

Arguments

x
A fitted model.

...
Currently not used.

weighted
For survey designs, returns the weighted sample size.

select
Optional name(s) of response variables for which to extract values. Can be used in case of regression models with multiple response variables.

Value

The number of observations used to fit the model, or NULL if this information is not available.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
n_obs(m)

[Package insight version 0.14.0 Index]