check_if_installed {insight}	R Documentation

Checking if needed package is installed

Description

Checking if needed package is installed

Usage

check_if_installed(package, stop = TRUE, ...)

Arguments

package
A string naming the package, whose installation needs to be checked in any of the libraries

stop
Logical that decides whether the function should stop if the needed package is not installed.

...
Currently ignored

[Package insight version 0.14.0 Index]