Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  1.64248991
#> 2    X2 -0.35524021
#> 3    X3 -0.01966118
#> 4    X4  0.11414644
#> 5    X5  2.45970784
#> 6    X1  1.53597616
#> 7    X2 -0.07806403
#> 8    X3  1.65939801
#> 9    X4  1.17660850
#> 10   X5 -0.73620768
#> 11   X1 -0.15341364
#> 12   X2 -0.26721094
#> 13   X3  1.37083630
#> 14   X4 -0.33957581
#> 15   X5 -0.49019164
#> 16   X1  1.12026894
#> 17   X2 -0.20961284
#> 18   X3  0.44654721
#> 19   X4  0.76261231
#> 20   X5  1.64625677
#> 21   X1  0.79510044
#> 22   X2  0.60377038
#> 23   X3  0.96976152
#> 24   X4  0.13897105
#> 25   X5 -0.47458578
#> 26   X1 -1.50260096
#> 27   X2 -1.34944539
#> 28   X3 -0.25245648
#> 29   X4  0.18849947
#> 30   X5  1.68340301
#> 31   X1 -0.78916355
#> 32   X2  1.09074304
#> 33   X3  1.08255829
#> 34   X4  0.10364997
#> 35   X5  1.12940153
#> 36   X1  0.15815909
#> 37   X2 -0.34369287
#> 38   X3 -1.32780515
#> 39   X4 -1.26040004
#> 40   X5 -2.17672219
#> 41   X1 -0.23251272
#> 42   X2  0.10326362
#> 43   X3 -0.31279703
#> 44   X4  1.29985885
#> 45   X5  0.40759303
#> 46   X1 -0.06629594
#> 47   X2 -0.52139711
#> 48   X3 -0.14935095
#> 49   X4  0.28752436
#> 50   X5 -0.43990300

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3         X4         X5 Row Column     Numbers
#> 1  -0.01966118  0.1141464  2.4597078   1     X1  1.64248991
#> 2  -0.01966118  0.1141464  2.4597078   1     X2 -0.35524021
#> 3   1.65939801  1.1766085 -0.7362077   2     X1  1.53597616
#> 4   1.65939801  1.1766085 -0.7362077   2     X2 -0.07806403
#> 5   1.37083630 -0.3395758 -0.4901916   3     X1 -0.15341364
#> 6   1.37083630 -0.3395758 -0.4901916   3     X2 -0.26721094
#> 7   0.44654721  0.7626123  1.6462568   4     X1  1.12026894
#> 8   0.44654721  0.7626123  1.6462568   4     X2 -0.20961284
#> 9   0.96976152  0.1389710 -0.4745858   5     X1  0.79510044
#> 10  0.96976152  0.1389710 -0.4745858   5     X2  0.60377038
#> 11 -0.25245648  0.1884995  1.6834030   6     X1 -1.50260096
#> 12 -0.25245648  0.1884995  1.6834030   6     X2 -1.34944539
#> 13  1.08255829  0.1036500  1.1294015   7     X1 -0.78916355
#> 14  1.08255829  0.1036500  1.1294015   7     X2  1.09074304
#> 15 -1.32780515 -1.2604000 -2.1767222   8     X1  0.15815909
#> 16 -1.32780515 -1.2604000 -2.1767222   8     X2 -0.34369287
#> 17 -0.31279703  1.2998588  0.4075930   9     X1 -0.23251272
#> 18 -0.31279703  1.2998588  0.4075930   9     X2  0.10326362
#> 19 -0.14935095  0.2875244 -0.4399030  10     X1 -0.06629594
#> 20 -0.14935095  0.2875244 -0.4399030  10     X2 -0.52139711

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2    Value_X3   Value_X4   Value_X5
#> 1       1  1.64248991 -0.35524021 -0.01966118  0.1141464  2.4597078
#> 2       2  1.53597616 -0.07806403  1.65939801  1.1766085 -0.7362077
#> 3       3 -0.15341364 -0.26721094  1.37083630 -0.3395758 -0.4901916
#> 4       4  1.12026894 -0.20961284  0.44654721  0.7626123  1.6462568
#> 5       5  0.79510044  0.60377038  0.96976152  0.1389710 -0.4745858
#> 6       6 -1.50260096 -1.34944539 -0.25245648  0.1884995  1.6834030
#> 7       7 -0.78916355  1.09074304  1.08255829  0.1036500  1.1294015
#> 8       8  0.15815909 -0.34369287 -1.32780515 -1.2604000 -2.1767222
#> 9       9 -0.23251272  0.10326362 -0.31279703  1.2998588  0.4075930
#> 10     10 -0.06629594 -0.52139711 -0.14935095  0.2875244 -0.4399030

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5