Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1 -0.765730078
#> 2    X2  0.119997823
#> 3    X3 -0.709825686
#> 4    X4  0.004469455
#> 5    X5  0.808052887
#> 6    X1  0.797781565
#> 7    X2 -1.211665907
#> 8    X3 -1.582321130
#> 9    X4 -0.216473882
#> 10   X5  0.018570686
#> 11   X1  2.574136904
#> 12   X2 -0.827106877
#> 13   X3  0.204800334
#> 14   X4 -0.729809521
#> 15   X5  1.960001685
#> 16   X1  0.900318832
#> 17   X2  0.835932397
#> 18   X3 -0.513378293
#> 19   X4 -2.249695129
#> 20   X5  0.578900528
#> 21   X1  0.713771112
#> 22   X2  0.022246440
#> 23   X3 -0.604853969
#> 24   X4 -1.135103701
#> 25   X5  0.861267475
#> 26   X1 -0.517375802
#> 27   X2  0.472134140
#> 28   X3  0.451904452
#> 29   X4 -0.082488492
#> 30   X5  1.015514897
#> 31   X1  0.739908448
#> 32   X2 -0.528979188
#> 33   X3 -0.662992251
#> 34   X4 -0.153945955
#> 35   X5  1.211250212
#> 36   X1  1.851657944
#> 37   X2  0.728196342
#> 38   X3 -2.171310894
#> 39   X4  0.742063198
#> 40   X5  1.563609785
#> 41   X1  0.076524505
#> 42   X2 -0.546984632
#> 43   X3 -0.230157096
#> 44   X4  0.009104546
#> 45   X5  0.100686946
#> 46   X1 -0.648787270
#> 47   X2 -1.529725205
#> 48   X3  0.949957851
#> 49   X4 -0.273461962
#> 50   X5  0.573281278

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3           X4         X5 Row Column     Numbers
#> 1  -0.7098257  0.004469455 0.80805289   1     X1 -0.76573008
#> 2  -0.7098257  0.004469455 0.80805289   1     X2  0.11999782
#> 3  -1.5823211 -0.216473882 0.01857069   2     X1  0.79778157
#> 4  -1.5823211 -0.216473882 0.01857069   2     X2 -1.21166591
#> 5   0.2048003 -0.729809521 1.96000168   3     X1  2.57413690
#> 6   0.2048003 -0.729809521 1.96000168   3     X2 -0.82710688
#> 7  -0.5133783 -2.249695129 0.57890053   4     X1  0.90031883
#> 8  -0.5133783 -2.249695129 0.57890053   4     X2  0.83593240
#> 9  -0.6048540 -1.135103701 0.86126747   5     X1  0.71377111
#> 10 -0.6048540 -1.135103701 0.86126747   5     X2  0.02224644
#> 11  0.4519045 -0.082488492 1.01551490   6     X1 -0.51737580
#> 12  0.4519045 -0.082488492 1.01551490   6     X2  0.47213414
#> 13 -0.6629923 -0.153945955 1.21125021   7     X1  0.73990845
#> 14 -0.6629923 -0.153945955 1.21125021   7     X2 -0.52897919
#> 15 -2.1713109  0.742063198 1.56360979   8     X1  1.85165794
#> 16 -2.1713109  0.742063198 1.56360979   8     X2  0.72819634
#> 17 -0.2301571  0.009104546 0.10068695   9     X1  0.07652450
#> 18 -0.2301571  0.009104546 0.10068695   9     X2 -0.54698463
#> 19  0.9499579 -0.273461962 0.57328128  10     X1 -0.64878727
#> 20  0.9499579 -0.273461962 0.57328128  10     X2 -1.52972520

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2   Value_X3     Value_X4   Value_X5
#> 1       1 -0.7657301  0.11999782 -0.7098257  0.004469455 0.80805289
#> 2       2  0.7977816 -1.21166591 -1.5823211 -0.216473882 0.01857069
#> 3       3  2.5741369 -0.82710688  0.2048003 -0.729809521 1.96000168
#> 4       4  0.9003188  0.83593240 -0.5133783 -2.249695129 0.57890053
#> 5       5  0.7137711  0.02224644 -0.6048540 -1.135103701 0.86126747
#> 6       6 -0.5173758  0.47213414  0.4519045 -0.082488492 1.01551490
#> 7       7  0.7399084 -0.52897919 -0.6629923 -0.153945955 1.21125021
#> 8       8  1.8516579  0.72819634 -2.1713109  0.742063198 1.56360979
#> 9       9  0.0765245 -0.54698463 -0.2301571  0.009104546 0.10068695
#> 10     10 -0.6487873 -1.52972520  0.9499579 -0.273461962 0.57328128

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5