Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name         Value
#> 1    X1  0.5161438503
#> 2    X2 -0.9510838107
#> 3    X3  0.6527596571
#> 4    X4 -0.6518749385
#> 5    X5 -0.0020339521
#> 6    X1 -1.2511735884
#> 7    X2  0.0108518434
#> 8    X3  1.8554281307
#> 9    X4  0.0007774867
#> 10   X5 -0.3019873359
#> 11   X1 -0.4433333313
#> 12   X2 -0.0024162993
#> 13   X3 -0.9598348577
#> 14   X4  1.2703968053
#> 15   X5  1.0978128300
#> 16   X1  0.5552912571
#> 17   X2  2.4250529870
#> 18   X3  0.7956564587
#> 19   X4 -0.2947420684
#> 20   X5  0.3007497949
#> 21   X1  1.0587258012
#> 22   X2 -0.7968392344
#> 23   X3  0.1385287202
#> 24   X4  1.7827054595
#> 25   X5 -1.4415143986
#> 26   X1 -0.0053214506
#> 27   X2  1.0354206272
#> 28   X3  1.2242324692
#> 29   X4  1.2548852698
#> 30   X5  0.1262011577
#> 31   X1  1.5583515835
#> 32   X2 -0.6029985826
#> 33   X3  0.5399391391
#> 34   X4 -0.8983230623
#> 35   X5  0.0144352435
#> 36   X1 -0.3214425143
#> 37   X2 -0.9275932041
#> 38   X3 -0.7344969676
#> 39   X4  0.7046739720
#> 40   X5 -0.9526286680
#> 41   X1  0.9416988445
#> 42   X2  0.2396301014
#> 43   X3 -0.4473544122
#> 44   X4 -0.2088347592
#> 45   X5  0.7609373612
#> 46   X1 -0.3409056818
#> 47   X2 -0.3786181495
#> 48   X3  0.3841116304
#> 49   X4 -0.0934685220
#> 50   X5  0.0283365848

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3            X4           X5 Row Column      Numbers
#> 1   0.6527597 -0.6518749385 -0.002033952   1     X1  0.516143850
#> 2   0.6527597 -0.6518749385 -0.002033952   1     X2 -0.951083811
#> 3   1.8554281  0.0007774867 -0.301987336   2     X1 -1.251173588
#> 4   1.8554281  0.0007774867 -0.301987336   2     X2  0.010851843
#> 5  -0.9598349  1.2703968053  1.097812830   3     X1 -0.443333331
#> 6  -0.9598349  1.2703968053  1.097812830   3     X2 -0.002416299
#> 7   0.7956565 -0.2947420684  0.300749795   4     X1  0.555291257
#> 8   0.7956565 -0.2947420684  0.300749795   4     X2  2.425052987
#> 9   0.1385287  1.7827054595 -1.441514399   5     X1  1.058725801
#> 10  0.1385287  1.7827054595 -1.441514399   5     X2 -0.796839234
#> 11  1.2242325  1.2548852698  0.126201158   6     X1 -0.005321451
#> 12  1.2242325  1.2548852698  0.126201158   6     X2  1.035420627
#> 13  0.5399391 -0.8983230623  0.014435243   7     X1  1.558351583
#> 14  0.5399391 -0.8983230623  0.014435243   7     X2 -0.602998583
#> 15 -0.7344970  0.7046739720 -0.952628668   8     X1 -0.321442514
#> 16 -0.7344970  0.7046739720 -0.952628668   8     X2 -0.927593204
#> 17 -0.4473544 -0.2088347592  0.760937361   9     X1  0.941698845
#> 18 -0.4473544 -0.2088347592  0.760937361   9     X2  0.239630101
#> 19  0.3841116 -0.0934685220  0.028336585  10     X1 -0.340905682
#> 20  0.3841116 -0.0934685220  0.028336585  10     X2 -0.378618150

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID     Value_X1     Value_X2   Value_X3      Value_X4     Value_X5
#> 1       1  0.516143850 -0.951083811  0.6527597 -0.6518749385 -0.002033952
#> 2       2 -1.251173588  0.010851843  1.8554281  0.0007774867 -0.301987336
#> 3       3 -0.443333331 -0.002416299 -0.9598349  1.2703968053  1.097812830
#> 4       4  0.555291257  2.425052987  0.7956565 -0.2947420684  0.300749795
#> 5       5  1.058725801 -0.796839234  0.1385287  1.7827054595 -1.441514399
#> 6       6 -0.005321451  1.035420627  1.2242325  1.2548852698  0.126201158
#> 7       7  1.558351583 -0.602998583  0.5399391 -0.8983230623  0.014435243
#> 8       8 -0.321442514 -0.927593204 -0.7344970  0.7046739720 -0.952628668
#> 9       9  0.941698845  0.239630101 -0.4473544 -0.2088347592  0.760937361
#> 10     10 -0.340905682 -0.378618150  0.3841116 -0.0934685220  0.028336585

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5