Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1 -0.10881166
#> 2    X2 -0.57541912
#> 3    X3 -0.21826216
#> 4    X4  0.10218486
#> 5    X5 -0.35493829
#> 6    X1 -1.15556966
#> 7    X2  1.19626176
#> 8    X3  3.00346971
#> 9    X4  0.12705416
#> 10   X5 -0.75333639
#> 11   X1  0.06266453
#> 12   X2  0.56816312
#> 13   X3  0.55655825
#> 14   X4  1.25645389
#> 15   X5  0.12170966
#> 16   X1 -0.15082807
#> 17   X2  0.49555609
#> 18   X3 -1.54557651
#> 19   X4  1.06290966
#> 20   X5 -0.55323465
#> 21   X1 -1.69025959
#> 22   X2 -0.08403702
#> 23   X3 -1.32097767
#> 24   X4  0.26163755
#> 25   X5  1.13702738
#> 26   X1  1.50527200
#> 27   X2  0.98038507
#> 28   X3  1.63456839
#> 29   X4  0.93961821
#> 30   X5 -0.67321450
#> 31   X1 -1.53208436
#> 32   X2  2.34752917
#> 33   X3 -2.20006951
#> 34   X4  0.35512108
#> 35   X5 -1.54501751
#> 36   X1 -1.28150491
#> 37   X2 -1.77120493
#> 38   X3  0.36973504
#> 39   X4 -1.50618890
#> 40   X5 -0.98089025
#> 41   X1 -0.15812814
#> 42   X2 -2.02863932
#> 43   X3  0.70305097
#> 44   X4  0.47629250
#> 45   X5  0.90931130
#> 46   X1 -1.33609907
#> 47   X2 -0.21568855
#> 48   X3  1.79227135
#> 49   X4  1.13210586
#> 50   X5  1.49696743

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4         X5 Row Column     Numbers
#> 1  -0.2182622  0.1021849 -0.3549383   1     X1 -0.10881166
#> 2  -0.2182622  0.1021849 -0.3549383   1     X2 -0.57541912
#> 3   3.0034697  0.1270542 -0.7533364   2     X1 -1.15556966
#> 4   3.0034697  0.1270542 -0.7533364   2     X2  1.19626176
#> 5   0.5565582  1.2564539  0.1217097   3     X1  0.06266453
#> 6   0.5565582  1.2564539  0.1217097   3     X2  0.56816312
#> 7  -1.5455765  1.0629097 -0.5532347   4     X1 -0.15082807
#> 8  -1.5455765  1.0629097 -0.5532347   4     X2  0.49555609
#> 9  -1.3209777  0.2616376  1.1370274   5     X1 -1.69025959
#> 10 -1.3209777  0.2616376  1.1370274   5     X2 -0.08403702
#> 11  1.6345684  0.9396182 -0.6732145   6     X1  1.50527200
#> 12  1.6345684  0.9396182 -0.6732145   6     X2  0.98038507
#> 13 -2.2000695  0.3551211 -1.5450175   7     X1 -1.53208436
#> 14 -2.2000695  0.3551211 -1.5450175   7     X2  2.34752917
#> 15  0.3697350 -1.5061889 -0.9808903   8     X1 -1.28150491
#> 16  0.3697350 -1.5061889 -0.9808903   8     X2 -1.77120493
#> 17  0.7030510  0.4762925  0.9093113   9     X1 -0.15812814
#> 18  0.7030510  0.4762925  0.9093113   9     X2 -2.02863932
#> 19  1.7922713  1.1321059  1.4969674  10     X1 -1.33609907
#> 20  1.7922713  1.1321059  1.4969674  10     X2 -0.21568855

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2   Value_X3   Value_X4   Value_X5
#> 1       1 -0.10881166 -0.57541912 -0.2182622  0.1021849 -0.3549383
#> 2       2 -1.15556966  1.19626176  3.0034697  0.1270542 -0.7533364
#> 3       3  0.06266453  0.56816312  0.5565582  1.2564539  0.1217097
#> 4       4 -0.15082807  0.49555609 -1.5455765  1.0629097 -0.5532347
#> 5       5 -1.69025959 -0.08403702 -1.3209777  0.2616376  1.1370274
#> 6       6  1.50527200  0.98038507  1.6345684  0.9396182 -0.6732145
#> 7       7 -1.53208436  2.34752917 -2.2000695  0.3551211 -1.5450175
#> 8       8 -1.28150491 -1.77120493  0.3697350 -1.5061889 -0.9808903
#> 9       9 -0.15812814 -2.02863932  0.7030510  0.4762925  0.9093113
#> 10     10 -1.33609907 -0.21568855  1.7922713  1.1321059  1.4969674

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5