Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name         Value
#> 1    X1 -1.3316825865
#> 2    X2  1.2498426691
#> 3    X3 -0.8308476286
#> 4    X4 -0.0934633255
#> 5    X5 -0.3620536259
#> 6    X1  1.2677934778
#> 7    X2  0.5265207212
#> 8    X3  0.4248490593
#> 9    X4  0.3903622106
#> 10   X5  1.2947546432
#> 11   X1 -2.0940601504
#> 12   X2 -0.3662615875
#> 13   X3  1.4600602517
#> 14   X4 -0.7032799100
#> 15   X5 -1.2946693323
#> 16   X1 -0.5452719065
#> 17   X2  0.2619706216
#> 18   X3  1.9968478616
#> 19   X4 -0.5445168835
#> 20   X5  0.6843329687
#> 21   X1  0.9732013333
#> 22   X2  1.3919793221
#> 23   X3  1.1784449150
#> 24   X4 -0.1476859058
#> 25   X5  0.6496169322
#> 26   X1  1.3704089058
#> 27   X2  1.0847636453
#> 28   X3  1.2532559407
#> 29   X4 -0.0009109992
#> 30   X5  0.3934752207
#> 31   X1 -0.1700995452
#> 32   X2  0.5995942877
#> 33   X3  0.1112251444
#> 34   X4  1.4094661383
#> 35   X5  1.7225041443
#> 36   X1  0.4143612636
#> 37   X2  1.5334382615
#> 38   X3 -1.0389666415
#> 39   X4 -1.1931603923
#> 40   X5 -1.3824114165
#> 41   X1 -0.6913873213
#> 42   X2  1.2800964295
#> 43   X3 -0.4290258696
#> 44   X4  0.2807531125
#> 45   X5 -0.6299743945
#> 46   X1  0.0367347672
#> 47   X2  0.5022591259
#> 48   X3 -0.7229032960
#> 49   X4  0.8754435398
#> 50   X5  1.2689610680

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3            X4         X5 Row Column     Numbers
#> 1  -0.8308476 -0.0934633255 -0.3620536   1     X1 -1.33168259
#> 2  -0.8308476 -0.0934633255 -0.3620536   1     X2  1.24984267
#> 3   0.4248491  0.3903622106  1.2947546   2     X1  1.26779348
#> 4   0.4248491  0.3903622106  1.2947546   2     X2  0.52652072
#> 5   1.4600603 -0.7032799100 -1.2946693   3     X1 -2.09406015
#> 6   1.4600603 -0.7032799100 -1.2946693   3     X2 -0.36626159
#> 7   1.9968479 -0.5445168835  0.6843330   4     X1 -0.54527191
#> 8   1.9968479 -0.5445168835  0.6843330   4     X2  0.26197062
#> 9   1.1784449 -0.1476859058  0.6496169   5     X1  0.97320133
#> 10  1.1784449 -0.1476859058  0.6496169   5     X2  1.39197932
#> 11  1.2532559 -0.0009109992  0.3934752   6     X1  1.37040891
#> 12  1.2532559 -0.0009109992  0.3934752   6     X2  1.08476365
#> 13  0.1112251  1.4094661383  1.7225041   7     X1 -0.17009955
#> 14  0.1112251  1.4094661383  1.7225041   7     X2  0.59959429
#> 15 -1.0389666 -1.1931603923 -1.3824114   8     X1  0.41436126
#> 16 -1.0389666 -1.1931603923 -1.3824114   8     X2  1.53343826
#> 17 -0.4290259  0.2807531125 -0.6299744   9     X1 -0.69138732
#> 18 -0.4290259  0.2807531125 -0.6299744   9     X2  1.28009643
#> 19 -0.7229033  0.8754435398  1.2689611  10     X1  0.03673477
#> 20 -0.7229033  0.8754435398  1.2689611  10     X2  0.50225913

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1   Value_X2   Value_X3      Value_X4   Value_X5
#> 1       1 -1.33168259  1.2498427 -0.8308476 -0.0934633255 -0.3620536
#> 2       2  1.26779348  0.5265207  0.4248491  0.3903622106  1.2947546
#> 3       3 -2.09406015 -0.3662616  1.4600603 -0.7032799100 -1.2946693
#> 4       4 -0.54527191  0.2619706  1.9968479 -0.5445168835  0.6843330
#> 5       5  0.97320133  1.3919793  1.1784449 -0.1476859058  0.6496169
#> 6       6  1.37040891  1.0847636  1.2532559 -0.0009109992  0.3934752
#> 7       7 -0.17009955  0.5995943  0.1112251  1.4094661383  1.7225041
#> 8       8  0.41436126  1.5334383 -1.0389666 -1.1931603923 -1.3824114
#> 9       9 -0.69138732  1.2800964 -0.4290259  0.2807531125 -0.6299744
#> 10     10  0.03673477  0.5022591 -0.7229033  0.8754435398  1.2689611

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5