Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  1.06331150
#> 2    X2  1.41447069
#> 3    X3  1.71269446
#> 4    X4  1.84343065
#> 5    X5  0.50759113
#> 6    X1  1.09512808
#> 7    X2 -0.33316123
#> 8    X3 -1.49912702
#> 9    X4 -0.14525525
#> 10   X5  2.55903725
#> 11   X1 -0.07582808
#> 12   X2  1.33682804
#> 13   X3  0.27994473
#> 14   X4  0.83666349
#> 15   X5  2.04973930
#> 16   X1 -1.83763765
#> 17   X2  0.16366396
#> 18   X3 -0.26621892
#> 19   X4  0.41788301
#> 20   X5 -0.39078263
#> 21   X1  0.17923517
#> 22   X2  0.20256702
#> 23   X3  1.40966508
#> 24   X4 -0.97581067
#> 25   X5 -1.21946478
#> 26   X1  0.20066494
#> 27   X2  0.58279556
#> 28   X3 -0.61873255
#> 29   X4  0.33487483
#> 30   X5  0.35893291
#> 31   X1 -0.57726384
#> 32   X2  1.29537906
#> 33   X3  0.51816965
#> 34   X4 -2.29298432
#> 35   X5  1.39639964
#> 36   X1 -0.56201159
#> 37   X2 -0.49883155
#> 38   X3  0.00908059
#> 39   X4 -1.57522577
#> 40   X5  0.35026753
#> 41   X1 -0.16479741
#> 42   X2  0.10898485
#> 43   X3 -0.24745990
#> 44   X4  0.34914701
#> 45   X5  0.53177921
#> 46   X1  0.50493861
#> 47   X2  0.36425184
#> 48   X3  0.26509855
#> 49   X4  1.33494918
#> 50   X5  1.19884212

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3         X4         X5 Row Column     Numbers
#> 1   1.71269446  1.8434307  0.5075911   1     X1  1.06331150
#> 2   1.71269446  1.8434307  0.5075911   1     X2  1.41447069
#> 3  -1.49912702 -0.1452552  2.5590372   2     X1  1.09512808
#> 4  -1.49912702 -0.1452552  2.5590372   2     X2 -0.33316123
#> 5   0.27994473  0.8366635  2.0497393   3     X1 -0.07582808
#> 6   0.27994473  0.8366635  2.0497393   3     X2  1.33682804
#> 7  -0.26621892  0.4178830 -0.3907826   4     X1 -1.83763765
#> 8  -0.26621892  0.4178830 -0.3907826   4     X2  0.16366396
#> 9   1.40966508 -0.9758107 -1.2194648   5     X1  0.17923517
#> 10  1.40966508 -0.9758107 -1.2194648   5     X2  0.20256702
#> 11 -0.61873255  0.3348748  0.3589329   6     X1  0.20066494
#> 12 -0.61873255  0.3348748  0.3589329   6     X2  0.58279556
#> 13  0.51816965 -2.2929843  1.3963996   7     X1 -0.57726384
#> 14  0.51816965 -2.2929843  1.3963996   7     X2  1.29537906
#> 15  0.00908059 -1.5752258  0.3502675   8     X1 -0.56201159
#> 16  0.00908059 -1.5752258  0.3502675   8     X2 -0.49883155
#> 17 -0.24745990  0.3491470  0.5317792   9     X1 -0.16479741
#> 18 -0.24745990  0.3491470  0.5317792   9     X2  0.10898485
#> 19  0.26509855  1.3349492  1.1988421  10     X1  0.50493861
#> 20  0.26509855  1.3349492  1.1988421  10     X2  0.36425184

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1   Value_X2    Value_X3   Value_X4   Value_X5
#> 1       1  1.06331150  1.4144707  1.71269446  1.8434307  0.5075911
#> 2       2  1.09512808 -0.3331612 -1.49912702 -0.1452552  2.5590372
#> 3       3 -0.07582808  1.3368280  0.27994473  0.8366635  2.0497393
#> 4       4 -1.83763765  0.1636640 -0.26621892  0.4178830 -0.3907826
#> 5       5  0.17923517  0.2025670  1.40966508 -0.9758107 -1.2194648
#> 6       6  0.20066494  0.5827956 -0.61873255  0.3348748  0.3589329
#> 7       7 -0.57726384  1.2953791  0.51816965 -2.2929843  1.3963996
#> 8       8 -0.56201159 -0.4988316  0.00908059 -1.5752258  0.3502675
#> 9       9 -0.16479741  0.1089849 -0.24745990  0.3491470  0.5317792
#> 10     10  0.50493861  0.3642518  0.26509855  1.3349492  1.1988421

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5