Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1 -0.45372895
#> 2    X2  0.39420405
#> 3    X3 -0.42792055
#> 4    X4  0.56665677
#> 5    X5 -1.15136853
#> 6    X1 -1.31569817
#> 7    X2  0.47207559
#> 8    X3 -0.63623630
#> 9    X4 -0.73670836
#> 10   X5  1.02355129
#> 11   X1  1.00335595
#> 12   X2  0.20243098
#> 13   X3  0.22564767
#> 14   X4  0.15124150
#> 15   X5 -1.44000294
#> 16   X1 -1.27791418
#> 17   X2 -0.86796577
#> 18   X3 -0.97089244
#> 19   X4 -1.19836449
#> 20   X5 -0.58756327
#> 21   X1 -1.87547372
#> 22   X2 -0.21806029
#> 23   X3  0.66738430
#> 24   X4  0.97360109
#> 25   X5 -0.58742187
#> 26   X1  0.43430162
#> 27   X2  2.59131608
#> 28   X3  0.45919457
#> 29   X4  0.15510850
#> 30   X5  0.67532222
#> 31   X1  0.98025717
#> 32   X2  1.38789183
#> 33   X3  0.62121723
#> 34   X4  1.54039211
#> 35   X5  1.19568978
#> 36   X1 -0.74134985
#> 37   X2  0.42048039
#> 38   X3 -1.33868469
#> 39   X4  0.55240494
#> 40   X5 -1.56688410
#> 41   X1  0.31533558
#> 42   X2 -1.57903392
#> 43   X3  0.03397142
#> 44   X4  1.63194565
#> 45   X5  1.36004680
#> 46   X1 -0.21517170
#> 47   X2  1.52760526
#> 48   X3  0.77185419
#> 49   X4  0.97305525
#> 50   X5  0.89770112

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3         X4         X5 Row Column    Numbers
#> 1  -0.42792055  0.5666568 -1.1513685   1     X1 -0.4537290
#> 2  -0.42792055  0.5666568 -1.1513685   1     X2  0.3942041
#> 3  -0.63623630 -0.7367084  1.0235513   2     X1 -1.3156982
#> 4  -0.63623630 -0.7367084  1.0235513   2     X2  0.4720756
#> 5   0.22564767  0.1512415 -1.4400029   3     X1  1.0033560
#> 6   0.22564767  0.1512415 -1.4400029   3     X2  0.2024310
#> 7  -0.97089244 -1.1983645 -0.5875633   4     X1 -1.2779142
#> 8  -0.97089244 -1.1983645 -0.5875633   4     X2 -0.8679658
#> 9   0.66738430  0.9736011 -0.5874219   5     X1 -1.8754737
#> 10  0.66738430  0.9736011 -0.5874219   5     X2 -0.2180603
#> 11  0.45919457  0.1551085  0.6753222   6     X1  0.4343016
#> 12  0.45919457  0.1551085  0.6753222   6     X2  2.5913161
#> 13  0.62121723  1.5403921  1.1956898   7     X1  0.9802572
#> 14  0.62121723  1.5403921  1.1956898   7     X2  1.3878918
#> 15 -1.33868469  0.5524049 -1.5668841   8     X1 -0.7413499
#> 16 -1.33868469  0.5524049 -1.5668841   8     X2  0.4204804
#> 17  0.03397142  1.6319456  1.3600468   9     X1  0.3153356
#> 18  0.03397142  1.6319456  1.3600468   9     X2 -1.5790339
#> 19  0.77185419  0.9730552  0.8977011  10     X1 -0.2151717
#> 20  0.77185419  0.9730552  0.8977011  10     X2  1.5276053

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1   Value_X2    Value_X3   Value_X4   Value_X5
#> 1       1 -0.4537290  0.3942041 -0.42792055  0.5666568 -1.1513685
#> 2       2 -1.3156982  0.4720756 -0.63623630 -0.7367084  1.0235513
#> 3       3  1.0033560  0.2024310  0.22564767  0.1512415 -1.4400029
#> 4       4 -1.2779142 -0.8679658 -0.97089244 -1.1983645 -0.5875633
#> 5       5 -1.8754737 -0.2180603  0.66738430  0.9736011 -0.5874219
#> 6       6  0.4343016  2.5913161  0.45919457  0.1551085  0.6753222
#> 7       7  0.9802572  1.3878918  0.62121723  1.5403921  1.1956898
#> 8       8 -0.7413499  0.4204804 -1.33868469  0.5524049 -1.5668841
#> 9       9  0.3153356 -1.5790339  0.03397142  1.6319456  1.3600468
#> 10     10 -0.2151717  1.5276053  0.77185419  0.9730552  0.8977011

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5