Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  0.07763078
#> 2    X2 -1.02711050
#> 3    X3 -0.35543555
#> 4    X4  0.96637063
#> 5    X5  1.18164870
#> 6    X1  0.10684170
#> 7    X2  0.78868227
#> 8    X3  1.16476355
#> 9    X4 -0.52359551
#> 10   X5 -0.14969979
#> 11   X1 -0.01888320
#> 12   X2  0.85501085
#> 13   X3  0.68526733
#> 14   X4  2.31839245
#> 15   X5  0.10108190
#> 16   X1 -0.60750218
#> 17   X2 -1.69576569
#> 18   X3  2.47335249
#> 19   X4  1.05761903
#> 20   X5 -0.36882307
#> 21   X1 -0.86781317
#> 22   X2  1.35936053
#> 23   X3 -0.88238995
#> 24   X4  0.56299243
#> 25   X5  0.18320606
#> 26   X1 -0.04688594
#> 27   X2  0.76828183
#> 28   X3  0.87781229
#> 29   X4  1.06894352
#> 30   X5 -0.57015562
#> 31   X1 -0.99639717
#> 32   X2 -0.30832276
#> 33   X3 -0.80094780
#> 34   X4 -0.16664645
#> 35   X5  1.02568296
#> 36   X1  0.97092201
#> 37   X2 -0.65090295
#> 38   X3 -0.95963605
#> 39   X4  1.50516572
#> 40   X5  0.77196471
#> 41   X1 -2.09976052
#> 42   X2 -1.26559042
#> 43   X3  0.73734387
#> 44   X4  0.30670145
#> 45   X5 -0.91912619
#> 46   X1  0.40267574
#> 47   X2 -0.48797674
#> 48   X3 -1.14372495
#> 49   X4 -0.34691556
#> 50   X5 -1.12477879

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4         X5 Row Column     Numbers
#> 1  -0.3554356  0.9663706  1.1816487   1     X1  0.07763078
#> 2  -0.3554356  0.9663706  1.1816487   1     X2 -1.02711050
#> 3   1.1647635 -0.5235955 -0.1496998   2     X1  0.10684170
#> 4   1.1647635 -0.5235955 -0.1496998   2     X2  0.78868227
#> 5   0.6852673  2.3183925  0.1010819   3     X1 -0.01888320
#> 6   0.6852673  2.3183925  0.1010819   3     X2  0.85501085
#> 7   2.4733525  1.0576190 -0.3688231   4     X1 -0.60750218
#> 8   2.4733525  1.0576190 -0.3688231   4     X2 -1.69576569
#> 9  -0.8823899  0.5629924  0.1832061   5     X1 -0.86781317
#> 10 -0.8823899  0.5629924  0.1832061   5     X2  1.35936053
#> 11  0.8778123  1.0689435 -0.5701556   6     X1 -0.04688594
#> 12  0.8778123  1.0689435 -0.5701556   6     X2  0.76828183
#> 13 -0.8009478 -0.1666464  1.0256830   7     X1 -0.99639717
#> 14 -0.8009478 -0.1666464  1.0256830   7     X2 -0.30832276
#> 15 -0.9596361  1.5051657  0.7719647   8     X1  0.97092201
#> 16 -0.9596361  1.5051657  0.7719647   8     X2 -0.65090295
#> 17  0.7373439  0.3067014 -0.9191262   9     X1 -2.09976052
#> 18  0.7373439  0.3067014 -0.9191262   9     X2 -1.26559042
#> 19 -1.1437249 -0.3469156 -1.1247788  10     X1  0.40267574
#> 20 -1.1437249 -0.3469156 -1.1247788  10     X2 -0.48797674

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1   Value_X2   Value_X3   Value_X4   Value_X5
#> 1       1  0.07763078 -1.0271105 -0.3554356  0.9663706  1.1816487
#> 2       2  0.10684170  0.7886823  1.1647635 -0.5235955 -0.1496998
#> 3       3 -0.01888320  0.8550109  0.6852673  2.3183925  0.1010819
#> 4       4 -0.60750218 -1.6957657  2.4733525  1.0576190 -0.3688231
#> 5       5 -0.86781317  1.3593605 -0.8823899  0.5629924  0.1832061
#> 6       6 -0.04688594  0.7682818  0.8778123  1.0689435 -0.5701556
#> 7       7 -0.99639717 -0.3083228 -0.8009478 -0.1666464  1.0256830
#> 8       8  0.97092201 -0.6509029 -0.9596361  1.5051657  0.7719647
#> 9       9 -2.09976052 -1.2655904  0.7373439  0.3067014 -0.9191262
#> 10     10  0.40267574 -0.4879767 -1.1437249 -0.3469156 -1.1247788

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5