Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name         Value
#> 1    X1  0.6639308734
#> 2    X2 -0.5154080874
#> 3    X3  1.5750254694
#> 4    X4  0.2982153482
#> 5    X5 -0.4354101735
#> 6    X1 -0.4449079599
#> 7    X2 -1.2482263531
#> 8    X3  0.8509552002
#> 9    X4  0.8603369131
#> 10   X5 -1.0063561241
#> 11   X1  0.1278467609
#> 12   X2  0.3544853136
#> 13   X3 -0.3443186920
#> 14   X4 -1.0168454182
#> 15   X5 -1.5162419957
#> 16   X1 -0.4189269623
#> 17   X2 -0.0423112436
#> 18   X3  1.5437313437
#> 19   X4 -0.6716515963
#> 20   X5 -0.7823142912
#> 21   X1  1.0172971380
#> 22   X2  0.0572284083
#> 23   X3  1.8965750368
#> 24   X4  1.0026545870
#> 25   X5 -0.6411492694
#> 26   X1 -0.6865951222
#> 27   X2 -0.7203194175
#> 28   X3  0.1721113460
#> 29   X4  0.2059617167
#> 30   X5  1.0019665952
#> 31   X1  0.9697838299
#> 32   X2 -0.9848426037
#> 33   X3  0.1789340006
#> 34   X4 -0.7100313160
#> 35   X5  0.1769668771
#> 36   X1  0.3822057482
#> 37   X2  0.8776547909
#> 38   X3  1.5311975835
#> 39   X4 -0.2563207066
#> 40   X5  1.0856776324
#> 41   X1 -0.7880428927
#> 42   X2 -0.0008519792
#> 43   X3 -1.2375653371
#> 44   X4 -0.4241506548
#> 45   X5 -0.2035856342
#> 46   X1  0.5828398752
#> 47   X2 -1.4990962614
#> 48   X3  1.1387391302
#> 49   X4  1.1022631635
#> 50   X5 -0.0791299995

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4         X5 Row Column       Numbers
#> 1   1.5750255  0.2982153 -0.4354102   1     X1  0.6639308734
#> 2   1.5750255  0.2982153 -0.4354102   1     X2 -0.5154080874
#> 3   0.8509552  0.8603369 -1.0063561   2     X1 -0.4449079599
#> 4   0.8509552  0.8603369 -1.0063561   2     X2 -1.2482263531
#> 5  -0.3443187 -1.0168454 -1.5162420   3     X1  0.1278467609
#> 6  -0.3443187 -1.0168454 -1.5162420   3     X2  0.3544853136
#> 7   1.5437313 -0.6716516 -0.7823143   4     X1 -0.4189269623
#> 8   1.5437313 -0.6716516 -0.7823143   4     X2 -0.0423112436
#> 9   1.8965750  1.0026546 -0.6411493   5     X1  1.0172971380
#> 10  1.8965750  1.0026546 -0.6411493   5     X2  0.0572284083
#> 11  0.1721113  0.2059617  1.0019666   6     X1 -0.6865951222
#> 12  0.1721113  0.2059617  1.0019666   6     X2 -0.7203194175
#> 13  0.1789340 -0.7100313  0.1769669   7     X1  0.9697838299
#> 14  0.1789340 -0.7100313  0.1769669   7     X2 -0.9848426037
#> 15  1.5311976 -0.2563207  1.0856776   8     X1  0.3822057482
#> 16  1.5311976 -0.2563207  1.0856776   8     X2  0.8776547909
#> 17 -1.2375653 -0.4241507 -0.2035856   9     X1 -0.7880428927
#> 18 -1.2375653 -0.4241507 -0.2035856   9     X2 -0.0008519792
#> 19  1.1387391  1.1022632 -0.0791300  10     X1  0.5828398752
#> 20  1.1387391  1.1022632 -0.0791300  10     X2 -1.4990962614

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1      Value_X2   Value_X3   Value_X4   Value_X5
#> 1       1  0.6639309 -0.5154080874  1.5750255  0.2982153 -0.4354102
#> 2       2 -0.4449080 -1.2482263531  0.8509552  0.8603369 -1.0063561
#> 3       3  0.1278468  0.3544853136 -0.3443187 -1.0168454 -1.5162420
#> 4       4 -0.4189270 -0.0423112436  1.5437313 -0.6716516 -0.7823143
#> 5       5  1.0172971  0.0572284083  1.8965750  1.0026546 -0.6411493
#> 6       6 -0.6865951 -0.7203194175  0.1721113  0.2059617  1.0019666
#> 7       7  0.9697838 -0.9848426037  0.1789340 -0.7100313  0.1769669
#> 8       8  0.3822057  0.8776547909  1.5311976 -0.2563207  1.0856776
#> 9       9 -0.7880429 -0.0008519792 -1.2375653 -0.4241507 -0.2035856
#> 10     10  0.5828399 -1.4990962614  1.1387391  1.1022632 -0.0791300

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5