Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name         Value
#> 1    X1  0.8163739532
#> 2    X2  0.2112418273
#> 3    X3 -0.7109347352
#> 4    X4  1.0454608183
#> 5    X5  2.1793075060
#> 6    X1 -1.3486824715
#> 7    X2 -0.7798949678
#> 8    X3 -1.0369366302
#> 9    X4 -0.8518754937
#> 10   X5  0.9149254633
#> 11   X1 -1.0763039495
#> 12   X2  0.6816321595
#> 13   X3  1.1543351413
#> 14   X4  0.7939213552
#> 15   X5 -0.0259745488
#> 16   X1 -1.3020188507
#> 17   X2 -1.0256637918
#> 18   X3 -0.4040327020
#> 19   X4 -0.6500764700
#> 20   X5 -0.2485461639
#> 21   X1 -0.1883506176
#> 22   X2  1.1397972507
#> 23   X3 -1.6623839751
#> 24   X4  0.2235782228
#> 25   X5 -0.7469008579
#> 26   X1  0.5888210905
#> 27   X2 -0.2821926015
#> 28   X3 -0.9068106316
#> 29   X4 -1.8178899394
#> 30   X5  0.6493940098
#> 31   X1 -0.4876234300
#> 32   X2  0.0001546157
#> 33   X3  0.6042642468
#> 34   X4 -0.6001891021
#> 35   X5 -1.2707234144
#> 36   X1  0.7371872053
#> 37   X2 -0.8473367628
#> 38   X3 -0.9035913536
#> 39   X4  1.2981008000
#> 40   X5  0.6022475778
#> 41   X1  0.3614575393
#> 42   X2  1.8719753670
#> 43   X3  0.6139871356
#> 44   X4  0.8552283693
#> 45   X5  1.1000444582
#> 46   X1  0.5305835081
#> 47   X2  0.6700200917
#> 48   X3  0.4003110049
#> 49   X4  2.4392507631
#> 50   X5 -0.8450572104

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4          X5 Row Column       Numbers
#> 1  -0.7109347  1.0454608  2.17930751   1     X1  0.8163739532
#> 2  -0.7109347  1.0454608  2.17930751   1     X2  0.2112418273
#> 3  -1.0369366 -0.8518755  0.91492546   2     X1 -1.3486824715
#> 4  -1.0369366 -0.8518755  0.91492546   2     X2 -0.7798949678
#> 5   1.1543351  0.7939214 -0.02597455   3     X1 -1.0763039495
#> 6   1.1543351  0.7939214 -0.02597455   3     X2  0.6816321595
#> 7  -0.4040327 -0.6500765 -0.24854616   4     X1 -1.3020188507
#> 8  -0.4040327 -0.6500765 -0.24854616   4     X2 -1.0256637918
#> 9  -1.6623840  0.2235782 -0.74690086   5     X1 -0.1883506176
#> 10 -1.6623840  0.2235782 -0.74690086   5     X2  1.1397972507
#> 11 -0.9068106 -1.8178899  0.64939401   6     X1  0.5888210905
#> 12 -0.9068106 -1.8178899  0.64939401   6     X2 -0.2821926015
#> 13  0.6042642 -0.6001891 -1.27072341   7     X1 -0.4876234300
#> 14  0.6042642 -0.6001891 -1.27072341   7     X2  0.0001546157
#> 15 -0.9035914  1.2981008  0.60224758   8     X1  0.7371872053
#> 16 -0.9035914  1.2981008  0.60224758   8     X2 -0.8473367628
#> 17  0.6139871  0.8552284  1.10004446   9     X1  0.3614575393
#> 18  0.6139871  0.8552284  1.10004446   9     X2  1.8719753670
#> 19  0.4003110  2.4392508 -0.84505721  10     X1  0.5305835081
#> 20  0.4003110  2.4392508 -0.84505721  10     X2  0.6700200917

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1      Value_X2   Value_X3   Value_X4    Value_X5
#> 1       1  0.8163740  0.2112418273 -0.7109347  1.0454608  2.17930751
#> 2       2 -1.3486825 -0.7798949678 -1.0369366 -0.8518755  0.91492546
#> 3       3 -1.0763039  0.6816321595  1.1543351  0.7939214 -0.02597455
#> 4       4 -1.3020189 -1.0256637918 -0.4040327 -0.6500765 -0.24854616
#> 5       5 -0.1883506  1.1397972507 -1.6623840  0.2235782 -0.74690086
#> 6       6  0.5888211 -0.2821926015 -0.9068106 -1.8178899  0.64939401
#> 7       7 -0.4876234  0.0001546157  0.6042642 -0.6001891 -1.27072341
#> 8       8  0.7371872 -0.8473367628 -0.9035914  1.2981008  0.60224758
#> 9       9  0.3614575  1.8719753670  0.6139871  0.8552284  1.10004446
#> 10     10  0.5305835  0.6700200917  0.4003110  2.4392508 -0.84505721

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5