Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  1.12080300
#> 2    X2  0.63148090
#> 3    X3 -1.29020986
#> 4    X4 -0.39326143
#> 5    X5 -0.87690651
#> 6    X1 -0.48342584
#> 7    X2 -1.23340605
#> 8    X3  0.92290968
#> 9    X4 -0.41315896
#> 10   X5  1.11356028
#> 11   X1 -2.51996293
#> 12   X2 -0.28537974
#> 13   X3  0.04350110
#> 14   X4 -0.21133509
#> 15   X5  0.94077383
#> 16   X1  0.66805012
#> 17   X2  0.46163176
#> 18   X3  0.07685319
#> 19   X4 -0.12055895
#> 20   X5 -0.32767884
#> 21   X1 -1.19667238
#> 22   X2  0.78483642
#> 23   X3 -0.51392651
#> 24   X4  0.13930561
#> 25   X5 -0.77814664
#> 26   X1 -2.44281576
#> 27   X2 -0.79629090
#> 28   X3  0.13094774
#> 29   X4  1.70341028
#> 30   X5 -2.27323689
#> 31   X1 -0.29807073
#> 32   X2  0.61755575
#> 33   X3  0.55761028
#> 34   X4  0.61284703
#> 35   X5  1.96050178
#> 36   X1 -0.58075657
#> 37   X2 -1.40102113
#> 38   X3  0.35752150
#> 39   X4  0.74754125
#> 40   X5  0.93076506
#> 41   X1 -0.48087240
#> 42   X2  1.10879774
#> 43   X3 -1.50609042
#> 44   X4  0.18287087
#> 45   X5  0.31856623
#> 46   X1  1.26166231
#> 47   X2  1.18324612
#> 48   X3 -0.10839858
#> 49   X4  0.70118404
#> 50   X5 -0.74174084

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3         X4         X5 Row Column    Numbers
#> 1  -1.29020986 -0.3932614 -0.8769065   1     X1  1.1208030
#> 2  -1.29020986 -0.3932614 -0.8769065   1     X2  0.6314809
#> 3   0.92290968 -0.4131590  1.1135603   2     X1 -0.4834258
#> 4   0.92290968 -0.4131590  1.1135603   2     X2 -1.2334060
#> 5   0.04350110 -0.2113351  0.9407738   3     X1 -2.5199629
#> 6   0.04350110 -0.2113351  0.9407738   3     X2 -0.2853797
#> 7   0.07685319 -0.1205589 -0.3276788   4     X1  0.6680501
#> 8   0.07685319 -0.1205589 -0.3276788   4     X2  0.4616318
#> 9  -0.51392651  0.1393056 -0.7781466   5     X1 -1.1966724
#> 10 -0.51392651  0.1393056 -0.7781466   5     X2  0.7848364
#> 11  0.13094774  1.7034103 -2.2732369   6     X1 -2.4428158
#> 12  0.13094774  1.7034103 -2.2732369   6     X2 -0.7962909
#> 13  0.55761028  0.6128470  1.9605018   7     X1 -0.2980707
#> 14  0.55761028  0.6128470  1.9605018   7     X2  0.6175557
#> 15  0.35752150  0.7475413  0.9307651   8     X1 -0.5807566
#> 16  0.35752150  0.7475413  0.9307651   8     X2 -1.4010211
#> 17 -1.50609042  0.1828709  0.3185662   9     X1 -0.4808724
#> 18 -1.50609042  0.1828709  0.3185662   9     X2  1.1087977
#> 19 -0.10839858  0.7011840 -0.7417408  10     X1  1.2616623
#> 20 -0.10839858  0.7011840 -0.7417408  10     X2  1.1832461

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1   Value_X2    Value_X3   Value_X4   Value_X5
#> 1       1  1.1208030  0.6314809 -1.29020986 -0.3932614 -0.8769065
#> 2       2 -0.4834258 -1.2334060  0.92290968 -0.4131590  1.1135603
#> 3       3 -2.5199629 -0.2853797  0.04350110 -0.2113351  0.9407738
#> 4       4  0.6680501  0.4616318  0.07685319 -0.1205589 -0.3276788
#> 5       5 -1.1966724  0.7848364 -0.51392651  0.1393056 -0.7781466
#> 6       6 -2.4428158 -0.7962909  0.13094774  1.7034103 -2.2732369
#> 7       7 -0.2980707  0.6175557  0.55761028  0.6128470  1.9605018
#> 8       8 -0.5807566 -1.4010211  0.35752150  0.7475413  0.9307651
#> 9       9 -0.4808724  1.1087977 -1.50609042  0.1828709  0.3185662
#> 10     10  1.2616623  1.1832461 -0.10839858  0.7011840 -0.7417408

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5