Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1 -1.41619972
#> 2    X2  0.74097432
#> 3    X3  0.20497872
#> 4    X4 -0.25879996
#> 5    X5 -1.19983684
#> 6    X1  0.52459492
#> 7    X2  0.08810772
#> 8    X3  0.23267496
#> 9    X4 -0.36298830
#> 10   X5  0.44452784
#> 11   X1 -0.27912013
#> 12   X2 -1.23760797
#> 13   X3  0.59223669
#> 14   X4  0.03589099
#> 15   X5 -0.23684962
#> 16   X1  0.32335550
#> 17   X2  0.07360660
#> 18   X3  0.45168754
#> 19   X4  0.51600809
#> 20   X5 -1.39585439
#> 21   X1 -0.15411210
#> 22   X2 -1.02056430
#> 23   X3 -1.88917975
#> 24   X4  1.56166137
#> 25   X5  0.95869497
#> 26   X1 -0.42435040
#> 27   X2  0.86188716
#> 28   X3  0.16584280
#> 29   X4 -0.87469499
#> 30   X5  1.70231769
#> 31   X1 -0.86944776
#> 32   X2  0.17897195
#> 33   X3  0.07492413
#> 34   X4 -0.26683701
#> 35   X5 -0.50097031
#> 36   X1 -0.31894577
#> 37   X2 -1.36810887
#> 38   X3 -0.23694997
#> 39   X4  0.42983607
#> 40   X5 -0.07951285
#> 41   X1  0.29566726
#> 42   X2  0.21809886
#> 43   X3 -0.49081476
#> 44   X4 -1.87627222
#> 45   X5 -1.86807510
#> 46   X1 -0.67788414
#> 47   X2  0.66939653
#> 48   X3 -1.31615668
#> 49   X4  0.45183947
#> 50   X5  0.84402996

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4          X5 Row Column     Numbers
#> 1   0.20497872 -0.25879996 -1.19983684   1     X1 -1.41619972
#> 2   0.20497872 -0.25879996 -1.19983684   1     X2  0.74097432
#> 3   0.23267496 -0.36298830  0.44452784   2     X1  0.52459492
#> 4   0.23267496 -0.36298830  0.44452784   2     X2  0.08810772
#> 5   0.59223669  0.03589099 -0.23684962   3     X1 -0.27912013
#> 6   0.59223669  0.03589099 -0.23684962   3     X2 -1.23760797
#> 7   0.45168754  0.51600809 -1.39585439   4     X1  0.32335550
#> 8   0.45168754  0.51600809 -1.39585439   4     X2  0.07360660
#> 9  -1.88917975  1.56166137  0.95869497   5     X1 -0.15411210
#> 10 -1.88917975  1.56166137  0.95869497   5     X2 -1.02056430
#> 11  0.16584280 -0.87469499  1.70231769   6     X1 -0.42435040
#> 12  0.16584280 -0.87469499  1.70231769   6     X2  0.86188716
#> 13  0.07492413 -0.26683701 -0.50097031   7     X1 -0.86944776
#> 14  0.07492413 -0.26683701 -0.50097031   7     X2  0.17897195
#> 15 -0.23694997  0.42983607 -0.07951285   8     X1 -0.31894577
#> 16 -0.23694997  0.42983607 -0.07951285   8     X2 -1.36810887
#> 17 -0.49081476 -1.87627222 -1.86807510   9     X1  0.29566726
#> 18 -0.49081476 -1.87627222 -1.86807510   9     X2  0.21809886
#> 19 -1.31615668  0.45183947  0.84402996  10     X1 -0.67788414
#> 20 -1.31615668  0.45183947  0.84402996  10     X2  0.66939653

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2    Value_X3    Value_X4    Value_X5
#> 1       1 -1.4161997  0.74097432  0.20497872 -0.25879996 -1.19983684
#> 2       2  0.5245949  0.08810772  0.23267496 -0.36298830  0.44452784
#> 3       3 -0.2791201 -1.23760797  0.59223669  0.03589099 -0.23684962
#> 4       4  0.3233555  0.07360660  0.45168754  0.51600809 -1.39585439
#> 5       5 -0.1541121 -1.02056430 -1.88917975  1.56166137  0.95869497
#> 6       6 -0.4243504  0.86188716  0.16584280 -0.87469499  1.70231769
#> 7       7 -0.8694478  0.17897195  0.07492413 -0.26683701 -0.50097031
#> 8       8 -0.3189458 -1.36810887 -0.23694997  0.42983607 -0.07951285
#> 9       9  0.2956673  0.21809886 -0.49081476 -1.87627222 -1.86807510
#> 10     10 -0.6778841  0.66939653 -1.31615668  0.45183947  0.84402996

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5