Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  1.97736399
#> 2    X2  0.80028453
#> 3    X3  1.03409943
#> 4    X4  0.25952231
#> 5    X5  2.68275351
#> 6    X1  1.83398167
#> 7    X2  1.23415208
#> 8    X3 -2.23529135
#> 9    X4 -0.95251235
#> 10   X5 -1.73312494
#> 11   X1  0.34695412
#> 12   X2  0.66444086
#> 13   X3 -0.12431685
#> 14   X4 -0.16921024
#> 15   X5  0.32030125
#> 16   X1 -0.38946214
#> 17   X2  1.07573000
#> 18   X3  1.08614941
#> 19   X4  0.79502117
#> 20   X5  0.69652102
#> 21   X1  1.62085453
#> 22   X2  0.29173961
#> 23   X3  1.93222637
#> 24   X4  1.19153912
#> 25   X5 -1.53397168
#> 26   X1  2.10450963
#> 27   X2 -0.48003147
#> 28   X3 -0.94443819
#> 29   X4  1.42807913
#> 30   X5 -1.86492567
#> 31   X1  0.43247864
#> 32   X2  0.80119154
#> 33   X3  1.45816365
#> 34   X4  0.62789958
#> 35   X5 -0.04638139
#> 36   X1 -1.51913861
#> 37   X2 -1.09077994
#> 38   X3 -1.72611750
#> 39   X4 -0.31373181
#> 40   X5 -1.20492001
#> 41   X1 -0.58896092
#> 42   X2  0.80024513
#> 43   X3 -2.16525770
#> 44   X4 -0.41530398
#> 45   X5  0.21087909
#> 46   X1 -0.55911248
#> 47   X2 -0.25797283
#> 48   X3  0.88613159
#> 49   X4 -1.88452729
#> 50   X5 -0.93400519

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4          X5 Row Column    Numbers
#> 1   1.0340994  0.2595223  2.68275351   1     X1  1.9773640
#> 2   1.0340994  0.2595223  2.68275351   1     X2  0.8002845
#> 3  -2.2352913 -0.9525124 -1.73312494   2     X1  1.8339817
#> 4  -2.2352913 -0.9525124 -1.73312494   2     X2  1.2341521
#> 5  -0.1243168 -0.1692102  0.32030125   3     X1  0.3469541
#> 6  -0.1243168 -0.1692102  0.32030125   3     X2  0.6644409
#> 7   1.0861494  0.7950212  0.69652102   4     X1 -0.3894621
#> 8   1.0861494  0.7950212  0.69652102   4     X2  1.0757300
#> 9   1.9322264  1.1915391 -1.53397168   5     X1  1.6208545
#> 10  1.9322264  1.1915391 -1.53397168   5     X2  0.2917396
#> 11 -0.9444382  1.4280791 -1.86492567   6     X1  2.1045096
#> 12 -0.9444382  1.4280791 -1.86492567   6     X2 -0.4800315
#> 13  1.4581637  0.6278996 -0.04638139   7     X1  0.4324786
#> 14  1.4581637  0.6278996 -0.04638139   7     X2  0.8011915
#> 15 -1.7261175 -0.3137318 -1.20492001   8     X1 -1.5191386
#> 16 -1.7261175 -0.3137318 -1.20492001   8     X2 -1.0907799
#> 17 -2.1652577 -0.4153040  0.21087909   9     X1 -0.5889609
#> 18 -2.1652577 -0.4153040  0.21087909   9     X2  0.8002451
#> 19  0.8861316 -1.8845273 -0.93400519  10     X1 -0.5591125
#> 20  0.8861316 -1.8845273 -0.93400519  10     X2 -0.2579728

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1   Value_X2   Value_X3   Value_X4    Value_X5
#> 1       1  1.9773640  0.8002845  1.0340994  0.2595223  2.68275351
#> 2       2  1.8339817  1.2341521 -2.2352913 -0.9525124 -1.73312494
#> 3       3  0.3469541  0.6644409 -0.1243168 -0.1692102  0.32030125
#> 4       4 -0.3894621  1.0757300  1.0861494  0.7950212  0.69652102
#> 5       5  1.6208545  0.2917396  1.9322264  1.1915391 -1.53397168
#> 6       6  2.1045096 -0.4800315 -0.9444382  1.4280791 -1.86492567
#> 7       7  0.4324786  0.8011915  1.4581637  0.6278996 -0.04638139
#> 8       8 -1.5191386 -1.0907799 -1.7261175 -0.3137318 -1.20492001
#> 9       9 -0.5889609  0.8002451 -2.1652577 -0.4153040  0.21087909
#> 10     10 -0.5591125 -0.2579728  0.8861316 -1.8845273 -0.93400519

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5