Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1 -0.412947581
#> 2    X2  0.190565236
#> 3    X3  0.033042687
#> 4    X4  0.026300206
#> 5    X5  0.899727247
#> 6    X1 -0.068229396
#> 7    X2  0.532909975
#> 8    X3  1.113073187
#> 9    X4 -1.258605925
#> 10   X5  1.060938941
#> 11   X1 -0.224967106
#> 12   X2 -0.690448093
#> 13   X3  0.045688843
#> 14   X4  0.338215180
#> 15   X5 -0.215857078
#> 16   X1 -0.640541930
#> 17   X2 -0.375132486
#> 18   X3 -1.310635300
#> 19   X4 -0.120887660
#> 20   X5 -0.301746588
#> 21   X1 -1.423965124
#> 22   X2  2.090077257
#> 23   X3  1.002337750
#> 24   X4  0.852555214
#> 25   X5 -0.324361588
#> 26   X1 -1.414982977
#> 27   X2 -0.341987866
#> 28   X3  0.003822226
#> 29   X4  1.146589129
#> 30   X5 -0.182130692
#> 31   X1 -0.165107985
#> 32   X2  0.573196869
#> 33   X3  0.154489750
#> 34   X4  1.458254975
#> 35   X5 -0.964135160
#> 36   X1  0.772817160
#> 37   X2 -0.872914083
#> 38   X3  0.404348293
#> 39   X4  1.149110855
#> 40   X5 -1.358417633
#> 41   X1 -0.630701670
#> 42   X2 -0.463787198
#> 43   X3 -0.687282145
#> 44   X4  1.303469525
#> 45   X5 -3.248626786
#> 46   X1  0.420163099
#> 47   X2  0.950313414
#> 48   X3 -1.047515857
#> 49   X4 -0.215199678
#> 50   X5 -0.789129172

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>              X3          X4         X5 Row Column    Numbers
#> 1   0.033042687  0.02630021  0.8997272   1     X1 -0.4129476
#> 2   0.033042687  0.02630021  0.8997272   1     X2  0.1905652
#> 3   1.113073187 -1.25860593  1.0609389   2     X1 -0.0682294
#> 4   1.113073187 -1.25860593  1.0609389   2     X2  0.5329100
#> 5   0.045688843  0.33821518 -0.2158571   3     X1 -0.2249671
#> 6   0.045688843  0.33821518 -0.2158571   3     X2 -0.6904481
#> 7  -1.310635300 -0.12088766 -0.3017466   4     X1 -0.6405419
#> 8  -1.310635300 -0.12088766 -0.3017466   4     X2 -0.3751325
#> 9   1.002337750  0.85255521 -0.3243616   5     X1 -1.4239651
#> 10  1.002337750  0.85255521 -0.3243616   5     X2  2.0900773
#> 11  0.003822226  1.14658913 -0.1821307   6     X1 -1.4149830
#> 12  0.003822226  1.14658913 -0.1821307   6     X2 -0.3419879
#> 13  0.154489750  1.45825498 -0.9641352   7     X1 -0.1651080
#> 14  0.154489750  1.45825498 -0.9641352   7     X2  0.5731969
#> 15  0.404348293  1.14911085 -1.3584176   8     X1  0.7728172
#> 16  0.404348293  1.14911085 -1.3584176   8     X2 -0.8729141
#> 17 -0.687282145  1.30346952 -3.2486268   9     X1 -0.6307017
#> 18 -0.687282145  1.30346952 -3.2486268   9     X2 -0.4637872
#> 19 -1.047515857 -0.21519968 -0.7891292  10     X1  0.4201631
#> 20 -1.047515857 -0.21519968 -0.7891292  10     X2  0.9503134

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1   Value_X2     Value_X3    Value_X4   Value_X5
#> 1       1 -0.4129476  0.1905652  0.033042687  0.02630021  0.8997272
#> 2       2 -0.0682294  0.5329100  1.113073187 -1.25860593  1.0609389
#> 3       3 -0.2249671 -0.6904481  0.045688843  0.33821518 -0.2158571
#> 4       4 -0.6405419 -0.3751325 -1.310635300 -0.12088766 -0.3017466
#> 5       5 -1.4239651  2.0900773  1.002337750  0.85255521 -0.3243616
#> 6       6 -1.4149830 -0.3419879  0.003822226  1.14658913 -0.1821307
#> 7       7 -0.1651080  0.5731969  0.154489750  1.45825498 -0.9641352
#> 8       8  0.7728172 -0.8729141  0.404348293  1.14911085 -1.3584176
#> 9       9 -0.6307017 -0.4637872 -0.687282145  1.30346952 -3.2486268
#> 10     10  0.4201631  0.9503134 -1.047515857 -0.21519968 -0.7891292

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5