Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1 -0.650901669
#> 2    X2  2.343391368
#> 3    X3 -0.580805033
#> 4    X4 -2.113240146
#> 5    X5  1.182755249
#> 6    X1  2.617845670
#> 7    X2 -2.462671117
#> 8    X3 -0.093532179
#> 9    X4  0.350277674
#> 10   X5 -0.289266857
#> 11   X1 -0.671777421
#> 12   X2 -1.969431504
#> 13   X3  0.415586575
#> 14   X4  0.654373033
#> 15   X5  0.066851923
#> 16   X1  1.192005255
#> 17   X2 -0.215112438
#> 18   X3 -0.056437973
#> 19   X4 -1.034072065
#> 20   X5  0.659935768
#> 21   X1  0.495417113
#> 22   X2  0.007216844
#> 23   X3 -0.148765640
#> 24   X4  1.249763503
#> 25   X5 -0.473825990
#> 26   X1  2.018323460
#> 27   X2 -0.053825882
#> 28   X3  0.465465821
#> 29   X4 -0.593048051
#> 30   X5  0.401918002
#> 31   X1 -0.974312448
#> 32   X2 -0.858423303
#> 33   X3 -1.023501541
#> 34   X4 -0.150781213
#> 35   X5 -1.278197389
#> 36   X1 -0.065771000
#> 37   X2 -1.079111363
#> 38   X3  0.651091567
#> 39   X4  2.745868055
#> 40   X5 -0.161608072
#> 41   X1  0.545568283
#> 42   X2  2.150975505
#> 43   X3  1.073499819
#> 44   X4 -1.226277264
#> 45   X5  2.092613643
#> 46   X1  0.694360079
#> 47   X2 -0.049323315
#> 48   X3 -0.409819739
#> 49   X4  0.086994334
#> 50   X5 -0.957509137

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4          X5 Row Column      Numbers
#> 1  -0.58080503 -2.11324015  1.18275525   1     X1 -0.650901669
#> 2  -0.58080503 -2.11324015  1.18275525   1     X2  2.343391368
#> 3  -0.09353218  0.35027767 -0.28926686   2     X1  2.617845670
#> 4  -0.09353218  0.35027767 -0.28926686   2     X2 -2.462671117
#> 5   0.41558657  0.65437303  0.06685192   3     X1 -0.671777421
#> 6   0.41558657  0.65437303  0.06685192   3     X2 -1.969431504
#> 7  -0.05643797 -1.03407206  0.65993577   4     X1  1.192005255
#> 8  -0.05643797 -1.03407206  0.65993577   4     X2 -0.215112438
#> 9  -0.14876564  1.24976350 -0.47382599   5     X1  0.495417113
#> 10 -0.14876564  1.24976350 -0.47382599   5     X2  0.007216844
#> 11  0.46546582 -0.59304805  0.40191800   6     X1  2.018323460
#> 12  0.46546582 -0.59304805  0.40191800   6     X2 -0.053825882
#> 13 -1.02350154 -0.15078121 -1.27819739   7     X1 -0.974312448
#> 14 -1.02350154 -0.15078121 -1.27819739   7     X2 -0.858423303
#> 15  0.65109157  2.74586806 -0.16160807   8     X1 -0.065771000
#> 16  0.65109157  2.74586806 -0.16160807   8     X2 -1.079111363
#> 17  1.07349982 -1.22627726  2.09261364   9     X1  0.545568283
#> 18  1.07349982 -1.22627726  2.09261364   9     X2  2.150975505
#> 19 -0.40981974  0.08699433 -0.95750914  10     X1  0.694360079
#> 20 -0.40981974  0.08699433 -0.95750914  10     X2 -0.049323315

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1     Value_X2    Value_X3    Value_X4    Value_X5
#> 1       1 -0.6509017  2.343391368 -0.58080503 -2.11324015  1.18275525
#> 2       2  2.6178457 -2.462671117 -0.09353218  0.35027767 -0.28926686
#> 3       3 -0.6717774 -1.969431504  0.41558657  0.65437303  0.06685192
#> 4       4  1.1920053 -0.215112438 -0.05643797 -1.03407206  0.65993577
#> 5       5  0.4954171  0.007216844 -0.14876564  1.24976350 -0.47382599
#> 6       6  2.0183235 -0.053825882  0.46546582 -0.59304805  0.40191800
#> 7       7 -0.9743124 -0.858423303 -1.02350154 -0.15078121 -1.27819739
#> 8       8 -0.0657710 -1.079111363  0.65109157  2.74586806 -0.16160807
#> 9       9  0.5455683  2.150975505  1.07349982 -1.22627726  2.09261364
#> 10     10  0.6943601 -0.049323315 -0.40981974  0.08699433 -0.95750914

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5