Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1 -1.37900646
#> 2    X2 -0.57980048
#> 3    X3  0.17974992
#> 4    X4  0.45699572
#> 5    X5  2.07717182
#> 6    X1 -0.06384330
#> 7    X2  0.57096831
#> 8    X3  0.44206498
#> 9    X4  0.46282543
#> 10   X5  0.48118318
#> 11   X1  0.57759805
#> 12   X2 -1.62219080
#> 13   X3  0.77913928
#> 14   X4  1.82375903
#> 15   X5 -0.26983524
#> 16   X1 -0.19013407
#> 17   X2  1.18920848
#> 18   X3 -0.80617490
#> 19   X4 -0.97623736
#> 20   X5 -0.87304328
#> 21   X1  1.01148802
#> 22   X2  0.57816554
#> 23   X3 -0.45342723
#> 24   X4  0.17668810
#> 25   X5  0.80838848
#> 26   X1 -1.31968226
#> 27   X2  0.09675303
#> 28   X3 -0.31085501
#> 29   X4  0.45502102
#> 30   X5 -0.10508257
#> 31   X1  1.01112131
#> 32   X2 -0.66664623
#> 33   X3 -0.40606958
#> 34   X4 -1.25889269
#> 35   X5 -0.32306047
#> 36   X1  0.89416769
#> 37   X2  1.01411985
#> 38   X3 -0.24922872
#> 39   X4 -1.98403034
#> 40   X5 -1.76369333
#> 41   X1  1.45237260
#> 42   X2  0.68752493
#> 43   X3 -0.58407685
#> 44   X4  0.06398312
#> 45   X5  0.18838185
#> 46   X1  0.09832482
#> 47   X2 -0.42345039
#> 48   X3 -1.34224616
#> 49   X4 -0.03101575
#> 50   X5  0.82031359

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3          X4         X5 Row Column     Numbers
#> 1   0.1797499  0.45699572  2.0771718   1     X1 -1.37900646
#> 2   0.1797499  0.45699572  2.0771718   1     X2 -0.57980048
#> 3   0.4420650  0.46282543  0.4811832   2     X1 -0.06384330
#> 4   0.4420650  0.46282543  0.4811832   2     X2  0.57096831
#> 5   0.7791393  1.82375903 -0.2698352   3     X1  0.57759805
#> 6   0.7791393  1.82375903 -0.2698352   3     X2 -1.62219080
#> 7  -0.8061749 -0.97623736 -0.8730433   4     X1 -0.19013407
#> 8  -0.8061749 -0.97623736 -0.8730433   4     X2  1.18920848
#> 9  -0.4534272  0.17668810  0.8083885   5     X1  1.01148802
#> 10 -0.4534272  0.17668810  0.8083885   5     X2  0.57816554
#> 11 -0.3108550  0.45502102 -0.1050826   6     X1 -1.31968226
#> 12 -0.3108550  0.45502102 -0.1050826   6     X2  0.09675303
#> 13 -0.4060696 -1.25889269 -0.3230605   7     X1  1.01112131
#> 14 -0.4060696 -1.25889269 -0.3230605   7     X2 -0.66664623
#> 15 -0.2492287 -1.98403034 -1.7636933   8     X1  0.89416769
#> 16 -0.2492287 -1.98403034 -1.7636933   8     X2  1.01411985
#> 17 -0.5840768  0.06398312  0.1883818   9     X1  1.45237260
#> 18 -0.5840768  0.06398312  0.1883818   9     X2  0.68752493
#> 19 -1.3422462 -0.03101575  0.8203136  10     X1  0.09832482
#> 20 -1.3422462 -0.03101575  0.8203136  10     X2 -0.42345039

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2   Value_X3    Value_X4   Value_X5
#> 1       1 -1.37900646 -0.57980048  0.1797499  0.45699572  2.0771718
#> 2       2 -0.06384330  0.57096831  0.4420650  0.46282543  0.4811832
#> 3       3  0.57759805 -1.62219080  0.7791393  1.82375903 -0.2698352
#> 4       4 -0.19013407  1.18920848 -0.8061749 -0.97623736 -0.8730433
#> 5       5  1.01148802  0.57816554 -0.4534272  0.17668810  0.8083885
#> 6       6 -1.31968226  0.09675303 -0.3108550  0.45502102 -0.1050826
#> 7       7  1.01112131 -0.66664623 -0.4060696 -1.25889269 -0.3230605
#> 8       8  0.89416769  1.01411985 -0.2492287 -1.98403034 -1.7636933
#> 9       9  1.45237260  0.68752493 -0.5840768  0.06398312  0.1883818
#> 10     10  0.09832482 -0.42345039 -1.3422462 -0.03101575  0.8203136

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5