Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1 -1.72480971
#> 2    X2 -0.23542049
#> 3    X3 -0.30924241
#> 4    X4 -0.37315943
#> 5    X5  0.48809512
#> 6    X1  0.30145697
#> 7    X2  0.87296091
#> 8    X3 -0.31226058
#> 9    X4 -1.21778683
#> 10   X5  0.13801942
#> 11   X1 -2.21791488
#> 12   X2 -1.09576824
#> 13   X3 -1.89350219
#> 14   X4  0.26554461
#> 15   X5 -0.33871361
#> 16   X1 -1.46933430
#> 17   X2 -0.94772314
#> 18   X3  1.00901703
#> 19   X4  0.66558116
#> 20   X5  0.96572848
#> 21   X1  1.43058948
#> 22   X2  0.45397041
#> 23   X3 -0.56399662
#> 24   X4  0.29861514
#> 25   X5 -0.08096197
#> 26   X1  1.70553980
#> 27   X2  0.33201657
#> 28   X3 -2.01042342
#> 29   X4 -0.12563732
#> 30   X5 -0.80735993
#> 31   X1  0.00814141
#> 32   X2 -0.05388969
#> 33   X3  0.45637090
#> 34   X4 -0.42446781
#> 35   X5  0.16823065
#> 36   X1 -0.15791466
#> 37   X2 -2.08087027
#> 38   X3  0.90296819
#> 39   X4 -1.06376615
#> 40   X5 -0.62566068
#> 41   X1 -0.39467160
#> 42   X2 -0.09848980
#> 43   X3 -0.63997518
#> 44   X4  0.57466893
#> 45   X5  1.22058242
#> 46   X1 -0.42327977
#> 47   X2 -0.82700291
#> 48   X3  0.82369459
#> 49   X4 -0.11941538
#> 50   X5 -0.05822132

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4          X5 Row Column     Numbers
#> 1  -0.3092424 -0.3731594  0.48809512   1     X1 -1.72480971
#> 2  -0.3092424 -0.3731594  0.48809512   1     X2 -0.23542049
#> 3  -0.3122606 -1.2177868  0.13801942   2     X1  0.30145697
#> 4  -0.3122606 -1.2177868  0.13801942   2     X2  0.87296091
#> 5  -1.8935022  0.2655446 -0.33871361   3     X1 -2.21791488
#> 6  -1.8935022  0.2655446 -0.33871361   3     X2 -1.09576824
#> 7   1.0090170  0.6655812  0.96572848   4     X1 -1.46933430
#> 8   1.0090170  0.6655812  0.96572848   4     X2 -0.94772314
#> 9  -0.5639966  0.2986151 -0.08096197   5     X1  1.43058948
#> 10 -0.5639966  0.2986151 -0.08096197   5     X2  0.45397041
#> 11 -2.0104234 -0.1256373 -0.80735993   6     X1  1.70553980
#> 12 -2.0104234 -0.1256373 -0.80735993   6     X2  0.33201657
#> 13  0.4563709 -0.4244678  0.16823065   7     X1  0.00814141
#> 14  0.4563709 -0.4244678  0.16823065   7     X2 -0.05388969
#> 15  0.9029682 -1.0637662 -0.62566068   8     X1 -0.15791466
#> 16  0.9029682 -1.0637662 -0.62566068   8     X2 -2.08087027
#> 17 -0.6399752  0.5746689  1.22058242   9     X1 -0.39467160
#> 18 -0.6399752  0.5746689  1.22058242   9     X2 -0.09848980
#> 19  0.8236946 -0.1194154 -0.05822132  10     X1 -0.42327977
#> 20  0.8236946 -0.1194154 -0.05822132  10     X2 -0.82700291

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2   Value_X3   Value_X4    Value_X5
#> 1       1 -1.72480971 -0.23542049 -0.3092424 -0.3731594  0.48809512
#> 2       2  0.30145697  0.87296091 -0.3122606 -1.2177868  0.13801942
#> 3       3 -2.21791488 -1.09576824 -1.8935022  0.2655446 -0.33871361
#> 4       4 -1.46933430 -0.94772314  1.0090170  0.6655812  0.96572848
#> 5       5  1.43058948  0.45397041 -0.5639966  0.2986151 -0.08096197
#> 6       6  1.70553980  0.33201657 -2.0104234 -0.1256373 -0.80735993
#> 7       7  0.00814141 -0.05388969  0.4563709 -0.4244678  0.16823065
#> 8       8 -0.15791466 -2.08087027  0.9029682 -1.0637662 -0.62566068
#> 9       9 -0.39467160 -0.09848980 -0.6399752  0.5746689  1.22058242
#> 10     10 -0.42327977 -0.82700291  0.8236946 -0.1194154 -0.05822132

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5