Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1  0.239539060
#> 2    X2 -0.927876706
#> 3    X3  2.098708459
#> 4    X4  0.915555990
#> 5    X5 -0.632662932
#> 6    X1 -0.705062349
#> 7    X2  0.041515672
#> 8    X3  1.565263262
#> 9    X4 -1.087411412
#> 10   X5 -1.368803968
#> 11   X1 -0.939041934
#> 12   X2  0.723019468
#> 13   X3 -0.730488352
#> 14   X4  1.165968810
#> 15   X5 -0.050727315
#> 16   X1 -0.267863598
#> 17   X2  0.070385998
#> 18   X3  0.694812761
#> 19   X4 -0.164950954
#> 20   X5  0.628876065
#> 21   X1 -1.534398802
#> 22   X2 -1.154643454
#> 23   X3  1.820108555
#> 24   X4 -0.998783018
#> 25   X5  0.203711409
#> 26   X1 -0.140484344
#> 27   X2 -0.981236154
#> 28   X3 -0.756490066
#> 29   X4  0.685844013
#> 30   X5 -0.209610402
#> 31   X1 -1.241137406
#> 32   X2  0.011251580
#> 33   X3 -1.139871478
#> 34   X4  0.102385157
#> 35   X5  1.191327723
#> 36   X1 -0.208498814
#> 37   X2  0.189904749
#> 38   X3  0.397581582
#> 39   X4  0.282803125
#> 40   X5 -0.002599895
#> 41   X1  0.656900860
#> 42   X2  0.690348974
#> 43   X3  0.917804562
#> 44   X4 -0.536661622
#> 45   X5 -0.359061837
#> 46   X1 -0.645569618
#> 47   X2 -1.492607308
#> 48   X3 -1.208875713
#> 49   X4  0.051932666
#> 50   X5  0.207817950

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3          X4           X5 Row Column     Numbers
#> 1   2.0987085  0.91555599 -0.632662932   1     X1  0.23953906
#> 2   2.0987085  0.91555599 -0.632662932   1     X2 -0.92787671
#> 3   1.5652633 -1.08741141 -1.368803968   2     X1 -0.70506235
#> 4   1.5652633 -1.08741141 -1.368803968   2     X2  0.04151567
#> 5  -0.7304884  1.16596881 -0.050727315   3     X1 -0.93904193
#> 6  -0.7304884  1.16596881 -0.050727315   3     X2  0.72301947
#> 7   0.6948128 -0.16495095  0.628876065   4     X1 -0.26786360
#> 8   0.6948128 -0.16495095  0.628876065   4     X2  0.07038600
#> 9   1.8201086 -0.99878302  0.203711409   5     X1 -1.53439880
#> 10  1.8201086 -0.99878302  0.203711409   5     X2 -1.15464345
#> 11 -0.7564901  0.68584401 -0.209610402   6     X1 -0.14048434
#> 12 -0.7564901  0.68584401 -0.209610402   6     X2 -0.98123615
#> 13 -1.1398715  0.10238516  1.191327723   7     X1 -1.24113741
#> 14 -1.1398715  0.10238516  1.191327723   7     X2  0.01125158
#> 15  0.3975816  0.28280312 -0.002599895   8     X1 -0.20849881
#> 16  0.3975816  0.28280312 -0.002599895   8     X2  0.18990475
#> 17  0.9178046 -0.53666162 -0.359061837   9     X1  0.65690086
#> 18  0.9178046 -0.53666162 -0.359061837   9     X2  0.69034897
#> 19 -1.2088757  0.05193267  0.207817950  10     X1 -0.64556962
#> 20 -1.2088757  0.05193267  0.207817950  10     X2 -1.49260731

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2   Value_X3    Value_X4     Value_X5
#> 1       1  0.2395391 -0.92787671  2.0987085  0.91555599 -0.632662932
#> 2       2 -0.7050623  0.04151567  1.5652633 -1.08741141 -1.368803968
#> 3       3 -0.9390419  0.72301947 -0.7304884  1.16596881 -0.050727315
#> 4       4 -0.2678636  0.07038600  0.6948128 -0.16495095  0.628876065
#> 5       5 -1.5343988 -1.15464345  1.8201086 -0.99878302  0.203711409
#> 6       6 -0.1404843 -0.98123615 -0.7564901  0.68584401 -0.209610402
#> 7       7 -1.2411374  0.01125158 -1.1398715  0.10238516  1.191327723
#> 8       8 -0.2084988  0.18990475  0.3975816  0.28280312 -0.002599895
#> 9       9  0.6569009  0.69034897  0.9178046 -0.53666162 -0.359061837
#> 10     10 -0.6455696 -1.49260731 -1.2088757  0.05193267  0.207817950

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5