Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  0.93397427
#> 2    X2 -0.06664828
#> 3    X3  0.49334391
#> 4    X4 -1.77290123
#> 5    X5  0.44942466
#> 6    X1  0.29607688
#> 7    X2  0.45025845
#> 8    X3 -0.07409649
#> 9    X4 -0.71578800
#> 10   X5  1.13401006
#> 11   X1 -0.25018248
#> 12   X2  1.23371252
#> 13   X3 -0.56496172
#> 14   X4  0.04031461
#> 15   X5 -3.45759236
#> 16   X1 -1.10609722
#> 17   X2 -0.35713032
#> 18   X3  0.19283881
#> 19   X4 -0.71182538
#> 20   X5 -1.28970591
#> 21   X1  0.96883056
#> 22   X2  1.88217260
#> 23   X3  0.36223548
#> 24   X4 -1.38737744
#> 25   X5  0.50261139
#> 26   X1 -0.94919694
#> 27   X2  0.44329202
#> 28   X3  0.70477974
#> 29   X4  0.58820766
#> 30   X5  0.08704074
#> 31   X1  0.51071161
#> 32   X2 -0.13829773
#> 33   X3  0.41784980
#> 34   X4 -0.11815729
#> 35   X5 -2.34432680
#> 36   X1  0.13270040
#> 37   X2 -1.00236603
#> 38   X3  0.48135328
#> 39   X4 -1.64176056
#> 40   X5  0.81600713
#> 41   X1 -1.11072762
#> 42   X2  0.22104906
#> 43   X3  1.64561747
#> 44   X4 -0.46743876
#> 45   X5  2.98708075
#> 46   X1  0.32705011
#> 47   X2 -1.50567872
#> 48   X3 -1.16244284
#> 49   X4 -0.11676792
#> 50   X5  3.20464966

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4          X5 Row Column     Numbers
#> 1   0.49334391 -1.77290123  0.44942466   1     X1  0.93397427
#> 2   0.49334391 -1.77290123  0.44942466   1     X2 -0.06664828
#> 3  -0.07409649 -0.71578800  1.13401006   2     X1  0.29607688
#> 4  -0.07409649 -0.71578800  1.13401006   2     X2  0.45025845
#> 5  -0.56496172  0.04031461 -3.45759236   3     X1 -0.25018248
#> 6  -0.56496172  0.04031461 -3.45759236   3     X2  1.23371252
#> 7   0.19283881 -0.71182538 -1.28970591   4     X1 -1.10609722
#> 8   0.19283881 -0.71182538 -1.28970591   4     X2 -0.35713032
#> 9   0.36223548 -1.38737744  0.50261139   5     X1  0.96883056
#> 10  0.36223548 -1.38737744  0.50261139   5     X2  1.88217260
#> 11  0.70477974  0.58820766  0.08704074   6     X1 -0.94919694
#> 12  0.70477974  0.58820766  0.08704074   6     X2  0.44329202
#> 13  0.41784980 -0.11815729 -2.34432680   7     X1  0.51071161
#> 14  0.41784980 -0.11815729 -2.34432680   7     X2 -0.13829773
#> 15  0.48135328 -1.64176056  0.81600713   8     X1  0.13270040
#> 16  0.48135328 -1.64176056  0.81600713   8     X2 -1.00236603
#> 17  1.64561747 -0.46743876  2.98708075   9     X1 -1.11072762
#> 18  1.64561747 -0.46743876  2.98708075   9     X2  0.22104906
#> 19 -1.16244284 -0.11676792  3.20464966  10     X1  0.32705011
#> 20 -1.16244284 -0.11676792  3.20464966  10     X2 -1.50567872

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2    Value_X3    Value_X4    Value_X5
#> 1       1  0.9339743 -0.06664828  0.49334391 -1.77290123  0.44942466
#> 2       2  0.2960769  0.45025845 -0.07409649 -0.71578800  1.13401006
#> 3       3 -0.2501825  1.23371252 -0.56496172  0.04031461 -3.45759236
#> 4       4 -1.1060972 -0.35713032  0.19283881 -0.71182538 -1.28970591
#> 5       5  0.9688306  1.88217260  0.36223548 -1.38737744  0.50261139
#> 6       6 -0.9491969  0.44329202  0.70477974  0.58820766  0.08704074
#> 7       7  0.5107116 -0.13829773  0.41784980 -0.11815729 -2.34432680
#> 8       8  0.1327004 -1.00236603  0.48135328 -1.64176056  0.81600713
#> 9       9 -1.1107276  0.22104906  1.64561747 -0.46743876  2.98708075
#> 10     10  0.3270501 -1.50567872 -1.16244284 -0.11676792  3.20464966

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5