Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1  0.610979436
#> 2    X2  1.045761927
#> 3    X3 -0.708935050
#> 4    X4 -0.847981432
#> 5    X5  0.436425872
#> 6    X1 -0.270800087
#> 7    X2 -0.466006463
#> 8    X3 -0.390638354
#> 9    X4  1.638572142
#> 10   X5  0.664057163
#> 11   X1  0.671782339
#> 12   X2 -2.458947821
#> 13   X3  1.815968512
#> 14   X4  0.568732102
#> 15   X5  0.980681446
#> 16   X1  0.209624522
#> 17   X2  1.064695425
#> 18   X3  1.424277544
#> 19   X4  0.927953357
#> 20   X5  1.180407675
#> 21   X1  0.332328600
#> 22   X2  0.266879606
#> 23   X3  0.734877704
#> 24   X4 -0.033058781
#> 25   X5  0.230736089
#> 26   X1 -0.067354510
#> 27   X2  1.696495321
#> 28   X3 -0.005450332
#> 29   X4  0.358158076
#> 30   X5  1.022611106
#> 31   X1  0.517307838
#> 32   X2  1.163088441
#> 33   X3 -0.910119655
#> 34   X4 -0.355649589
#> 35   X5  0.506781525
#> 36   X1  1.698718497
#> 37   X2 -0.031548562
#> 38   X3  0.441233654
#> 39   X4  0.215812776
#> 40   X5 -1.568096979
#> 41   X1  0.544713718
#> 42   X2 -1.285702687
#> 43   X3  2.154838284
#> 44   X4 -0.030911614
#> 45   X5 -0.444918821
#> 46   X1 -1.566505840
#> 47   X2  1.829345833
#> 48   X3 -0.402044839
#> 49   X4  0.239857217
#> 50   X5 -0.853830190

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>              X3          X4         X5 Row Column     Numbers
#> 1  -0.708935050 -0.84798143  0.4364259   1     X1  0.61097944
#> 2  -0.708935050 -0.84798143  0.4364259   1     X2  1.04576193
#> 3  -0.390638354  1.63857214  0.6640572   2     X1 -0.27080009
#> 4  -0.390638354  1.63857214  0.6640572   2     X2 -0.46600646
#> 5   1.815968512  0.56873210  0.9806814   3     X1  0.67178234
#> 6   1.815968512  0.56873210  0.9806814   3     X2 -2.45894782
#> 7   1.424277544  0.92795336  1.1804077   4     X1  0.20962452
#> 8   1.424277544  0.92795336  1.1804077   4     X2  1.06469543
#> 9   0.734877704 -0.03305878  0.2307361   5     X1  0.33232860
#> 10  0.734877704 -0.03305878  0.2307361   5     X2  0.26687961
#> 11 -0.005450332  0.35815808  1.0226111   6     X1 -0.06735451
#> 12 -0.005450332  0.35815808  1.0226111   6     X2  1.69649532
#> 13 -0.910119655 -0.35564959  0.5067815   7     X1  0.51730784
#> 14 -0.910119655 -0.35564959  0.5067815   7     X2  1.16308844
#> 15  0.441233654  0.21581278 -1.5680970   8     X1  1.69871850
#> 16  0.441233654  0.21581278 -1.5680970   8     X2 -0.03154856
#> 17  2.154838284 -0.03091161 -0.4449188   9     X1  0.54471372
#> 18  2.154838284 -0.03091161 -0.4449188   9     X2 -1.28570269
#> 19 -0.402044839  0.23985722 -0.8538302  10     X1 -1.56650584
#> 20 -0.402044839  0.23985722 -0.8538302  10     X2  1.82934583

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2     Value_X3    Value_X4   Value_X5
#> 1       1  0.61097944  1.04576193 -0.708935050 -0.84798143  0.4364259
#> 2       2 -0.27080009 -0.46600646 -0.390638354  1.63857214  0.6640572
#> 3       3  0.67178234 -2.45894782  1.815968512  0.56873210  0.9806814
#> 4       4  0.20962452  1.06469543  1.424277544  0.92795336  1.1804077
#> 5       5  0.33232860  0.26687961  0.734877704 -0.03305878  0.2307361
#> 6       6 -0.06735451  1.69649532 -0.005450332  0.35815808  1.0226111
#> 7       7  0.51730784  1.16308844 -0.910119655 -0.35564959  0.5067815
#> 8       8  1.69871850 -0.03154856  0.441233654  0.21581278 -1.5680970
#> 9       9  0.54471372 -1.28570269  2.154838284 -0.03091161 -0.4449188
#> 10     10 -1.56650584  1.82934583 -0.402044839  0.23985722 -0.8538302

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5