Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name         Value
#> 1    X1 -0.1432485271
#> 2    X2  0.7042309343
#> 3    X3 -0.2400421733
#> 4    X4  0.4823757602
#> 5    X5 -0.4091980078
#> 6    X1 -1.1275651153
#> 7    X2  1.2017982101
#> 8    X3 -0.5595326553
#> 9    X4  1.6532078679
#> 10   X5 -1.0070652116
#> 11   X1  0.7629740717
#> 12   X2  2.2352488036
#> 13   X3 -0.0008402205
#> 14   X4  0.4239906635
#> 15   X5 -2.3930213743
#> 16   X1 -1.1401029668
#> 17   X2  0.6201796861
#> 18   X3  1.9212003992
#> 19   X4  2.0987819234
#> 20   X5  0.4374673418
#> 21   X1 -0.1807870655
#> 22   X2  0.8912020409
#> 23   X3  0.6522021859
#> 24   X4  1.5244497158
#> 25   X5  0.9074729040
#> 26   X1  1.5523133542
#> 27   X2  0.7595854204
#> 28   X3  0.9738369890
#> 29   X4  0.4749856634
#> 30   X5 -0.7581426266
#> 31   X1  1.5487377690
#> 32   X2  2.1425464479
#> 33   X3 -0.3105836495
#> 34   X4  0.5954707444
#> 35   X5  0.5665630856
#> 36   X1  0.3787512793
#> 37   X2  0.8873636129
#> 38   X3 -0.6318503751
#> 39   X4 -0.3139545267
#> 40   X5  0.9628444774
#> 41   X1 -1.3506999615
#> 42   X2 -0.8685157666
#> 43   X3  0.1063923773
#> 44   X4  0.5150715661
#> 45   X5 -2.1242203750
#> 46   X1  0.7292599615
#> 47   X2  0.6087858516
#> 48   X3 -0.7741013206
#> 49   X4  1.2145517909
#> 50   X5  0.7972324604

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>               X3         X4         X5 Row Column    Numbers
#> 1  -0.2400421733  0.4823758 -0.4091980   1     X1 -0.1432485
#> 2  -0.2400421733  0.4823758 -0.4091980   1     X2  0.7042309
#> 3  -0.5595326553  1.6532079 -1.0070652   2     X1 -1.1275651
#> 4  -0.5595326553  1.6532079 -1.0070652   2     X2  1.2017982
#> 5  -0.0008402205  0.4239907 -2.3930214   3     X1  0.7629741
#> 6  -0.0008402205  0.4239907 -2.3930214   3     X2  2.2352488
#> 7   1.9212003992  2.0987819  0.4374673   4     X1 -1.1401030
#> 8   1.9212003992  2.0987819  0.4374673   4     X2  0.6201797
#> 9   0.6522021859  1.5244497  0.9074729   5     X1 -0.1807871
#> 10  0.6522021859  1.5244497  0.9074729   5     X2  0.8912020
#> 11  0.9738369890  0.4749857 -0.7581426   6     X1  1.5523134
#> 12  0.9738369890  0.4749857 -0.7581426   6     X2  0.7595854
#> 13 -0.3105836495  0.5954707  0.5665631   7     X1  1.5487378
#> 14 -0.3105836495  0.5954707  0.5665631   7     X2  2.1425464
#> 15 -0.6318503751 -0.3139545  0.9628445   8     X1  0.3787513
#> 16 -0.6318503751 -0.3139545  0.9628445   8     X2  0.8873636
#> 17  0.1063923773  0.5150716 -2.1242204   9     X1 -1.3507000
#> 18  0.1063923773  0.5150716 -2.1242204   9     X2 -0.8685158
#> 19 -0.7741013206  1.2145518  0.7972325  10     X1  0.7292600
#> 20 -0.7741013206  1.2145518  0.7972325  10     X2  0.6087859

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1   Value_X2      Value_X3   Value_X4   Value_X5
#> 1       1 -0.1432485  0.7042309 -0.2400421733  0.4823758 -0.4091980
#> 2       2 -1.1275651  1.2017982 -0.5595326553  1.6532079 -1.0070652
#> 3       3  0.7629741  2.2352488 -0.0008402205  0.4239907 -2.3930214
#> 4       4 -1.1401030  0.6201797  1.9212003992  2.0987819  0.4374673
#> 5       5 -0.1807871  0.8912020  0.6522021859  1.5244497  0.9074729
#> 6       6  1.5523134  0.7595854  0.9738369890  0.4749857 -0.7581426
#> 7       7  1.5487378  2.1425464 -0.3105836495  0.5954707  0.5665631
#> 8       8  0.3787513  0.8873636 -0.6318503751 -0.3139545  0.9628445
#> 9       9 -1.3507000 -0.8685158  0.1063923773  0.5150716 -2.1242204
#> 10     10  0.7292600  0.6087859 -0.7741013206  1.2145518  0.7972325

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5