Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1 -0.745433889
#> 2    X2 -1.135435337
#> 3    X3 -2.934800044
#> 4    X4 -0.376312259
#> 5    X5 -1.537157715
#> 6    X1 -0.049106700
#> 7    X2 -1.521694113
#> 8    X3  0.145120747
#> 9    X4 -1.700291838
#> 10   X5 -0.264739311
#> 11   X1 -0.065548872
#> 12   X2  0.001494434
#> 13   X3 -1.242034014
#> 14   X4  0.953370473
#> 15   X5 -1.402411892
#> 16   X1  1.893964671
#> 17   X2 -0.549188401
#> 18   X3  2.808129333
#> 19   X4 -0.306236909
#> 20   X5  1.566479887
#> 21   X1  0.821298448
#> 22   X2  0.153903211
#> 23   X3 -0.589490247
#> 24   X4  1.415676208
#> 25   X5  0.804099919
#> 26   X1  1.109734707
#> 27   X2  1.078088253
#> 28   X3  0.777103028
#> 29   X4 -2.101724387
#> 30   X5 -0.831140723
#> 31   X1 -0.026070160
#> 32   X2 -0.100445871
#> 33   X3 -0.745036090
#> 34   X4  0.446425847
#> 35   X5  2.184963090
#> 36   X1  0.282657861
#> 37   X2  0.078037101
#> 38   X3 -0.109642986
#> 39   X4 -0.591436266
#> 40   X5  0.330531987
#> 41   X1  0.155164091
#> 42   X2  0.287437732
#> 43   X3  0.550777412
#> 44   X4 -0.969667642
#> 45   X5  1.049583341
#> 46   X1  0.667457554
#> 47   X2  0.112249495
#> 48   X3  0.561288482
#> 49   X4 -0.920222018
#> 50   X5 -1.088024102

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4         X5 Row Column      Numbers
#> 1  -2.9348000 -0.3763123 -1.5371577   1     X1 -0.745433889
#> 2  -2.9348000 -0.3763123 -1.5371577   1     X2 -1.135435337
#> 3   0.1451207 -1.7002918 -0.2647393   2     X1 -0.049106700
#> 4   0.1451207 -1.7002918 -0.2647393   2     X2 -1.521694113
#> 5  -1.2420340  0.9533705 -1.4024119   3     X1 -0.065548872
#> 6  -1.2420340  0.9533705 -1.4024119   3     X2  0.001494434
#> 7   2.8081293 -0.3062369  1.5664799   4     X1  1.893964671
#> 8   2.8081293 -0.3062369  1.5664799   4     X2 -0.549188401
#> 9  -0.5894902  1.4156762  0.8040999   5     X1  0.821298448
#> 10 -0.5894902  1.4156762  0.8040999   5     X2  0.153903211
#> 11  0.7771030 -2.1017244 -0.8311407   6     X1  1.109734707
#> 12  0.7771030 -2.1017244 -0.8311407   6     X2  1.078088253
#> 13 -0.7450361  0.4464258  2.1849631   7     X1 -0.026070160
#> 14 -0.7450361  0.4464258  2.1849631   7     X2 -0.100445871
#> 15 -0.1096430 -0.5914363  0.3305320   8     X1  0.282657861
#> 16 -0.1096430 -0.5914363  0.3305320   8     X2  0.078037101
#> 17  0.5507774 -0.9696676  1.0495833   9     X1  0.155164091
#> 18  0.5507774 -0.9696676  1.0495833   9     X2  0.287437732
#> 19  0.5612885 -0.9202220 -1.0880241  10     X1  0.667457554
#> 20  0.5612885 -0.9202220 -1.0880241  10     X2  0.112249495

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1     Value_X2   Value_X3   Value_X4   Value_X5
#> 1       1 -0.74543389 -1.135435337 -2.9348000 -0.3763123 -1.5371577
#> 2       2 -0.04910670 -1.521694113  0.1451207 -1.7002918 -0.2647393
#> 3       3 -0.06554887  0.001494434 -1.2420340  0.9533705 -1.4024119
#> 4       4  1.89396467 -0.549188401  2.8081293 -0.3062369  1.5664799
#> 5       5  0.82129845  0.153903211 -0.5894902  1.4156762  0.8040999
#> 6       6  1.10973471  1.078088253  0.7771030 -2.1017244 -0.8311407
#> 7       7 -0.02607016 -0.100445871 -0.7450361  0.4464258  2.1849631
#> 8       8  0.28265786  0.078037101 -0.1096430 -0.5914363  0.3305320
#> 9       9  0.15516409  0.287437732  0.5507774 -0.9696676  1.0495833
#> 10     10  0.66745755  0.112249495  0.5612885 -0.9202220 -1.0880241

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5