Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  2.25215051
#> 2    X2 -0.54870587
#> 3    X3 -1.96546704
#> 4    X4  0.09274927
#> 5    X5 -2.27608392
#> 6    X1 -0.69343806
#> 7    X2 -1.14835367
#> 8    X3 -1.49097182
#> 9    X4  0.87254639
#> 10   X5  0.72874304
#> 11   X1 -0.88650552
#> 12   X2  1.19872207
#> 13   X3 -0.01616382
#> 14   X4 -1.46690279
#> 15   X5  0.33593992
#> 16   X1 -0.52504719
#> 17   X2 -0.94708957
#> 18   X3 -0.10182999
#> 19   X4  0.72163841
#> 20   X5 -0.16553371
#> 21   X1  1.52309067
#> 22   X2 -0.08465994
#> 23   X3  1.39186775
#> 24   X4  1.60435296
#> 25   X5  0.15289609
#> 26   X1 -0.28137033
#> 27   X2  0.30358199
#> 28   X3  0.32320491
#> 29   X4 -0.55106281
#> 30   X5 -1.07804797
#> 31   X1 -0.66006885
#> 32   X2 -0.18428405
#> 33   X3 -0.40555108
#> 34   X4  0.83363159
#> 35   X5  0.07722016
#> 36   X1 -0.75040402
#> 37   X2 -0.37610327
#> 38   X3  2.34071682
#> 39   X4 -1.44827788
#> 40   X5 -0.66461177
#> 41   X1  0.60687467
#> 42   X2  1.80232304
#> 43   X3 -1.31267168
#> 44   X4  1.23825269
#> 45   X5 -0.74278474
#> 46   X1  0.61443570
#> 47   X2  0.76792096
#> 48   X3  1.62509672
#> 49   X4 -0.70094456
#> 50   X5  0.93580551

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4          X5 Row Column     Numbers
#> 1  -1.96546704  0.09274927 -2.27608392   1     X1  2.25215051
#> 2  -1.96546704  0.09274927 -2.27608392   1     X2 -0.54870587
#> 3  -1.49097182  0.87254639  0.72874304   2     X1 -0.69343806
#> 4  -1.49097182  0.87254639  0.72874304   2     X2 -1.14835367
#> 5  -0.01616382 -1.46690279  0.33593992   3     X1 -0.88650552
#> 6  -0.01616382 -1.46690279  0.33593992   3     X2  1.19872207
#> 7  -0.10182999  0.72163841 -0.16553371   4     X1 -0.52504719
#> 8  -0.10182999  0.72163841 -0.16553371   4     X2 -0.94708957
#> 9   1.39186775  1.60435296  0.15289609   5     X1  1.52309067
#> 10  1.39186775  1.60435296  0.15289609   5     X2 -0.08465994
#> 11  0.32320491 -0.55106281 -1.07804797   6     X1 -0.28137033
#> 12  0.32320491 -0.55106281 -1.07804797   6     X2  0.30358199
#> 13 -0.40555108  0.83363159  0.07722016   7     X1 -0.66006885
#> 14 -0.40555108  0.83363159  0.07722016   7     X2 -0.18428405
#> 15  2.34071682 -1.44827788 -0.66461177   8     X1 -0.75040402
#> 16  2.34071682 -1.44827788 -0.66461177   8     X2 -0.37610327
#> 17 -1.31267168  1.23825269 -0.74278474   9     X1  0.60687467
#> 18 -1.31267168  1.23825269 -0.74278474   9     X2  1.80232304
#> 19  1.62509672 -0.70094456  0.93580551  10     X1  0.61443570
#> 20  1.62509672 -0.70094456  0.93580551  10     X2  0.76792096

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2    Value_X3    Value_X4    Value_X5
#> 1       1  2.2521505 -0.54870587 -1.96546704  0.09274927 -2.27608392
#> 2       2 -0.6934381 -1.14835367 -1.49097182  0.87254639  0.72874304
#> 3       3 -0.8865055  1.19872207 -0.01616382 -1.46690279  0.33593992
#> 4       4 -0.5250472 -0.94708957 -0.10182999  0.72163841 -0.16553371
#> 5       5  1.5230907 -0.08465994  1.39186775  1.60435296  0.15289609
#> 6       6 -0.2813703  0.30358199  0.32320491 -0.55106281 -1.07804797
#> 7       7 -0.6600689 -0.18428405 -0.40555108  0.83363159  0.07722016
#> 8       8 -0.7504040 -0.37610327  2.34071682 -1.44827788 -0.66461177
#> 9       9  0.6068747  1.80232304 -1.31267168  1.23825269 -0.74278474
#> 10     10  0.6144357  0.76792096  1.62509672 -0.70094456  0.93580551

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5