Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  1.42551737
#> 2    X2  1.22014561
#> 3    X3 -0.32788777
#> 4    X4  0.36127367
#> 5    X5 -0.99322354
#> 6    X1  0.71736604
#> 7    X2 -1.33909094
#> 8    X3  0.93643008
#> 9    X4 -0.69335621
#> 10   X5 -1.30577626
#> 11   X1 -1.13414872
#> 12   X2 -0.32121807
#> 13   X3 -0.33478054
#> 14   X4  0.64042915
#> 15   X5 -0.78119152
#> 16   X1 -1.00940145
#> 17   X2 -0.27959562
#> 18   X3  0.28543808
#> 19   X4 -0.36023218
#> 20   X5 -0.38291775
#> 21   X1 -0.15289546
#> 22   X2  0.70216738
#> 23   X3  0.67595057
#> 24   X4  0.19114845
#> 25   X5 -1.61218968
#> 26   X1  0.09269266
#> 27   X2  0.39975928
#> 28   X3 -1.51740626
#> 29   X4  0.64871529
#> 30   X5  0.42001809
#> 31   X1  1.14113081
#> 32   X2 -0.13539443
#> 33   X3  0.29146243
#> 34   X4  0.09611514
#> 35   X5 -2.41553810
#> 36   X1 -0.78986070
#> 37   X2 -0.71240022
#> 38   X3  0.70392089
#> 39   X4 -0.70590516
#> 40   X5  1.13368420
#> 41   X1 -0.69545708
#> 42   X2 -3.07957241
#> 43   X3 -0.71328519
#> 44   X4  1.52440376
#> 45   X5 -1.01725567
#> 46   X1 -0.12945773
#> 47   X2 -0.33178299
#> 48   X3 -1.25146505
#> 49   X4 -0.88225278
#> 50   X5  0.41045453

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3          X4         X5 Row Column     Numbers
#> 1  -0.3278878  0.36127367 -0.9932235   1     X1  1.42551737
#> 2  -0.3278878  0.36127367 -0.9932235   1     X2  1.22014561
#> 3   0.9364301 -0.69335621 -1.3057763   2     X1  0.71736604
#> 4   0.9364301 -0.69335621 -1.3057763   2     X2 -1.33909094
#> 5  -0.3347805  0.64042915 -0.7811915   3     X1 -1.13414872
#> 6  -0.3347805  0.64042915 -0.7811915   3     X2 -0.32121807
#> 7   0.2854381 -0.36023218 -0.3829178   4     X1 -1.00940145
#> 8   0.2854381 -0.36023218 -0.3829178   4     X2 -0.27959562
#> 9   0.6759506  0.19114845 -1.6121897   5     X1 -0.15289546
#> 10  0.6759506  0.19114845 -1.6121897   5     X2  0.70216738
#> 11 -1.5174063  0.64871529  0.4200181   6     X1  0.09269266
#> 12 -1.5174063  0.64871529  0.4200181   6     X2  0.39975928
#> 13  0.2914624  0.09611514 -2.4155381   7     X1  1.14113081
#> 14  0.2914624  0.09611514 -2.4155381   7     X2 -0.13539443
#> 15  0.7039209 -0.70590516  1.1336842   8     X1 -0.78986070
#> 16  0.7039209 -0.70590516  1.1336842   8     X2 -0.71240022
#> 17 -0.7132852  1.52440376 -1.0172557   9     X1 -0.69545708
#> 18 -0.7132852  1.52440376 -1.0172557   9     X2 -3.07957241
#> 19 -1.2514651 -0.88225278  0.4104545  10     X1 -0.12945773
#> 20 -1.2514651 -0.88225278  0.4104545  10     X2 -0.33178299

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1   Value_X2   Value_X3    Value_X4   Value_X5
#> 1       1  1.42551737  1.2201456 -0.3278878  0.36127367 -0.9932235
#> 2       2  0.71736604 -1.3390909  0.9364301 -0.69335621 -1.3057763
#> 3       3 -1.13414872 -0.3212181 -0.3347805  0.64042915 -0.7811915
#> 4       4 -1.00940145 -0.2795956  0.2854381 -0.36023218 -0.3829178
#> 5       5 -0.15289546  0.7021674  0.6759506  0.19114845 -1.6121897
#> 6       6  0.09269266  0.3997593 -1.5174063  0.64871529  0.4200181
#> 7       7  1.14113081 -0.1353944  0.2914624  0.09611514 -2.4155381
#> 8       8 -0.78986070 -0.7124002  0.7039209 -0.70590516  1.1336842
#> 9       9 -0.69545708 -3.0795724 -0.7132852  1.52440376 -1.0172557
#> 10     10 -0.12945773 -0.3317830 -1.2514651 -0.88225278  0.4104545

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5