Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1 -0.438234025
#> 2    X2 -1.431917764
#> 3    X3  0.710538936
#> 4    X4  0.027066935
#> 5    X5 -0.675114074
#> 6    X1  0.861161938
#> 7    X2 -1.045797059
#> 8    X3 -0.672367566
#> 9    X4  1.250352944
#> 10   X5  0.916970545
#> 11   X1 -1.284477069
#> 12   X2 -1.036742060
#> 13   X3 -0.725313105
#> 14   X4  1.534141079
#> 15   X5  0.408941443
#> 16   X1  0.395452919
#> 17   X2  0.595118387
#> 18   X3 -0.932703659
#> 19   X4 -0.134861661
#> 20   X5 -0.003659161
#> 21   X1  0.137396997
#> 22   X2  1.252334344
#> 23   X3  1.622259267
#> 24   X4 -0.622296389
#> 25   X5  0.090736987
#> 26   X1 -1.545229483
#> 27   X2  1.049457451
#> 28   X3  1.251254259
#> 29   X4 -0.852870447
#> 30   X5  0.995909238
#> 31   X1 -1.066075216
#> 32   X2  0.026154047
#> 33   X3 -0.778329997
#> 34   X4 -0.232092883
#> 35   X5 -1.380795143
#> 36   X1  0.220982034
#> 37   X2 -0.790153498
#> 38   X3 -0.391952764
#> 39   X4 -1.609895680
#> 40   X5  0.338482515
#> 41   X1  0.287288631
#> 42   X2 -0.368599109
#> 43   X3 -0.794045859
#> 44   X4  0.847579322
#> 45   X5  1.174390484
#> 46   X1  0.809203915
#> 47   X2 -0.504068682
#> 48   X3  0.058454887
#> 49   X4 -0.353716482
#> 50   X5 -1.081652457

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>             X3          X4           X5 Row Column     Numbers
#> 1   0.71053894  0.02706694 -0.675114074   1     X1 -0.43823402
#> 2   0.71053894  0.02706694 -0.675114074   1     X2 -1.43191776
#> 3  -0.67236757  1.25035294  0.916970545   2     X1  0.86116194
#> 4  -0.67236757  1.25035294  0.916970545   2     X2 -1.04579706
#> 5  -0.72531310  1.53414108  0.408941443   3     X1 -1.28447707
#> 6  -0.72531310  1.53414108  0.408941443   3     X2 -1.03674206
#> 7  -0.93270366 -0.13486166 -0.003659161   4     X1  0.39545292
#> 8  -0.93270366 -0.13486166 -0.003659161   4     X2  0.59511839
#> 9   1.62225927 -0.62229639  0.090736987   5     X1  0.13739700
#> 10  1.62225927 -0.62229639  0.090736987   5     X2  1.25233434
#> 11  1.25125426 -0.85287045  0.995909238   6     X1 -1.54522948
#> 12  1.25125426 -0.85287045  0.995909238   6     X2  1.04945745
#> 13 -0.77833000 -0.23209288 -1.380795143   7     X1 -1.06607522
#> 14 -0.77833000 -0.23209288 -1.380795143   7     X2  0.02615405
#> 15 -0.39195276 -1.60989568  0.338482515   8     X1  0.22098203
#> 16 -0.39195276 -1.60989568  0.338482515   8     X2 -0.79015350
#> 17 -0.79404586  0.84757932  1.174390484   9     X1  0.28728863
#> 18 -0.79404586  0.84757932  1.174390484   9     X2 -0.36859911
#> 19  0.05845489 -0.35371648 -1.081652457  10     X1  0.80920392
#> 20  0.05845489 -0.35371648 -1.081652457  10     X2 -0.50406868

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID   Value_X1    Value_X2    Value_X3    Value_X4     Value_X5
#> 1       1 -0.4382340 -1.43191776  0.71053894  0.02706694 -0.675114074
#> 2       2  0.8611619 -1.04579706 -0.67236757  1.25035294  0.916970545
#> 3       3 -1.2844771 -1.03674206 -0.72531310  1.53414108  0.408941443
#> 4       4  0.3954529  0.59511839 -0.93270366 -0.13486166 -0.003659161
#> 5       5  0.1373970  1.25233434  1.62225927 -0.62229639  0.090736987
#> 6       6 -1.5452295  1.04945745  1.25125426 -0.85287045  0.995909238
#> 7       7 -1.0660752  0.02615405 -0.77833000 -0.23209288 -1.380795143
#> 8       8  0.2209820 -0.79015350 -0.39195276 -1.60989568  0.338482515
#> 9       9  0.2872886 -0.36859911 -0.79404586  0.84757932  1.174390484
#> 10     10  0.8092039 -0.50406868  0.05845489 -0.35371648 -1.081652457

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5