Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  0.02001287
#> 2    X2 -0.08282682
#> 3    X3 -0.13879804
#> 4    X4 -0.44686537
#> 5    X5 -2.29646053
#> 6    X1  0.31916566
#> 7    X2  0.75195717
#> 8    X3  0.28716731
#> 9    X4 -0.08493407
#> 10   X5 -2.25362980
#> 11   X1 -0.35931393
#> 12   X2 -0.13214418
#> 13   X3 -1.35249137
#> 14   X4  0.08063496
#> 15   X5  0.03970517
#> 16   X1 -0.57293655
#> 17   X2 -0.46340569
#> 18   X3 -0.99092384
#> 19   X4 -1.96914989
#> 20   X5 -0.20499792
#> 21   X1 -0.80829329
#> 22   X2  0.72674125
#> 23   X3  1.39673753
#> 24   X4 -0.46629230
#> 25   X5  0.87677485
#> 26   X1  0.47255849
#> 27   X2  1.52554877
#> 28   X3  0.49102209
#> 29   X4 -0.56137270
#> 30   X5  0.92881089
#> 31   X1 -0.73417051
#> 32   X2  1.82275284
#> 33   X3 -1.01549871
#> 34   X4 -2.28426812
#> 35   X5  0.31132982
#> 36   X1  0.59506033
#> 37   X2  0.07742849
#> 38   X3 -0.17239268
#> 39   X4 -1.43015262
#> 40   X5 -0.27059500
#> 41   X1 -0.03195081
#> 42   X2 -1.39173604
#> 43   X3  0.64321239
#> 44   X4  2.37466284
#> 45   X5  0.74008197
#> 46   X1  0.76131360
#> 47   X2 -0.99430975
#> 48   X3  0.25789876
#> 49   X4 -0.02800201
#> 50   X5  1.47349198

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3          X4          X5 Row Column     Numbers
#> 1  -0.1387980 -0.44686537 -2.29646053   1     X1  0.02001287
#> 2  -0.1387980 -0.44686537 -2.29646053   1     X2 -0.08282682
#> 3   0.2871673 -0.08493407 -2.25362980   2     X1  0.31916566
#> 4   0.2871673 -0.08493407 -2.25362980   2     X2  0.75195717
#> 5  -1.3524914  0.08063496  0.03970517   3     X1 -0.35931393
#> 6  -1.3524914  0.08063496  0.03970517   3     X2 -0.13214418
#> 7  -0.9909238 -1.96914989 -0.20499792   4     X1 -0.57293655
#> 8  -0.9909238 -1.96914989 -0.20499792   4     X2 -0.46340569
#> 9   1.3967375 -0.46629230  0.87677485   5     X1 -0.80829329
#> 10  1.3967375 -0.46629230  0.87677485   5     X2  0.72674125
#> 11  0.4910221 -0.56137270  0.92881089   6     X1  0.47255849
#> 12  0.4910221 -0.56137270  0.92881089   6     X2  1.52554877
#> 13 -1.0154987 -2.28426812  0.31132982   7     X1 -0.73417051
#> 14 -1.0154987 -2.28426812  0.31132982   7     X2  1.82275284
#> 15 -0.1723927 -1.43015262 -0.27059500   8     X1  0.59506033
#> 16 -0.1723927 -1.43015262 -0.27059500   8     X2  0.07742849
#> 17  0.6432124  2.37466284  0.74008197   9     X1 -0.03195081
#> 18  0.6432124  2.37466284  0.74008197   9     X2 -1.39173604
#> 19  0.2578988 -0.02800201  1.47349198  10     X1  0.76131360
#> 20  0.2578988 -0.02800201  1.47349198  10     X2 -0.99430975

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2   Value_X3    Value_X4    Value_X5
#> 1       1  0.02001287 -0.08282682 -0.1387980 -0.44686537 -2.29646053
#> 2       2  0.31916566  0.75195717  0.2871673 -0.08493407 -2.25362980
#> 3       3 -0.35931393 -0.13214418 -1.3524914  0.08063496  0.03970517
#> 4       4 -0.57293655 -0.46340569 -0.9909238 -1.96914989 -0.20499792
#> 5       5 -0.80829329  0.72674125  1.3967375 -0.46629230  0.87677485
#> 6       6  0.47255849  1.52554877  0.4910221 -0.56137270  0.92881089
#> 7       7 -0.73417051  1.82275284 -1.0154987 -2.28426812  0.31132982
#> 8       8  0.59506033  0.07742849 -0.1723927 -1.43015262 -0.27059500
#> 9       9 -0.03195081 -1.39173604  0.6432124  2.37466284  0.74008197
#> 10     10  0.76131360 -0.99430975  0.2578988 -0.02800201  1.47349198

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5