Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name        Value
#> 1    X1 -1.400043517
#> 2    X2 -0.553699384
#> 3    X3  0.468154420
#> 4    X4  0.935363190
#> 5    X5  0.070034850
#> 6    X1  0.255317055
#> 7    X2  0.628982042
#> 8    X3  0.362951256
#> 9    X4  0.176488611
#> 10   X5 -0.639123324
#> 11   X1 -2.437263611
#> 12   X2  2.065024895
#> 13   X3 -1.304543545
#> 14   X4  0.243685465
#> 15   X5 -0.049964899
#> 16   X1 -0.005571287
#> 17   X2 -1.630989402
#> 18   X3  0.737776321
#> 19   X4  1.623548883
#> 20   X5 -0.251483443
#> 21   X1  0.621552721
#> 22   X2  0.512426950
#> 23   X3  1.888504929
#> 24   X4  0.112038083
#> 25   X5  0.444797116
#> 26   X1  1.148411606
#> 27   X2 -1.863011492
#> 28   X3 -0.097445104
#> 29   X4 -0.133997013
#> 30   X5  2.755417575
#> 31   X1 -1.821817661
#> 32   X2 -0.522012515
#> 33   X3 -0.935847354
#> 34   X4 -1.910087468
#> 35   X5  0.046531380
#> 36   X1 -0.247325302
#> 37   X2 -0.052601910
#> 38   X3 -0.015950311
#> 39   X4 -0.279237242
#> 40   X5  0.577709069
#> 41   X1 -0.244199607
#> 42   X2  0.542996343
#> 43   X3 -0.826788954
#> 44   X4 -0.313445978
#> 45   X5  0.118194874
#> 46   X1 -0.282705449
#> 47   X2 -0.914074827
#> 48   X3 -1.512399651
#> 49   X4  1.067307879
#> 50   X5 -1.911720491

# Customizing the names
data_to_long(wide_data,
             cols = c(1, 2),
             colnames_to = "Column",
             values_to = "Numbers",
             rows_to = "Row")
#>             X3         X4          X5 Row Column      Numbers
#> 1   0.46815442  0.9353632  0.07003485   1     X1 -1.400043517
#> 2   0.46815442  0.9353632  0.07003485   1     X2 -0.553699384
#> 3   0.36295126  0.1764886 -0.63912332   2     X1  0.255317055
#> 4   0.36295126  0.1764886 -0.63912332   2     X2  0.628982042
#> 5  -1.30454355  0.2436855 -0.04996490   3     X1 -2.437263611
#> 6  -1.30454355  0.2436855 -0.04996490   3     X2  2.065024895
#> 7   0.73777632  1.6235489 -0.25148344   4     X1 -0.005571287
#> 8   0.73777632  1.6235489 -0.25148344   4     X2 -1.630989402
#> 9   1.88850493  0.1120381  0.44479712   5     X1  0.621552721
#> 10  1.88850493  0.1120381  0.44479712   5     X2  0.512426950
#> 11 -0.09744510 -0.1339970  2.75541758   6     X1  1.148411606
#> 12 -0.09744510 -0.1339970  2.75541758   6     X2 -1.863011492
#> 13 -0.93584735 -1.9100875  0.04653138   7     X1 -1.821817661
#> 14 -0.93584735 -1.9100875  0.04653138   7     X2 -0.522012515
#> 15 -0.01595031 -0.2792372  0.57770907   8     X1 -0.247325302
#> 16 -0.01595031 -0.2792372  0.57770907   8     X2 -0.052601910
#> 17 -0.82678895 -0.3134460  0.11819487   9     X1 -0.244199607
#> 18 -0.82678895 -0.3134460  0.11819487   9     X2  0.542996343
#> 19 -1.51239965  1.0673079 -1.91172049  10     X1 -0.282705449
#> 20 -1.51239965  1.0673079 -1.91172049  10     X2 -0.914074827

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID")  # Save row number
data_to_wide(long_data,
             colnames_from = "Name",
             values_from = "Value",
             rows_from = "Row_ID")
#>    Row_ID     Value_X1    Value_X2    Value_X3   Value_X4    Value_X5
#> 1       1 -1.400043517 -0.55369938  0.46815442  0.9353632  0.07003485
#> 2       2  0.255317055  0.62898204  0.36295126  0.1764886 -0.63912332
#> 3       3 -2.437263611  2.06502490 -1.30454355  0.2436855 -0.04996490
#> 4       4 -0.005571287 -1.63098940  0.73777632  1.6235489 -0.25148344
#> 5       5  0.621552721  0.51242695  1.88850493  0.1120381  0.44479712
#> 6       6  1.148411606 -1.86301149 -0.09744510 -0.1339970  2.75541758
#> 7       7 -1.821817661 -0.52201251 -0.93584735 -1.9100875  0.04653138
#> 8       8 -0.247325302 -0.05260191 -0.01595031 -0.2792372  0.57770907
#> 9       9 -0.244199607  0.54299634 -0.82678895 -0.3134460  0.11819487
#> 10     10 -0.282705449 -0.91407483 -1.51239965  1.0673079 -1.91172049

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi  # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
                       cols = "\\d",  # Select all columns that contain a digit
                       colnames_to = "Item",
                       values_to = "Score",
                       rows_to = "Participant")

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
                       colnames_from = "Item",
                       values_from = "Score")
   head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5