Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

reshape_longer(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

reshape_wider(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data
A data frame to pivot.

cols
A vector of column names or indices to pivot into longer format.

colnames_to
The name of the new column that will contain the column names.

values_to
The name of the new column that will contain the values of the pivoted variables.

rows_to
The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...
Additional arguments passed on to methods.

names_to, names_from
Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from
The name of the column that contains the values of the put in the columns.

colnames_from
The name of the column that contains the levels to be used as future columns

rows_from
The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep
The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  0.07643317
#> 2    X2 -1.70190591
#> 3    X3  0.28818185
#> 4    X4  0.60590795
#> 5    X5  0.69753323
#> 6    X1 -1.04652097
#> 7    X2 -2.41223364
#> 8    X3 -0.19515071
#> 9    X4 -0.44439254
#> 10   X5  0.73450152
#> 11   X1 -0.53201721
#> 12   X2 -1.17111815
#> 13   X3  0.36542646
#> 14   X4  1.12234573
#> 15   X5  1.32554490
#> 16   X1  0.26823962
#> 17   X2  0.14809270
#> 18   X3  0.31192105
#> 19   X4 -1.79082488
#> 20   X5  0.17086334
#> 21   X1 -1.51386008
#> 22   X2  1.54654871
#> 23   X3  0.68051641
#> 24   X4 -0.29974713
#> 25   X5 -1.38144095
#> 26   X1  0.09360624
#> 27   X2 -0.60711946
#> 28   X3  1.70390673
#> 29   X4 -0.47529655
#> 30   X5  0.61886405
#> 31   X1 -1.24405756
#> 32   X2 -0.09933875
#> 33   X3 -2.37064805
#> 34   X4  0.97084452
#> 35   X5 -0.17575744
#> 36   X1 -0.79955139
#> 37   X2  0.55084256
#> 38   X3 -0.65999582
#> 39   X4 -0.80914988
#> 40   X5  0.86189776
#> 41   X1 -0.64544524
#> 42   X2  0.68783388
#> 43   X3  0.79301962
#> 44   X4 -1.51965195
#> 45   X5  0.10134123
#> 46   X1 -0.88883544
#> 47   X2 -2.11932901
#> 48   X3  0.25743700
#> 49   X4  1.53630723
#> 50   X5 -0.05537317

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3         X4          X5 Row Column     Numbers
#> 1   0.2881819  0.6059079  0.69753323   1     X1  0.07643317
#> 2   0.2881819  0.6059079  0.69753323   1     X2 -1.70190591
#> 3  -0.1951507 -0.4443925  0.73450152   2     X1 -1.04652097
#> 4  -0.1951507 -0.4443925  0.73450152   2     X2 -2.41223364
#> 5   0.3654265  1.1223457  1.32554490   3     X1 -0.53201721
#> 6   0.3654265  1.1223457  1.32554490   3     X2 -1.17111815
#> 7   0.3119211 -1.7908249  0.17086334   4     X1  0.26823962
#> 8   0.3119211 -1.7908249  0.17086334   4     X2  0.14809270
#> 9   0.6805164 -0.2997471 -1.38144095   5     X1 -1.51386008
#> 10  0.6805164 -0.2997471 -1.38144095   5     X2  1.54654871
#> 11  1.7039067 -0.4752966  0.61886405   6     X1  0.09360624
#> 12  1.7039067 -0.4752966  0.61886405   6     X2 -0.60711946
#> 13 -2.3706480  0.9708445 -0.17575744   7     X1 -1.24405756
#> 14 -2.3706480  0.9708445 -0.17575744   7     X2 -0.09933875
#> 15 -0.6599958 -0.8091499  0.86189776   8     X1 -0.79955139
#> 16 -0.6599958 -0.8091499  0.86189776   8     X2  0.55084256
#> 17  0.7930196 -1.5196519  0.10134123   9     X1 -0.64544524
#> 18  0.7930196 -1.5196519  0.10134123   9     X2  0.68783388
#> 19  0.2574370  1.5363072 -0.05537317  10     X1 -0.88883544
#> 20  0.2574370  1.5363072 -0.05537317  10     X2 -2.11932901

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1    Value_X2   Value_X3   Value_X4    Value_X5
#> 1       1  0.07643317 -1.70190591  0.2881819  0.6059079  0.69753323
#> 2       2 -1.04652097 -2.41223364 -0.1951507 -0.4443925  0.73450152
#> 3       3 -0.53201721 -1.17111815  0.3654265  1.1223457  1.32554490
#> 4       4  0.26823962  0.14809270  0.3119211 -1.7908249  0.17086334
#> 5       5 -1.51386008  1.54654871  0.6805164 -0.2997471 -1.38144095
#> 6       6  0.09360624 -0.60711946  1.7039067 -0.4752966  0.61886405
#> 7       7 -1.24405756 -0.09933875 -2.3706480  0.9708445 -0.17575744
#> 8       8 -0.79955139  0.55084256 -0.6599958 -0.8091499  0.86189776
#> 9       9 -0.64544524  0.68783388  0.7930196 -1.5196519  0.10134123
#> 10     10 -0.88883544 -2.11932901  0.2574370  1.5363072 -0.05537317

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5