Reshape (pivot) data from wide to long

This function "lengthens" data, increasing the number of rows and decreasing the number of columns. This is a dependency-free base-R equivalent of tidyr::pivot_longer().

Usage

data_to_long(
  data,
  cols = "all",
  colnames_to = "Name",
  values_to = "Value",
  rows_to = NULL,
  ...,
  names_to = colnames_to
)

data_to_wide(
  data,
  values_from = "Value",
  colnames_from = "Name",
  rows_from = NULL,
  sep = "_",
  ...,
  names_from = colnames_from
)

Arguments

data

A data frame to pivot.

cols

A vector of column names or indices to pivot into longer format.

colnames_to

The name of the new column that will contain the column names.

values_to

The name of the new column that will contain the values of the pivoted variables.

rows_to

The name of the column that will contain the row-number from the original data. If NULL, will be removed.

...

Additional arguments passed on to methods.

names_to, names_from

Same as colnames_to, is there for compatibility with tidyr::pivot_longer().

values_from

The name of the column that contains the values of the put in the columns.

colnames_from

The name of the column that contains the levels to be used as future columns.

rows_from

The name of the column that identifies the rows. If NULL, will use all the unique rows.

sep

The indicating a separating character in the variable names in the wide format.

Value

data.frame

Examples

wide_data <- data.frame(replicate(5, rnorm(10)))

# From wide to long
# ------------------
# Default behaviour (equivalent to tidyr::pivot_longer(wide_data, cols = 1:5))
data_to_long(wide_data)
#>    Name       Value
#> 1    X1  0.82305095
#> 2    X2  0.56317115
#> 3    X3 -2.52061076
#> 4    X4  0.37696850
#> 5    X5 -0.81819454
#> 6    X1 -0.84103556
#> 7    X2  0.30467272
#> 8    X3  0.61970601
#> 9    X4  0.25988828
#> 10   X5 -1.25366307
#> 11   X1 -0.60267378
#> 12   X2 -0.84385823
#> 13   X3 -1.13636751
#> 14   X4 -0.76637385
#> 15   X5  0.89937725
#> 16   X1  1.34683895
#> 17   X2 -0.19482996
#> 18   X3 -1.37206044
#> 19   X4  1.26251410
#> 20   X5 -1.03112806
#> 21   X1 -0.36370975
#> 22   X2  1.35051681
#> 23   X3 -2.09296992
#> 24   X4  0.63877805
#> 25   X5  0.91018128
#> 26   X1 -0.77820054
#> 27   X2  0.86345870
#> 28   X3  0.53743137
#> 29   X4  0.67609708
#> 30   X5 -0.47225262
#> 31   X1  0.64574910
#> 32   X2  1.10967103
#> 33   X3  0.67224124
#> 34   X4  0.14854358
#> 35   X5  0.24571529
#> 36   X1  0.12376862
#> 37   X2  0.50734621
#> 38   X3 -0.13844555
#> 39   X4  0.60251838
#> 40   X5 -0.65830583
#> 41   X1  1.49184549
#> 42   X2 -1.63398974
#> 43   X3 -0.54475063
#> 44   X4 -0.03937592
#> 45   X5 -0.11054275
#> 46   X1  0.04542114
#> 47   X2  0.92004145
#> 48   X3 -1.31929936
#> 49   X4 -0.09757986
#> 50   X5  0.36903080

# Customizing the names
data_to_long(wide_data,
  cols = c(1, 2),
  colnames_to = "Column",
  values_to = "Numbers",
  rows_to = "Row"
)
#>            X3          X4         X5 Row Column     Numbers
#> 1  -2.5206108  0.37696850 -0.8181945   1     X1  0.82305095
#> 2  -2.5206108  0.37696850 -0.8181945   1     X2  0.56317115
#> 3   0.6197060  0.25988828 -1.2536631   2     X1 -0.84103556
#> 4   0.6197060  0.25988828 -1.2536631   2     X2  0.30467272
#> 5  -1.1363675 -0.76637385  0.8993772   3     X1 -0.60267378
#> 6  -1.1363675 -0.76637385  0.8993772   3     X2 -0.84385823
#> 7  -1.3720604  1.26251410 -1.0311281   4     X1  1.34683895
#> 8  -1.3720604  1.26251410 -1.0311281   4     X2 -0.19482996
#> 9  -2.0929699  0.63877805  0.9101813   5     X1 -0.36370975
#> 10 -2.0929699  0.63877805  0.9101813   5     X2  1.35051681
#> 11  0.5374314  0.67609708 -0.4722526   6     X1 -0.77820054
#> 12  0.5374314  0.67609708 -0.4722526   6     X2  0.86345870
#> 13  0.6722412  0.14854358  0.2457153   7     X1  0.64574910
#> 14  0.6722412  0.14854358  0.2457153   7     X2  1.10967103
#> 15 -0.1384455  0.60251838 -0.6583058   8     X1  0.12376862
#> 16 -0.1384455  0.60251838 -0.6583058   8     X2  0.50734621
#> 17 -0.5447506 -0.03937592 -0.1105428   9     X1  1.49184549
#> 18 -0.5447506 -0.03937592 -0.1105428   9     X2 -1.63398974
#> 19 -1.3192994 -0.09757986  0.3690308  10     X1  0.04542114
#> 20 -1.3192994 -0.09757986  0.3690308  10     X2  0.92004145

# From long to wide
# -----------------
long_data <- data_to_long(wide_data, rows_to = "Row_ID") # Save row number
data_to_wide(long_data,
  colnames_from = "Name",
  values_from = "Value",
  rows_from = "Row_ID"
)
#>    Row_ID    Value_X1   Value_X2   Value_X3    Value_X4   Value_X5
#> 1       1  0.82305095  0.5631711 -2.5206108  0.37696850 -0.8181945
#> 2       2 -0.84103556  0.3046727  0.6197060  0.25988828 -1.2536631
#> 3       3 -0.60267378 -0.8438582 -1.1363675 -0.76637385  0.8993772
#> 4       4  1.34683895 -0.1948300 -1.3720604  1.26251410 -1.0311281
#> 5       5 -0.36370975  1.3505168 -2.0929699  0.63877805  0.9101813
#> 6       6 -0.77820054  0.8634587  0.5374314  0.67609708 -0.4722526
#> 7       7  0.64574910  1.1096710  0.6722412  0.14854358  0.2457153
#> 8       8  0.12376862  0.5073462 -0.1384455  0.60251838 -0.6583058
#> 9       9  1.49184549 -1.6339897 -0.5447506 -0.03937592 -0.1105428
#> 10     10  0.04542114  0.9200415 -1.3192994 -0.09757986  0.3690308

# Full example
# ------------------
if (require("psych")) {
  data <- psych::bfi # Wide format with one row per participant's personality test

  # Pivot long format
  long <- data_to_long(data,
    cols = "\\d", # Select all columns that contain a digit
    colnames_to = "Item",
    values_to = "Score",
    rows_to = "Participant"
  )

  # Separate facet and question number
  long$Facet <- gsub("\\d", "", long$Item)
  long$Item <- gsub("[A-Z]", "", long$Item)
  long$Item <- paste0("I", long$Item)

  wide <- data_to_wide(long,
    colnames_from = "Item",
    values_from = "Score"
  )
  head(wide)
}
#> Loading required package: psych
#>   gender education age Participant Facet Score_I1 Score_I2 Score_I3 Score_I4
#> 1      1        NA  16       61617     A        2        4        3        4
#> 2      1        NA  16       61617     C        2        3        3        4
#> 3      1        NA  16       61617     E        3        3        3        4
#> 4      1        NA  16       61617     N        3        4        2        2
#> 5      1        NA  16       61617     O        3        6        3        4
#> 6      2        NA  18       61618     A        2        4        5        2
#>   Score_I5
#> 1        4
#> 2        4
#> 3        4
#> 4        3
#> 5        3
#> 6        5