Generic export of data frames into formatted tables

display() is a generic function to export data frames into various table formats (like plain text, markdown, ...). print_md() usually is a convenient wrapper for display(format = "markdown"). Similar, print_html() is a shortcut for display(format = "html"). See the documentation for the specific objects' classes.

Usage

display(object, ...)

print_md(x, ...)

print_html(x, ...)

# S3 method for class 'data.frame'
display(object, format = "markdown", ...)

# S3 method for class 'data.frame'
print_md(x, ...)

# S3 method for class 'data.frame'
print_html(x, ...)

Arguments

object, x

A data frame.

...

Arguments passed to other methods.

format

String, indicating the output format. Can be "markdown" or "html".

Value

Depending on format, either an object of class gt_tbl or a character vector of class knitr_kable.

Examples

display(iris[1:5, ], format = "html")

  

Sepal.Length
      	Sepal.Width
      	Petal.Length
      	Petal.Width
      	Species
    

5.10
	3.50
	1.40
	0.20
	setosa

4.90
	3.00
	1.40
	0.20
	setosa

4.70
	3.20
	1.30
	0.20
	setosa

4.60
	3.10
	1.50
	0.20
	setosa

5.00
	3.60
	1.40
	0.20
	setosa