find_algorithm {insight}	R Documentation

Find sampling algorithm and optimizers

Description

Returns information on the sampling or estimation algorithm as well as optimization functions, or for Bayesian model information on chains, iterations and warmup-samples.

Usage

find_algorithm(x, ...)

Arguments

x
A fitted model.

...
Currently not used.

Value

A list with elements depending on the model.
For frequentist models:

algorithm, for instance "OLS" or "ML"

optimizer, name of optimizing function, only applies to specific models (like gam)

For frequentist mixed models:

algorithm, for instance "REML" or "ML"

optimizer, name of optimizing function

For Bayesian models:

algorithm, the algorithm

chains, number of chains

iterations, number of iterations per chain

warmup, number of warmups per chain

Examples

if (require("lme4")) {
  data(sleepstudy)
  m <- lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
  find_algorithm(m)
}
## Not run: 
library(rstanarm)
m <- stan_lmer(Reaction ~ Days + (1 | Subject), data = sleepstudy)
find_algorithm(m)

## End(Not run)

[Package insight version 0.14.0 Index]