Find model formula

Returns the formula(s) for the different parts of a model (like fixed or random effects, zero-inflated component, ...).

find_formula(x, ...)

Arguments

x
A fitted model.

...
Currently not used.

Value

A list of formulas that describe the model. For simple models, only one list-element, conditional, is returned. For more complex models, the returned list may have following elements:

conditional, the "fixed effects" part from the model

random, the "random effects" part from the model (or the id for gee-models and similar)

zero_inflated, the "fixed effects" part from the zero-inflation component of the model

zero_inflated_random, the "random effects" part from the zero-inflation component of the model

dispersion, the dispersion formula

instruments, for fixed-effects regressions like ivreg, felm or plm, the instrumental variables

cluster, for fixed-effects regressions like felm, the cluster specification

correlation, for models with correlation-component like gls, the formula that describes the correlation structure

slopes, for fixed-effects individual-slope models like feis, the formula for the slope parameters

Note

For models of class lme or gls the correlation-component is only returned, when it is explicitly defined as named argument (form), e.g. corAR1(form = ~1 | Mare)

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_formula(m)

#> $conditional
#> mpg ~ wt + cyl + vs
#> <environment: 0x000000002e510b98>
#>