find_parameters {insight}	R Documentation

Find names of model parameters

Description

Returns the names of model parameters, like they typically appear in the summary() output. For Bayesian models, the parameter names equal the column names of the posterior samples after coercion from as.data.frame(). See the documentation for your object's class:

Bayesian models (rstanarm, brms, MCMCglmm, ...)

Generalized additive models (mgcv, VGAM, ...)

Marginal effects models (mfx)

Estimated marginal means (emmeans)

Mixed models (lme4, glmmTMB, GLMMadaptive, ...)

Zero-inflated and hurdle models (pscl, ...)

Models with special components (betareg, MuMIn, ...)

Usage

find_parameters(x, ...)

## Default S3 method:
find_parameters(x, flatten = FALSE, verbose = TRUE, ...)

Arguments

x
A fitted model.

...
Currently not used.

flatten
Logical, if TRUE, the values are returned as character vector, not as list. Duplicated values are removed.

verbose
Toggle messages and warnings.

Value

A list of parameter names. For simple models, only one list-element, conditional, is returned.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_parameters(m)

[Package insight version 0.14.0 Index]