find_parameters {insight}	R Documentation

Find names of model parameters

Description

Returns the names of model parameters, like they typically appear in the summary() output. For Bayesian models, the parameter names equal the column names of the posterior samples after coercion from as.data.frame(). See the documentation for your object's class:

Bayesian models (rstanarm, brms, MCMCglmm, ...)

Generalized additive models (mgcv, VGAM, ...)

Marginal effects models (mfx)

Estimated marginal means (emmeans)

Mixed models (lme4, glmmTMB, GLMMadaptive, ...)

Zero-inflated and hurdle models (pscl, ...)

Models with special components (betareg, MuMIn, ...)

Usage


find_parameters(x, ...)

## Default S3 method:
find_parameters(x, flatten = FALSE, verbose = TRUE, ...)

## S3 method for class 'pgmm'
find_parameters(x, component = c("conditional", "all"), flatten = FALSE, ...)

## S3 method for class 'nls'
find_parameters(
  x,
  component = c("all", "conditional", "nonlinear"),
  flatten = FALSE,
  ...
)

Arguments

x
A fitted model.

...
Currently not used.

flatten
Logical, if TRUE, the values are returned as character vector, not as list. Duplicated values are removed.

verbose
Toggle messages and warnings.

component
Should all predictor variables, predictor variables for the conditional model, the zero-inflated part of the model, the dispersion term or the instrumental variables be returned? Applies to models with zero-inflated and/or dispersion formula, or to models with instrumental variable (so called fixed-effects regressions). May be abbreviated. Note that the conditional component is also called count or mean component, depending on the model.

Value

A list of parameter names. For simple models, only one list-element, conditional, is returned.

Model components

Possible values for the component argument depend on the model class. Following are valid options:

"all": returns all model components, applies to all models, but will only have an effect for models with more than just the conditional model component.

"conditional": only returns the conditional component, i.e. "fixed effects" terms from the model. Will only have an effect for models with more than just the conditional model component.

"smooth_terms": returns smooth terms, only applies to GAMs (or similar models that may contain smooth terms).

"zero_inflated" (or "zi"): returns the zero-inflation component.

"dispersion": returns the dispersion model component. This is common for models with zero-inflation or that can model the dispersion parameter.

"instruments": for instrumental-variable or some fixed effects regression, returns the instruments.

"location": returns location parameters such as conditional, zero_inflated, smooth_terms, or instruments (everything that are fixed or random effects - depending on the effects argument - but no auxiliary parameters).

"distributional" (or "auxiliary"): components like sigma, dispersion, beta or precision (and other auxiliary parameters) are returned.

Examples


data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_parameters(m)

[Package insight version 0.19.10 Index]