Find names of model parameters from Bayesian models

Returns the names of model parameters, like they typically appear in the summary() output. For Bayesian models, the parameter names equal the column names of the posterior samples after coercion from as.data.frame().

Usage

# S3 method for class 'BGGM'
find_parameters(x, component = "correlation", flatten = FALSE, ...)

# S3 method for class 'brmsfit'
find_parameters(
  x,
  effects = "all",
  component = "all",
  flatten = FALSE,
  parameters = NULL,
  ...
)

Arguments

x

A fitted model.

component

Which type of parameters to return, such as parameters for the conditional model, the zero-inflated part of the model, the dispersion term, the instrumental variables or marginal effects be returned? Applies to models with zero-inflated and/or dispersion formula, or to models with instrumental variables (so called fixed-effects regressions), or models with marginal effects (from mfx). See details in section Model Components .May be abbreviated. Note that the conditional component also refers to the count or mean component - names may differ, depending on the modeling package. There are three convenient shortcuts (not applicable to all model classes):

component = "all" returns all possible parameters.

If component = "location", location parameters such as conditional, zero_inflated, smooth_terms, or instruments are returned (everything that are fixed or random effects - depending on the effects argument - but no auxiliary parameters).

For component = "distributional" (or "auxiliary"), components like sigma, dispersion, beta or precision (and other auxiliary parameters) are returned.

flatten

Logical, if TRUE, the values are returned as character vector, not as list. Duplicated values are removed.

...

Currently not used.

effects

Should variables for fixed effects ("fixed"), random effects ("random") or both ("all") be returned? Only applies to mixed models. May be abbreviated.

parameters

Regular expression pattern that describes the parameters that should be returned.

Value

A list of parameter names. For simple models, only one list-element, conditional, is returned. For more complex models, the returned list may have following elements:

conditional, the "fixed effects" part from the model

random, the "random effects" part from the model

zero_inflated, the "fixed effects" part from the zero-inflation component of the model

zero_inflated_random, the "random effects" part from the zero-inflation component of the model

smooth_terms, the smooth parameters

Furthermore, some models, especially from brms, can also return auxiliary parameters. These may be one of the following:

sigma, the residual standard deviation (auxiliary parameter)

dispersion, the dispersion parameters (auxiliary parameter)

beta, the beta parameter (auxiliary parameter)

simplex, simplex parameters of monotonic effects (brms only)

mix, mixture parameters (brms only)

shiftprop, shifted proportion parameters (brms only)

Models of class BGGM additionally can return the elements correlation and intercept.

Models of class BFBayesFactor additionally can return the element extra.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_parameters(m)
#> $conditional
#> [1] "(Intercept)" "wt"          "cyl"         "vs"         
#>