find_parameters.gamlss {insight}	R Documentation

Find names of model parameters from generalized additive models

Description

Returns the names of model parameters, like they typically appear in the summary() output.

Usage


## S3 method for class 'gamlss'
find_parameters(x, flatten = FALSE, ...)

## S3 method for class 'gam'
find_parameters(
  x,
  component = c("all", "conditional", "smooth_terms", "location"),
  flatten = FALSE,
  ...
)

Arguments

x
A fitted model.

flatten
Logical, if TRUE, the values are returned as character vector, not as list. Duplicated values are removed.

...
Currently not used.

component
Should all predictor variables, predictor variables for the conditional model, the zero-inflated part of the model, the dispersion term or the instrumental variables be returned? Applies to models with zero-inflated and/or dispersion formula, or to models with instrumental variable (so called fixed-effects regressions). May be abbreviated. Note that the conditional component is also called count or mean component, depending on the model.

Value

A list of parameter names. The returned list may have following elements:

conditional, the "fixed effects" part from the model.

smooth_terms, the smooth parameters.

Examples


data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_parameters(m)

[Package insight version 0.19.10 Index]