Find names of model parameters from mixed models

Returns the names of model parameters, like they typically appear in the summary() output.

# S3 method for glmmTMB
find_parameters(
  x,
  effects = c("all", "fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated", "dispersion"),
  flatten = FALSE,
  ...
)

# S3 method for merMod
find_parameters(x, effects = c("all", "fixed", "random"), flatten = FALSE, ...)

Arguments

x
A fitted model.

effects
Should parameters for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

component
Which type of parameters to return, such as parameters for the conditional model, the zero-inflated part of the model or the dispersion term? Applies to models with zero-inflated and/or dispersion formula. Note that the conditional component is also called count or mean component, depending on the model. There are three convenient shortcuts: component = "all" returns all possible parameters. If component = "location", location parameters such as conditional or zero_inflated are returned (everything that are fixed or random effects - depending on the effects argument - but no auxiliary parameters). For component = "distributional" (or "auxiliary"), components like sigma or dispersion (and other auxiliary parameters) are returned.

flatten
Logical, if TRUE, the values are returned as character vector, not as list. Duplicated values are removed.

...
Currently not used.

Value

A list of parameter names. The returned list may have following elements:

conditional, the "fixed effects" part from the model.

random, the "random effects" part from the model.

zero_inflated, the "fixed effects" part from the zero-inflation component of the model.

zero_inflated_random, the "random effects" part from the zero-inflation component of the model.

dispersion, the dispersion parameters (auxiliary parameter)

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_parameters(m)
#> $conditional
#> [1] "(Intercept)" "wt"          "cyl"         "vs"         
#>