find_predictors {insight}	R Documentation

Find names of model predictors

Description

Returns the names of the predictor variables for the different parts of a model (like fixed or random effects, zero-inflated component, ...). Unlike find_parameters, the names from find_predictors() match the original variable names from the data that was used to fit the model.

Usage

find_predictors(
  x,
  effects = c("fixed", "random", "all"),
  component = c("all", "conditional", "zi", "zero_inflated", "dispersion",
    "instruments", "correlation", "smooth_terms"),
  flatten = FALSE,
  verbose = TRUE
)

Arguments

x
A fitted model.

effects
Should variables for fixed effects, random effects or both be returned? Only applies to mixed models. May be abbreviated.

component
Should all predictor variables, predictor variables for the conditional model, the zero-inflated part of the model, the dispersion term or the instrumental variables be returned? Applies to models with zero-inflated and/or dispersion formula, or to models with instrumental variable (so called fixed-effects regressions). May be abbreviated. Note that the conditional component is also called count or mean component, depending on the model.

flatten
Logical, if TRUE, the values are returned as character vector, not as list. Duplicated values are removed.

verbose
Toggle warnings.

Value

A list of character vectors that represent the name(s) of the predictor variables. Depending on the combination of the arguments effects and component, the returned list has following elements:

conditional, the "fixed effects" terms from the model

random, the "random effects" terms from the model

zero_inflated, the "fixed effects" terms from the zero-inflation component of the model

zero_inflated_random, the "random effects" terms from the zero-inflation component of the model

dispersion, the dispersion terms

instruments, for fixed-effects regressions like ivreg, felm or plm, the instrumental variables

correlation, for models with correlation-component like gls, the variables used to describe the correlation structure

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
find_predictors(m)

[Package insight version 0.14.0 Index]