Confidence/Credible Interval (CI) Formatting

Confidence/Credible Interval (CI) Formatting

format_ci(
  CI_low,
  CI_high,
  ci = 0.95,
  digits = 2,
  brackets = TRUE,
  width = NULL,
  width_low = width,
  width_high = width,
  missing = ""
)

Arguments

CI_low
Lower CI bound.

CI_high
Upper CI bound.

ci
CI level in percentage.

digits
Number of significant digits.

brackets
Logical, if TRUE (default), values are encompassed in square brackets.

width
Minimum width of the returned string. If not NULL and width is larger than the string's length, leading whitespaces are added to the string. If width="auto", width will be set to the length of the longest string.

width_low, width_high
Like width, but only applies to the lower or higher confidence interval value. This can be used when the values for the lower and upper CI are of very different length.

missing
Value by which NA values are replaced. By default, an empty string (i.e. "") is returned for NA.

Value

A formatted string.

Examples

format_ci(1.20, 3.57, ci = 0.90)

#> [1] "90% CI [1.20, 3.57]"

format_ci(1.20, 3.57, ci = NULL)

#> [1] "[1.20, 3.57]"

format_ci(1.20, 3.57, ci = NULL, brackets = FALSE)

#> [1] "1.20, 3.57"

format_ci(c(1.205645, 23.4), c(3.57, -1.35), ci = 0.90)

#> [1] "90% CI [1.21, 3.57]"   "90% CI [23.40, -1.35]"

format_ci(c(1.20, NA, NA), c(3.57, -1.35, NA), ci = 0.90)

#> [1] "90% CI [1.20, 3.57]" "90% CI [, -1.35]"    ""                   

# automatic alignment of width, useful for printing multiple CIs in columns
x <- format_ci(c(1.205, 23.4, 100.43), c(3.57, -13.35, 9.4))
cat(x, sep = "\n")

#> 95% CI [1.21, 3.57]
#> 95% CI [23.40, -13.35]
#> 95% CI [100.43, 9.40]

x <- format_ci(c(1.205, 23.4, 100.43), c(3.57, -13.35, 9.4), width = "auto")
cat(x, sep = "\n")

#> 95% CI [  1.21,   3.57]
#> 95% CI [ 23.40, -13.35]
#> 95% CI [100.43,   9.40]