format_message {insight}	R Documentation

Format messages and warnings

Description

Inserts line breaks into a longer message or warning string. Line length is adjusted to maximum length of the console, if the width can be accessed. By default, new lines are indented by two whitespace.

Usage

format_message(string, ..., line_length = options()$width)

Arguments

string
A string.

...
Further strings that will be concatenated as indented new lines.

line_length
Numeric, the maximum length of a line.

Value

A formatted string.

Examples

msg <- format_message("Much too long string for just one line, I guess!",
  line_length = 15
)
message(msg)

msg <- format_message("Much too long string for just one line, I guess!",
  "First new line",
  "Second new line",
  "(both indented)",
  line_length = 30
)
message(msg)

[Package insight version 0.14.0 Index]