Format messages and warnings

Inserts line breaks into a longer message or warning string. Line length is adjusted to maximum length of the console, if the width can be accessed. By default, new lines are indented by two whitespace.

format_message(string, line_length = options()$width)

Arguments

string
A string.

line_length
Numeric, the maximum length of a line.

Value

A formatted string.

Examples

format_message("Much too long string for just one line, I guess!", 15)

#> [1] "Much too long\n  string for just\n  one line, I\n  guess!\n  "