get_data {insight}	R Documentation

Get the data that was used to fit the model

Description

This functions tries to get the data that was used to fit the model and returns it as data frame.

Usage

get_data(x, ...)

## S3 method for class 'gee'
get_data(x, effects = c("all", "fixed", "random"), verbose = TRUE, ...)

## S3 method for class 'rqss'
get_data(
  x,
  component = c("all", "conditional", "smooth_terms"),
  verbose = TRUE,
  ...
)

## S3 method for class 'hurdle'
get_data(
  x,
  component = c("all", "conditional", "zi", "zero_inflated", "dispersion"),
  verbose = TRUE,
  ...
)

## S3 method for class 'zcpglm'
get_data(
  x,
  component = c("all", "conditional", "zi", "zero_inflated"),
  verbose = TRUE,
  ...
)

## S3 method for class 'glmmTMB'
get_data(
  x,
  effects = c("all", "fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated", "dispersion"),
  verbose = TRUE,
  ...
)

## S3 method for class 'merMod'
get_data(x, effects = c("all", "fixed", "random"), verbose = TRUE, ...)

## S3 method for class 'glmmadmb'
get_data(x, effects = c("all", "fixed", "random"), verbose = TRUE, ...)

## S3 method for class 'rlmerMod'
get_data(x, effects = c("all", "fixed", "random"), ...)

## S3 method for class 'clmm'
get_data(x, effects = c("all", "fixed", "random"), ...)

## S3 method for class 'mixed'
get_data(x, effects = c("all", "fixed", "random"), ...)

## S3 method for class 'lme'
get_data(x, effects = c("all", "fixed", "random"), ...)

## S3 method for class 'MixMod'
get_data(
  x,
  effects = c("all", "fixed", "random"),
  component = c("all", "conditional", "zi", "zero_inflated", "dispersion"),
  verbose = TRUE,
  ...
)

## S3 method for class 'brmsfit'
get_data(x, effects = "all", component = "all", verbose = TRUE, ...)

## S3 method for class 'stanreg'
get_data(x, effects = c("all", "fixed", "random"), verbose = TRUE, ...)

## S3 method for class 'MCMCglmm'
get_data(x, effects = c("all", "fixed", "random"), ...)

Arguments

x
A fitted model.

...
Currently not used.

effects
Should model data for fixed effects, random effects or both be returned? Only applies to mixed models.

verbose
Toggle messages and warnings.

component
Should all predictor variables, predictor variables for the conditional model, the zero-inflated part of the model, the dispersion term or the instrumental variables be returned? Applies to models with zero-inflated and/or dispersion formula, or to models with instrumental variable (so called fixed-effects regressions). May be abbreviated. Note that the conditional component is also called count or mean component, depending on the model.

Value

The data that was used to fit the model.

Note

Unlike model.frame(), which may contain transformed variables (e.g. if poly() or scale() was used inside the formula to specify the model), get_data() aims at returning the "original", untransformed data (if possible). Consequently, column names are changed accordingly, i.e. "log(x)" will become "x" etc. for all data columns with transformed values.

Examples

if (require("lme4")) {
  data(cbpp, package = "lme4")
  cbpp$trials <- cbpp$size - cbpp$incidence
  m <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)
  head(get_data(m))
}

[Package insight version 0.14.0 Index]