Create a reference grid

Create a reference matrix, useful for visualisation, with evenly spread and combined values.

Usage

get_datagrid(x, ...)

# S3 method for data.frame
get_datagrid(
  x,
  at = "all",
  factors = "reference",
  numerics = "mean",
  preserve_range = FALSE,
  reference = x,
  ...
)

# S3 method for numeric
get_datagrid(x, length = 10, range = "range", ...)

# S3 method for factor
get_datagrid(x, ...)

# S3 method for default
get_datagrid(
  x,
  at = "all",
  factors = "reference",
  numerics = "mean",
  preserve_range = TRUE,
  reference = x,
  include_smooth = TRUE,
  include_random = FALSE,
  include_response = FALSE,
  data = NULL,
  ...
)

Arguments

x

An object from which to construct the reference grid.

...

Arguments passed to or from other methods (for instance, length or range to control the spread of numeric variables.).

at

Can be "all" or list of characters indicating columns of interest. Can also contain assignments (as named list, e.g. at = list(c(Sepal.Length = c(2, 4), Species = "setosa")), or as string, e.g. at = "Sepal.Length = 2" or at = c("Sepal.Length = 2", "Species = 'setosa'") - note the usage of single and double quotes to assign strings within strings). The remaining variables will be fixed.

factors

Type of summary for factors. Can be "reference" (set at the reference level), "mode" (set at the most common level) or "all" to keep all levels.

numerics

Type of summary for numeric values. Can be "all" (will duplicate the grid for all unique values), any function ("mean", "median", ...) or a value (e.g., numerics = 0).

preserve_range

In the case of combinations between numeric variables and factors, setting preserve_range = TRUE will drop the observations where the value of the numeric variable is originally not present in the range of its factor level. This leads to an unbalanced grid. Also, if you want the minimum and the maximum to closely match the actual ranges, you should increase the length argument.

reference

The reference vector from which to compute the mean and SD.

length

Length of numeric "at" variables.

range

Can be one of "range", "iqr", "ci", "hdi" or "eti". If "range" (default), will use the minimum and maximun of the original vector as end-points. If any other interval, will spread within the range (the default CI width is 95% but this can be changed by setting something else, e.g., ci = 0.90). See IQR() and bayestestR::ci().

include_smooth

If x is a model object, decide whether smooth terms should be included in the data grid or not.

include_random

If x is a mixed model object, decide whether random effect terms should be included in the data grid or not. If include_random is FALSE, but x is a mixed model with random effects, these will still be included in the returned grid, but set to their "population level" value (e.g., NA for glmmTMB or 0 for merMod). This ensures that common predict() methods work properly, as these usually need data with all variables in the model included.

include_response

If x is a model object, decide whether the response variable should be included in the data grid or not.

data

Optional, the data frame that was used to fit the model. Usually, the data is retrieved via get_data().

Value

Reference grid data frame.

Examples

# Add one row to change the "mode" of Species
data <- rbind(iris, iris[149, ], make.row.names = FALSE)

# Single variable is of interest; all others are "fixed"
get_datagrid(data, at = "Sepal.Length")
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1           4.3    3.059603     3.768874    1.206623  setosa
#> 2           4.7    3.059603     3.768874    1.206623  setosa
#> 3           5.1    3.059603     3.768874    1.206623  setosa
#> 4           5.5    3.059603     3.768874    1.206623  setosa
#> 5           5.9    3.059603     3.768874    1.206623  setosa
#> 6           6.3    3.059603     3.768874    1.206623  setosa
#> 7           6.7    3.059603     3.768874    1.206623  setosa
#> 8           7.1    3.059603     3.768874    1.206623  setosa
#> 9           7.5    3.059603     3.768874    1.206623  setosa
#> 10          7.9    3.059603     3.768874    1.206623  setosa
get_datagrid(data, at = "Sepal.Length", length = 3)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          4.3    3.059603     3.768874    1.206623  setosa
#> 2          6.1    3.059603     3.768874    1.206623  setosa
#> 3          7.9    3.059603     3.768874    1.206623  setosa
get_datagrid(data, at = "Sepal.Length", range = "ci", ci = 0.90)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1      4.600000    3.059603     3.768874    1.206623  setosa
#> 2      4.894444    3.059603     3.768874    1.206623  setosa
#> 3      5.188889    3.059603     3.768874    1.206623  setosa
#> 4      5.483333    3.059603     3.768874    1.206623  setosa
#> 5      5.777778    3.059603     3.768874    1.206623  setosa
#> 6      6.072222    3.059603     3.768874    1.206623  setosa
#> 7      6.366667    3.059603     3.768874    1.206623  setosa
#> 8      6.661111    3.059603     3.768874    1.206623  setosa
#> 9      6.955556    3.059603     3.768874    1.206623  setosa
#> 10     7.250000    3.059603     3.768874    1.206623  setosa
get_datagrid(data, at = "Sepal.Length", factors = "mode")
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
#> 1           4.3    3.059603     3.768874    1.206623 virginica
#> 2           4.7    3.059603     3.768874    1.206623 virginica
#> 3           5.1    3.059603     3.768874    1.206623 virginica
#> 4           5.5    3.059603     3.768874    1.206623 virginica
#> 5           5.9    3.059603     3.768874    1.206623 virginica
#> 6           6.3    3.059603     3.768874    1.206623 virginica
#> 7           6.7    3.059603     3.768874    1.206623 virginica
#> 8           7.1    3.059603     3.768874    1.206623 virginica
#> 9           7.5    3.059603     3.768874    1.206623 virginica
#> 10          7.9    3.059603     3.768874    1.206623 virginica

# Multiple variables are of interest, creating a combination
get_datagrid(data, at = c("Sepal.Length", "Species"), length = 3)
#>   Sepal.Length    Species Sepal.Width Petal.Length Petal.Width
#> 1          4.3     setosa    3.059603     3.768874    1.206623
#> 2          6.1     setosa    3.059603     3.768874    1.206623
#> 3          7.9     setosa    3.059603     3.768874    1.206623
#> 4          4.3 versicolor    3.059603     3.768874    1.206623
#> 5          6.1 versicolor    3.059603     3.768874    1.206623
#> 6          7.9 versicolor    3.059603     3.768874    1.206623
#> 7          4.3  virginica    3.059603     3.768874    1.206623
#> 8          6.1  virginica    3.059603     3.768874    1.206623
#> 9          7.9  virginica    3.059603     3.768874    1.206623
get_datagrid(data, at = c(1, 3), length = 3)
#>   Sepal.Length Petal.Length Sepal.Width Petal.Width Species
#> 1          4.3         1.00    3.059603    1.206623  setosa
#> 2          6.1         1.00    3.059603    1.206623  setosa
#> 3          7.9         1.00    3.059603    1.206623  setosa
#> 4          4.3         3.95    3.059603    1.206623  setosa
#> 5          6.1         3.95    3.059603    1.206623  setosa
#> 6          7.9         3.95    3.059603    1.206623  setosa
#> 7          4.3         6.90    3.059603    1.206623  setosa
#> 8          6.1         6.90    3.059603    1.206623  setosa
#> 9          7.9         6.90    3.059603    1.206623  setosa
get_datagrid(data, at = c("Sepal.Length", "Species"), preserve_range = TRUE)
#>    Sepal.Length    Species Sepal.Width Petal.Length Petal.Width
#> 1           4.3     setosa    3.059603     3.768874    1.206623
#> 2           4.7     setosa    3.059603     3.768874    1.206623
#> 3           5.1     setosa    3.059603     3.768874    1.206623
#> 4           5.5     setosa    3.059603     3.768874    1.206623
#> 5           5.1 versicolor    3.059603     3.768874    1.206623
#> 6           5.5 versicolor    3.059603     3.768874    1.206623
#> 7           5.9 versicolor    3.059603     3.768874    1.206623
#> 8           6.3 versicolor    3.059603     3.768874    1.206623
#> 9           6.7 versicolor    3.059603     3.768874    1.206623
#> 10          5.1  virginica    3.059603     3.768874    1.206623
#> 11          5.5  virginica    3.059603     3.768874    1.206623
#> 12          5.9  virginica    3.059603     3.768874    1.206623
#> 13          6.3  virginica    3.059603     3.768874    1.206623
#> 14          6.7  virginica    3.059603     3.768874    1.206623
#> 15          7.1  virginica    3.059603     3.768874    1.206623
#> 16          7.5  virginica    3.059603     3.768874    1.206623
#> 17          7.9  virginica    3.059603     3.768874    1.206623
get_datagrid(data, at = c("Sepal.Length", "Species"), numerics = 0)
#>    Sepal.Length    Species Sepal.Width Petal.Length Petal.Width
#> 1           4.3     setosa           0            0           0
#> 2           4.7     setosa           0            0           0
#> 3           5.1     setosa           0            0           0
#> 4           5.5     setosa           0            0           0
#> 5           5.9     setosa           0            0           0
#> 6           6.3     setosa           0            0           0
#> 7           6.7     setosa           0            0           0
#> 8           7.1     setosa           0            0           0
#> 9           7.5     setosa           0            0           0
#> 10          7.9     setosa           0            0           0
#> 11          4.3 versicolor           0            0           0
#> 12          4.7 versicolor           0            0           0
#> 13          5.1 versicolor           0            0           0
#> 14          5.5 versicolor           0            0           0
#> 15          5.9 versicolor           0            0           0
#> 16          6.3 versicolor           0            0           0
#> 17          6.7 versicolor           0            0           0
#> 18          7.1 versicolor           0            0           0
#> 19          7.5 versicolor           0            0           0
#> 20          7.9 versicolor           0            0           0
#> 21          4.3  virginica           0            0           0
#> 22          4.7  virginica           0            0           0
#> 23          5.1  virginica           0            0           0
#> 24          5.5  virginica           0            0           0
#> 25          5.9  virginica           0            0           0
#> 26          6.3  virginica           0            0           0
#> 27          6.7  virginica           0            0           0
#> 28          7.1  virginica           0            0           0
#> 29          7.5  virginica           0            0           0
#> 30          7.9  virginica           0            0           0
get_datagrid(data, at = c("Sepal.Length = 3", "Species"))
#>   Sepal.Length    Species Sepal.Width Petal.Length Petal.Width
#> 1            3     setosa    3.059603     3.768874    1.206623
#> 2            3 versicolor    3.059603     3.768874    1.206623
#> 3            3  virginica    3.059603     3.768874    1.206623
get_datagrid(data, at = c("Sepal.Length = c(3, 1)", "Species = 'setosa'"))
#>   Sepal.Length Species Sepal.Width Petal.Length Petal.Width
#> 1            3  setosa    3.059603     3.768874    1.206623
#> 2            1  setosa    3.059603     3.768874    1.206623

# with list-style at-argument
get_datagrid(data, at = list(Sepal.Length = c(1, 3), Species = "setosa"))
#>   Sepal.Length Species Sepal.Width Petal.Length Petal.Width
#> 1            1  setosa    3.059603     3.768874    1.206623
#> 2            3  setosa    3.059603     3.768874    1.206623