Get the value at the intercept

Returns the value at the intercept (i.e., the intercept parameter), and NA if there isn't one.

get_intercept(x, ...)

Arguments

x
A model.

...
Not used.

Value

The value of the intercept.

Examples

get_intercept(lm(Sepal.Length ~ Petal.Width, data = iris))

#> [1] 4.777629

get_intercept(lm(Sepal.Length ~ 0 + Petal.Width, data = iris))

#> [1] NA

if (require("lme4")) {
  get_intercept(lme4::lmer(Sepal.Length ~ Sepal.Width + (1 | Species), data = iris))
}

#> [1] 3.406167

if (require("rstanarm")) {
  get_intercept(rstanarm::stan_glm(Sepal.Length ~ Petal.Width,
  data = iris, refresh = 0, iter = 200))
}

#> Warning: The largest R-hat is 1.05, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#r-hat

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#tail-ess

#>   [1] 4.732243 4.813314 4.861779 4.804299 4.778380 4.737728 4.760871 4.737482
#>   [9] 4.799112 4.666290 4.853937 4.843227 4.708128 4.761905 4.832012 4.842388
#>  [17] 4.663068 4.754412 4.906831 4.806599 4.881443 4.781191 4.760204 4.962490
#>  [25] 4.762044 4.856090 4.627825 4.777000 4.804846 4.682536 4.906926 4.744360
#>  [33] 4.822930 4.807385 4.671201 4.781382 4.848357 4.667826 4.813650 4.814147
#>  [41] 4.805706 4.954829 4.814549 4.736149 4.711019 4.875107 4.830930 4.877275
#>  [49] 4.780190 4.838260 4.813544 4.779363 4.652369 4.793853 4.762906 4.811559
#>  [57] 4.662182 4.896826 4.854389 4.770383 4.801141 4.760017 4.767056 4.874299
#>  [65] 4.764578 4.651108 4.713150 4.782306 4.902113 4.893844 4.750103 4.924574
#>  [73] 4.883928 4.875968 4.900860 4.718775 4.874059 4.852206 4.839456 4.761886
#>  [81] 4.762744 4.907624 4.798850 4.706131 4.826341 4.798322 4.745478 4.842398
#>  [89] 4.816969 4.689872 4.825705 4.882827 4.792662 4.710501 4.654866 4.888492
#>  [97] 4.805760 4.853414 4.780054 4.889213 4.941654 4.629145 4.812626 4.696590
#> [105] 4.720938 4.802390 4.841473 4.684030 4.746180 4.667326 4.691220 4.690906
#> [113] 4.795405 4.694107 4.792135 4.710100 4.717179 4.836711 4.842738 4.627392
#> [121] 4.650434 4.758712 4.679026 4.938503 4.843528 4.866480 4.817228 4.720641
#> [129] 4.784663 4.717391 4.794109 4.789040 4.820252 4.904821 4.855868 4.734292
#> [137] 4.789917 4.853899 4.816034 4.757068 4.797116 4.749208 4.730201 4.765605
#> [145] 4.749321 4.811421 4.754205 4.815983 4.789346 4.708869 4.642411 4.747466
#> [153] 4.804527 4.730571 4.841750 4.801902 4.753736 4.732276 4.866450 4.595097
#> [161] 4.704463 4.787550 4.750622 4.715400 4.767878 4.727785 4.756733 4.834480
#> [169] 4.745884 4.738346 4.717108 4.766033 4.807624 4.765170 4.812882 4.808942
#> [177] 4.779840 4.754410 4.780024 4.736430 4.792649 4.746543 4.757490 4.844886
#> [185] 4.750133 4.794068 4.743280 4.723778 4.710808 4.816268 4.731126 4.743412
#> [193] 4.827477 4.865611 4.800045 4.841919 4.733415 4.763835 4.782842 4.700874
#> [201] 4.716259 4.695432 4.800733 4.722872 4.739829 4.710226 4.716812 4.845437
#> [209] 4.716726 4.630735 4.765992 4.720033 4.741863 4.724462 4.865311 4.835854
#> [217] 4.846468 4.761658 4.732770 4.888204 4.725000 4.798912 4.712889 4.723858
#> [225] 4.746974 4.688460 4.829872 4.694993 4.782535 4.783573 4.839700 4.725027
#> [233] 4.799097 4.829840 4.776598 4.777370 4.755582 4.720121 4.849053 4.697372
#> [241] 4.875750 4.667626 4.865922 4.841892 4.797530 4.775466 4.761657 4.776939
#> [249] 4.742879 4.781554 4.785587 4.919612 4.941681 4.854838 4.677902 4.856045
#> [257] 4.781481 4.745401 4.831612 4.697469 4.799843 4.836350 4.910195 4.912395
#> [265] 4.825714 4.748746 4.856118 4.724290 4.750477 4.789284 4.718710 4.740714
#> [273] 4.851171 4.828126 4.665566 4.832313 4.675065 4.791008 4.692133 4.712785
#> [281] 4.732807 4.832516 4.784840 4.813855 4.881540 4.866314 4.783110 4.819247
#> [289] 4.801614 4.834530 4.771642 4.777436 4.764157 4.808690 4.752932 4.750242
#> [297] 4.725492 4.828935 4.672123 4.719100 4.803886 4.836562 4.802994 4.713444
#> [305] 4.723930 4.705260 4.841303 4.857231 4.835020 4.750805 4.891080 4.707829
#> [313] 4.821330 4.829939 4.677407 4.802231 4.732154 4.793383 4.714463 4.715975
#> [321] 4.733619 4.791703 4.752943 4.892546 4.769438 4.743722 4.782002 4.918426
#> [329] 4.846780 4.754346 4.835947 4.678530 4.698733 4.792941 4.849924 4.688766
#> [337] 4.828207 4.730759 4.948588 4.757770 4.890917 4.851772 4.796968 4.732265
#> [345] 4.747547 4.903567 4.837914 4.787384 4.747198 4.878934 4.795889 4.832492
#> [353] 4.736807 4.807933 4.787271 4.824443 4.863464 4.844574 4.792736 4.805232
#> [361] 4.687222 4.726042 4.624581 4.772250 4.746818 4.772294 4.781681 4.765855
#> [369] 4.852027 4.852546 4.719546 4.833055 4.656023 4.843536 4.812372 4.721774
#> [377] 4.870151 4.869946 4.737924 4.804637 4.830443 4.780076 4.795311 4.831899
#> [385] 4.875560 4.865100 4.935212 4.847369 4.677904 4.682949 4.682941 4.881674
#> [393] 4.890374 4.677274 4.787368 4.854764 4.805839 4.728431 4.781118 4.703499

if (require("gamm4")) {
  get_intercept(gamm4::gamm4(Sepal.Length ~ s(Petal.Width), data = iris))
}

#> Loading required package: gamm4

#> This is gamm4 0.2-6

#> [1] 5.843333