Get model parameters

Returns the coefficients (or posterior samples for Bayesian models) from a model. See the documentation for your object's class:

Bayesian models (rstanarm, brms, MCMCglmm, ...)

Estimated marginal means (emmeans)

Generalized additive models (mgcv, VGAM, ...)

Marginal effects models (mfx)

Mixed models (lme4, glmmTMB, GLMMadaptive, ...)

Zero-inflated and hurdle models (pscl, ...)

Models with special components (betareg, MuMIn, ...)

Hypothesis tests (htest)

Usage

get_parameters(x, ...)

# S3 method for default
get_parameters(x, verbose = TRUE, ...)

Arguments

x

A fitted model.

...

Currently not used.

verbose

Toggle messages and warnings.

Value

for non-Bayesian models, a data frame with two columns: the parameter names and the related point estimates.

for Anova (aov()) with error term, a list of parameters for the conditional and the random effects parameters

Details

In most cases when models either return different "effects" (fixed, random) or "components" (conditional, zero-inflated, ...), the arguments effects and component can be used.

get_parameters() is comparable to coef(), however, the coefficients are returned as data frame (with columns for names and point estimates of coefficients). For Bayesian models, the posterior samples of parameters are returned.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_parameters(m)
#>     Parameter   Estimate
#> 1 (Intercept) 38.7460642
#> 2          wt -3.2463673
#> 3         cyl -1.3641033
#> 4          vs  0.5241721