Model Predictions (robust)

The get_predicted() function is a robust, flexible and user-friendly alternative to base R predict() function. Additional features and advantages include availability of uncertainty intervals (CI), bootstrapping, a more intuitive API and the support of more models than base R's predict function. However, although the interface are simplified, it is still very important to read the documentation of the arguments. This is because making "predictions" (a lose term for a variety of things) is a non-trivial process, with lots of caveats and complications. Read the Details section for more information.

Usage

get_predicted(x, ...)

# S3 method for lm
get_predicted(
  x,
  data = NULL,
  predict = "expectation",
  iterations = NULL,
  verbose = TRUE,
  ...
)

# S3 method for stanreg
get_predicted(
  x,
  data = NULL,
  predict = "expectation",
  iterations = NULL,
  include_random = TRUE,
  include_smooth = TRUE,
  verbose = TRUE,
  ...
)

Arguments

x

A statistical model (can also be a data.frame, in which case the second argument has to be a model).

...

Other argument to be passed, for instance to get_predicted_ci(). This can be used to request confidence intervals based on robust standard errors, e.g. by specifying the vcov_* arguments from get_predicted_ci() directly in the call to get_predicted().

data

An optional data frame in which to look for variables with which to predict. If omitted, the data used to fit the model is used.

predict

string or NULL

"link" returns predictions on the model's link-scale (for logistic models, that means the log-odds scale) with a confidence interval (CI).

"expectation" (default) also returns confidence intervals, but this time the output is on the response scale (for logistic models, that means probabilities).

"prediction" also gives an output on the response scale, but this time associated with a prediction interval (PI), which is larger than a confidence interval (though it mostly make sense for linear models).

"classification" only differs from "prediction" for binomial models where it additionally transforms the predictions into the original response's type (for instance, to a factor).

Other strings are passed directly to the type argument of the predict() method supplied by the modelling package.

When predict = NULL, alternative arguments such as type will be captured by the ... ellipsis and passed directly to the predict() method supplied by the modelling package. Note that this might result in conflicts with multiple matching type arguments - thus, the recommendation is to use the predict argument for those values.

Notes: You can see the 4 options for predictions as on a gradient from "close to the model" to "close to the response data": "link", "expectation", "prediction", "classification". The predict argument modulates two things: the scale of the output and the type of certainty interval. Read more about in the Details section below.

iterations

For Bayesian models, this corresponds to the number of posterior draws. If NULL, will return all the draws (one for each iteration of the model). For frequentist models, if not NULL, will generate bootstrapped draws, from which bootstrapped CIs will be computed. Iterations can be accessed by running as.data.frame() on the output.

verbose

Toggle warnings.

include_random

If TRUE (default), include all random effects in the prediction. If FALSE, don't take them into account. Can also be a formula to specify which random effects to condition on when predicting (passed to the re.form argument). If include_random = TRUE and newdata is provided, make sure to include the random effect variables in newdata as well.

include_smooth

For General Additive Models (GAMs). If FALSE, will fix the value of the smooth to its average, so that the predictions are not depending on it. (default), mean(), or bayestestR::map_estimate().

Value

The fitted values (i.e. predictions for the response). For Bayesian or bootstrapped models (when iterations != NULL), iterations (as columns and observations are rows) can be accessed via as.data.frame.

Details

In insight::get_predicted(), the predict argument jointly modulates two separate concepts, the scale and the uncertainty interval.

Confidence Interval (CI) vs. Prediction Interval (PI))

Linear models - lm(): For linear models, Prediction intervals (predict="prediction") show the range that likely contains the value of a new observation (in what range it is likely to fall), whereas confidence intervals (predict="expectation" or predict="link") reflect the uncertainty around the estimated parameters (and gives the range of uncertainty of the regression line). In general, Prediction Intervals (PIs) account for both the uncertainty in the model's parameters, plus the random variation of the individual values. Thus, prediction intervals are always wider than confidence intervals. Moreover, prediction intervals will not necessarily become narrower as the sample size increases (as they do not reflect only the quality of the fit, but also the variability within the data).

Generalized Linear models - glm(): For binomial models, prediction intervals are somewhat useless (for instance, for a binomial (Bernoulli) model for which the dependent variable is a vector of 1s and 0s, the prediction interval is... [0, 1]).

Link scale vs. Response scale

When users set the predict argument to "expectation", the predictions are returned on the response scale, which is arguably the most convenient way to understand and visualize relationships of interest. When users set the predict argument to "link", predictions are returned on the link scale, and no transformation is applied. For instance, for a logistic regression model, the response scale corresponds to the predicted probabilities, whereas the link-scale makes predictions of log-odds (probabilities on the logit scale). Note that when users select predict="classification" in binomial models, the get_predicted() function will first calculate predictions as if the user had selected predict="expectation". Then, it will round the responses in order to return the most likely outcome.

Heteroscedasticity consistent standard errors

The arguments vcov_estimation, vcov_type and vcov_args can be used to calculate robust standard errors for confidence intervals of predictions. These arguments, when provided in get_predicted(), are passed down to get_predicted_ci(), thus, see the related documentation there for more details.

See also

Examples

data(mtcars)
x <- lm(mpg ~ cyl + hp, data = mtcars)

predictions <- get_predicted(x)
#> Warning: No random effects found in model.
#> Error in if (!all_levels_found) {    include_random <- FALSE    if (isTRUE(verbose)) {        warning(format_message("`include_random` was set to `TRUE`, but grouping factor(s) in 'data' has new levels not in the original data.",             "Either use `allow.new.levels=TRUE`, or make sure to include only valid values for grouping factor(s).",             "Setting `include_random = FALSE` now."), call. = FALSE)    }}: argument is of length zero
predictions
#> Error in eval(expr, envir, enclos): object 'predictions' not found

# Options and methods ---------------------
get_predicted(x, predict = "prediction")
#> Warning: No random effects found in model.
#> Error in if (!all_levels_found) {    include_random <- FALSE    if (isTRUE(verbose)) {        warning(format_message("`include_random` was set to `TRUE`, but grouping factor(s) in 'data' has new levels not in the original data.",             "Either use `allow.new.levels=TRUE`, or make sure to include only valid values for grouping factor(s).",             "Setting `include_random = FALSE` now."), call. = FALSE)    }}: argument is of length zero

# Get CI
as.data.frame(predictions)
#> Error in as.data.frame(predictions): object 'predictions' not found

# Bootstrapped
as.data.frame(get_predicted(x, iterations = 4))
#> Warning: No random effects found in model.
#> Error in if (!all_levels_found) {    include_random <- FALSE    if (isTRUE(verbose)) {        warning(format_message("`include_random` was set to `TRUE`, but grouping factor(s) in 'data' has new levels not in the original data.",             "Either use `allow.new.levels=TRUE`, or make sure to include only valid values for grouping factor(s).",             "Setting `include_random = FALSE` now."), call. = FALSE)    }}: argument is of length zero
summary(get_predicted(x, iterations = 4)) # Same as as.data.frame(..., keep_iterations = F)
#> Warning: No random effects found in model.
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'summary': argument is of length zero

# Different prediction types ------------------------
data(iris)
data <- droplevels(iris[1:100, ])

# Fit a logistic model
x <- glm(Species ~ Sepal.Length, data = data, family = "binomial")

# Expectation (default): response scale + CI
pred <- get_predicted(x, predict = "expectation")
#> Warning: No random effects found in model.
#> Error in if (!all_levels_found) {    include_random <- FALSE    if (isTRUE(verbose)) {        warning(format_message("`include_random` was set to `TRUE`, but grouping factor(s) in 'data' has new levels not in the original data.",             "Either use `allow.new.levels=TRUE`, or make sure to include only valid values for grouping factor(s).",             "Setting `include_random = FALSE` now."), call. = FALSE)    }}: argument is of length zero
head(as.data.frame(pred))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'pred' not found

# Prediction: response scale + PI
pred <- get_predicted(x, predict = "prediction")
#> Warning: No random effects found in model.
#> Error in if (!all_levels_found) {    include_random <- FALSE    if (isTRUE(verbose)) {        warning(format_message("`include_random` was set to `TRUE`, but grouping factor(s) in 'data' has new levels not in the original data.",             "Either use `allow.new.levels=TRUE`, or make sure to include only valid values for grouping factor(s).",             "Setting `include_random = FALSE` now."), call. = FALSE)    }}: argument is of length zero
head(as.data.frame(pred))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'pred' not found

# Link: link scale + CI
pred <- get_predicted(x, predict = "link")
#> Warning: No random effects found in model.
#> Error in if (!all_levels_found) {    include_random <- FALSE    if (isTRUE(verbose)) {        warning(format_message("`include_random` was set to `TRUE`, but grouping factor(s) in 'data' has new levels not in the original data.",             "Either use `allow.new.levels=TRUE`, or make sure to include only valid values for grouping factor(s).",             "Setting `include_random = FALSE` now."), call. = FALSE)    }}: argument is of length zero
head(as.data.frame(pred))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'pred' not found

# Classification: classification "type" + PI
pred <- get_predicted(x, predict = "classification")
#> Warning: No random effects found in model.
#> Error in if (!all_levels_found) {    include_random <- FALSE    if (isTRUE(verbose)) {        warning(format_message("`include_random` was set to `TRUE`, but grouping factor(s) in 'data' has new levels not in the original data.",             "Either use `allow.new.levels=TRUE`, or make sure to include only valid values for grouping factor(s).",             "Setting `include_random = FALSE` now."), call. = FALSE)    }}: argument is of length zero
head(as.data.frame(pred))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'head': object 'pred' not found