Predicted values

Returns values predicted by a model (i.e., fitted values). The Confidence/Credible Intervals (CI) are stored as an attribute, which one can easily extract with as.data.frame() (see examples below).

get_predicted(x, ...)

# S3 method for default
get_predicted(x, newdata = NULL, ...)

# S3 method for lm
get_predicted(
  x,
  newdata = NULL,
  ci = 0.95,
  ci_type = "confidence",
  vcov_estimation = NULL,
  vcov_type = NULL,
  vcov_args = NULL,
  ...
)

# S3 method for glm
get_predicted(x, newdata = NULL, ci = 0.95, transform = "response", ...)

# S3 method for merMod
get_predicted(
  x,
  newdata = NULL,
  ci = 0.95,
  ci_type = "confidence",
  transform = "response",
  include_random = TRUE,
  bootstrap = FALSE,
  vcov_estimation = NULL,
  vcov_type = NULL,
  vcov_args = NULL,
  ...
)

Arguments

x
A model.

...
Not used.

newdata
An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

ci
The interval level (default 0.95, i.e., 95% CI).

ci_type
Can be "prediction" or "confidence". Prediction intervals show the range that likely contains the value of a new observation (in what range it would fall), whereas confidence intervals reflect the uncertainty around the estimated parameters (and gives the range of the link; for instance of the regression line in a linear regressions). Prediction intervals account for both the uncertainty in the model's parameters, plus the random variation of the individual values. Thus, prediction intervals are always wider than confidence intervals. Moreover, prediction intervals will not necessarily become narrower as the sample size increases (as they do not reflect only the quality of the fit). This applies mostly for "simple" linear models (like lm), as for other models (e.g., glm), prediction intervals are somewhat useless (for instance, for a binomial model for which the dependent variable is a vector of 1s and 0s, the prediction interval is... [0, 1]).

vcov_estimation
String, indicating the suffix of the vcov*()-function from the sandwich or clubSandwich package, e.g. vcov_estimation = "CL" (which calls vcovCL to compute clustered covariance matrix estimators), or vcov_estimation = "HC" (which calls vcovHC() to compute heteroskedasticity-consistent covariance matrix estimators).

vcov_type
Character vector, specifying the estimation type for the robust covariance matrix estimation (see vcovHC() or clubSandwich::vcovCR() for details).

vcov_args
List of named vectors, used as additional arguments that are passed down to the sandwich-function specified in vcov_estimation.

transform
Either "response" (default) or "link". If "link", no transformation is applied and the values are on the scale of the linear predictors. If "response", the output is on the scale of the response variable. Thus for a default binomial model, "response" gives the predicted probabilities, and "link" makes predictions of log-odds (probabilities on logit scale).

include_random
If TRUE (default), include all random effects in the prediction. If FALSE, don't take them into account. Can also be a formula to specify which random effects to condition on when predicting (passed to the re.form argument). If include_random = TRUE and newdata is provided, make sure to include the random effect variables in newdata as well.

bootstrap
Should confidence intervals (CIs) be computed via bootstrapping rather than analytically. If TRUE, you can specify the number of iterations by modifying the argument iter = 500 (default).

Value

The fitted values (i.e. predictions for the response).

Note

Currently, this function just calls stats::fitted(), but will be extended to other objects that don't work with stats::fitted() in future updates.

Examples

data(mtcars)
x <- lm(mpg ~ cyl + hp, data = mtcars)
predicted <- get_predicted(x)

#> Error in model.frame.default(object, data, xlev = xlev): 'data' must be a data.frame, environment, or list

predicted

#> Error in eval(expr, envir, enclos): object 'predicted' not found

# Get CI
attributes(predicted)$CI_low # Or CI_high

#> Error in eval(expr, envir, enclos): object 'predicted' not found

as.data.frame(predicted)

#> Error in as.data.frame(predicted): object 'predicted' not found