Compute Model's Predictions

Compute Model's Predictions.

get_predicted(x, data = NULL, ...)

# S3 method for lm
get_predicted(x, data = NULL, predict = "relation", iterations = NULL, ...)

# S3 method for stanreg
get_predicted(
  x,
  data = NULL,
  predict = "relation",
  iterations = NULL,
  include_random = TRUE,
  include_smooth = TRUE,
  centrality_function = stats::median,
  ...
)

Arguments

x
A statistical model (can also be a data.frame, in which case the second argument has to be a model).

data
An optional data frame in which to look for variables with which to predict. If omitted, the data used to fit the model is used.

...
Other argument to be passed for instance to get_predicted_ci.

predict
Can be "link", "relation" (default), or "prediction". This modulates the scale of the output as well as the type of certainty interval. More specifically, "link" gives an output on the link-scale (for logistic models, that means the log-odds scale) with a confidence interval (CI). "relation" (default) also returns confidence intervals, but this time the output is on the response scale (for logistic models, that means probabilities). Finally, "predict" also gives an output on the response scale, but this time associated with a prediction interval (PI), which is larger than a confidence interval (though it mostly make sense for linear models). gives Read more about in the Details section below.

iterations
For Bayesian models, this corresponds to the number of posterior draws. If NULL, will return all the draws (one for each iteration of the model). For frequentist models, if not NULL, will generate bootstrapped draws, from which bootstrapped CIs will be computed.

include_random
If TRUE (default), include all random effects in the prediction. If FALSE, don't take them into account. Can also be a formula to specify which random effects to condition on when predicting (passed to the re.form argument). If include_random = TRUE and newdata is provided, make sure to include the random effect variables in newdata as well.

include_smooth
For General Additive Models (GAMs). If FALSE, will fix the value of the smooth to its average, so that the predictions are not depending on it.

centrality_function
The function used to obtain a centrality estimate for bootstrapped or Bayesian models. Usually median() (default), mean(), or bayestestR::map_estimate().

Value

The fitted values (i.e. predictions for the response). For Bayesian or bootstrapped models (when iterations != NULL), this will be a dataframe with all iterations as columns (observations are still rows).

Details

The predict argument jointly modulates two separate concepts, the scale and the uncertainty interval.

Confidence Interval vs. Prediction Interval)

Linear models - lm(): For linear models, Prediction intervals (predict = "prediction") show the range that likely contains the value of a new observation (in what range it is likely to fall), whereas confidence intervals (predict = "relation" or predict = "link") reflect the uncertainty around the estimated parameters (and gives the range of uncertainty of the regression line). lIn general, Prediction Intervals (PIs) account for both the uncertainty in the model's parameters, plus the random variation of the individual values. Thus, prediction intervals are always wider than confidence intervals. Moreover, prediction intervals will not necessarily become narrower as the sample size increases (as they do not reflect only the quality of the fit, but also the variability within the data).

General Linear models - glm(): For binomial models, prediction intervals are somewhat useless (for instance, for a binomial model for which the dependent variable is a vector of 1s and 0s, the prediction interval is... [0, 1]).

Link scale vs. Response scale

Having the output is on the scale of the response variable is arguably the most convenient to understand and visualize the relationships. If on the link-scale, no transformation is applied and the values are on the scale of the model's predictors. For instance, for a logistic model, the response scale corresponds to the predicted probabilities, whereas the link-scale makes predictions of log-odds (probabilities on the logit scale).

See also

Examples

data(mtcars)
x <- lm(mpg ~ cyl + hp, data = mtcars)
predictions <- get_predicted(x)
predictions

#> Predicted values:
#> 
#>  [1] 21.21678 21.21678 26.07124 21.21678 15.44448 21.31239 14.10597 26.66401
#>  [9] 26.03299 20.96820 20.96820 15.34888 15.34888 15.34888 14.87083 14.67962
#> [17] 14.39279 26.58752 26.85523 26.60665 25.99475 15.92253 15.92253 14.10597
#> [25] 15.44448 26.58752 26.10948 25.68880 13.74265 19.97387 12.38501 25.76529
#> 
#> NOTE: Confidence intervals, if available, are stored as attributes and can be acccessed using `as.data.frame()` on this output.

get_predicted(x, predict = "prediction")

#> Predicted values:
#> 
#>  [1] 21.21678 21.21678 26.07124 21.21678 15.44448 21.31239 14.10597 26.66401
#>  [9] 26.03299 20.96820 20.96820 15.34888 15.34888 15.34888 14.87083 14.67962
#> [17] 14.39279 26.58752 26.85523 26.60665 25.99475 15.92253 15.92253 14.10597
#> [25] 15.44448 26.58752 26.10948 25.68880 13.74265 19.97387 12.38501 25.76529
#> 
#> NOTE: Confidence intervals, if available, are stored as attributes and can be acccessed using `as.data.frame()` on this output.

# Get CI
as.data.frame(predictions)

#>                     Predicted        SE    CI_low  CI_high
#> Mazda RX4            21.21678 0.7281647 19.727518 22.70605
#> Mazda RX4 Wag        21.21678 0.7281647 19.727518 22.70605
#> Datsun 710           26.07124 0.9279509 24.173366 27.96911
#> Hornet 4 Drive       21.21678 0.7281647 19.727518 22.70605
#> Hornet Sportabout    15.44448 0.9200310 13.562810 17.32616
#> Valiant              21.31239 0.7777664 19.721680 22.90310
#> Duster 360           14.10597 1.0080670 12.044237 16.16769
#> Merc 240D            26.66401 0.9225132 24.777260 28.55076
#> Merc 230             26.03299 0.9362657 24.118117 27.94787
#> Merc 280             20.96820 0.6234320 19.693139 22.24326
#> Merc 280C            20.96820 0.6234320 19.693139 22.24326
#> Merc 450SE           15.34888 0.8862558 13.536280 17.16147
#> Merc 450SL           15.34888 0.8862558 13.536280 17.16147
#> Merc 450SLC          15.34888 0.8862558 13.536280 17.16147
#> Cadillac Fleetwood   14.87083 0.8057154 13.222961 16.51871
#> Lincoln Continental  14.67962 0.8206255 13.001249 16.35798
#> Chrysler Imperial    14.39279 0.8911693 12.570146 16.21544
#> Fiat 128             26.58752 0.9099596 24.726448 28.44860
#> Honda Civic          26.85523 0.9695585 24.872258 28.83820
#> Toyota Corolla       26.60665 0.9127445 24.739874 28.47342
#> Toyota Corona        25.99475 0.9454598 24.061069 27.92843
#> Dodge Challenger     15.92253 1.1490264 13.572504 18.27255
#> AMC Javelin          15.92253 1.1490264 13.572504 18.27255
#> Camaro Z28           14.10597 1.0080670 12.044237 16.16769
#> Pontiac Firebird     15.44448 0.9200310 13.562810 17.32616
#> Fiat X1-9            26.58752 0.9099596 24.726448 28.44860
#> Porsche 914-2        26.10948 0.9205392 24.226768 27.99220
#> Lotus Europa         25.68880 1.0474287 23.546572 27.83104
#> Ford Pantera L       13.74265 1.2011595 11.286007 16.19930
#> Ferrari Dino         19.97387 0.7635547 18.412227 21.53552
#> Maserati Bora        12.38501 2.1153615  8.058613 16.71141
#> Volvo 142E           25.76529 1.0175965 23.684073 27.84651

# Bootsrapped
as.data.frame(get_predicted(x, iterations = 100), include_iterations = FALSE)

#>    Predicted        SE    CI_low  CI_high
#> 1   21.28226 0.7294724 19.932848 22.69425
#> 2   21.28226 0.7294724 19.932848 22.69425
#> 3   26.19968 1.2159623 23.709626 28.49205
#> 4   21.28226 0.7294724 19.932848 22.69425
#> 5   15.27344 0.8024721 13.415287 16.55061
#> 6   21.38737 0.7815932 19.941101 22.83304
#> 7   13.93067 0.9679625 11.299462 15.16234
#> 8   26.92832 1.1865444 24.597097 29.17495
#> 9   26.17054 1.2249107 23.645951 28.44331
#> 10  21.02098 0.6234041 19.693782 22.08264
#> 11  21.02098 0.6234041 19.693782 22.08264
#> 12  15.15017 0.7614376 13.399981 16.38366
#> 13  15.15017 0.7614376 13.399981 16.38366
#> 14  15.15017 0.7614376 13.399981 16.38366
#> 15  14.65710 0.6747630 13.194831 15.63884
#> 16  14.51917 0.7050963 12.753683 15.46338
#> 17  14.16608 0.8112578 12.174440 15.27347
#> 18  26.84280 1.1784427 24.553484 29.08683
#> 19  27.22592 1.2216581 24.645970 29.39523
#> 20  26.86741 1.1801405 24.582436 29.10886
#> 21  26.12940 1.2346429 23.582277 28.39188
#> 22  15.85051 1.0776006 13.710199 17.93704
#> 23  15.85051 1.0776006 13.710199 17.93704
#> 24  13.93067 0.9679625 11.299462 15.16234
#> 25  15.27344 0.8024721 13.415287 16.55061
#> 26  26.84280 1.1784427 24.553484 29.08683
#> 27  26.23290 1.2078150 23.791173 28.53611
#> 28  25.70587 1.3383766 23.010074 28.05897
#> 29  13.59792 1.2080027 10.191156 15.06412
#> 30  19.93928 0.8364845 17.834366 21.18902
#> 31  12.16881 2.2548471  6.019652 14.71227
#> 32  25.79529 1.3084173 23.157365 28.11547

if (FALSE) {
# Bayesian models
if (require("rstanarm")) {
  x <- stan_glm(mpg ~ am, data = mtcars, refresh = 0)
  predictions <- get_predicted(x)
  predictions
  as.data.frame(predictions, include_iterations = FALSE)
}
}