Predicted values

Returns values predicted by a model (i.e., fitted values). The Confidence/Credible Intervals (CI) are stored as an attribute, which one can easily extract with as.data.frame() (see examples below).

get_predicted(x, ...)

# S3 method for default
get_predicted(x, newdata = NULL, ...)

# S3 method for lm
get_predicted(x, newdata = NULL, ci = 0.95, ci_type = "confidence", ...)

# S3 method for glm
get_predicted(x, newdata = NULL, ci = 0.95, transform = "response", ...)

# S3 method for merMod
get_predicted(
  x,
  newdata = NULL,
  ci = 0.95,
  ci_type = "confidence",
  transform = "response",
  include_random = TRUE,
  bootstrap = FALSE,
  ...
)

Arguments

x
A model.

...
Not used.

newdata
An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

ci
The interval level (default 0.95, i.e., 95% CI).

ci_type
Can be "prediction" or "confidence". Prediction intervals show the range that likely contains the value of a new observation (in what range it would fall), whereas confidence intervals reflect the uncertainty around the estimated parameters (and gives the range of the link; for instance of the regression line in a linear regressions). Prediction intervals account for both the uncertainty in the model's parameters, plus the random variation of the individual values. Thus, prediction intervals are always wider than confidence intervals. Moreover, prediction intervals will not necessarily become narrower as the sample size increases (as they do not reflect only the quality of the fit). This applies mostly for "simple" linear models (like lm), as for other models (e.g., glm), prediction intervals are somewhat useless (for instance, for a binomial model for which the dependent variable is a vector of 1s and 0s, the prediction interval is... [0, 1]).

transform
Either "response" (default) or "link". If "link", no transformation is applied and the values are on the scale of the linear predictors. If "response", the output is on the scale of the response variable. Thus for a default binomial model, "response" gives the predicted probabilities, and "link" makes predictions of log-odds (probabilities on logit scale).

include_random
If TRUE (default), include all random effects in the prediction. If FALSE, don't take them into account. Can also be a formula to specify which random effects to condition on when predicting (passed to the re.form argument). If include_random = TRUE and newdata is provided, make sure to include the random effect variables in newdata as well.

bootstrap
Should confidence intervals (CIs) be computed via bootstrapping rather than analytically. If TRUE, you can specify the number of iterations by modifying the argument iter = 500 (default).

Value

The fitted values (i.e. predictions for the response).

Note

Currently, this function just calls stats::fitted(), but will be extended to other objects that don't work with stats::fitted() in future updates.

Examples

data(mtcars)
x <- lm(mpg ~ cyl + hp, data = mtcars)
predicted <- get_predicted(x)
predicted

#> Predicted values:
#> 
#>  [1] 21.21678 21.21678 26.07124 21.21678 15.44448 21.31239 14.10597 26.66401
#>  [9] 26.03299 20.96820 20.96820 15.34888 15.34888 15.34888 14.87083 14.67962
#> [17] 14.39279 26.58752 26.85523 26.60665 25.99475 15.92253 15.92253 14.10597
#> [25] 15.44448 26.58752 26.10948 25.68880 13.74265 19.97387 12.38501 25.76529
#> 
#> NOTE: Confidence intervals, if available, are stored as attributes and can be acccessed using `as.data.frame()` on this output.

# Get CI
attributes(predicted)$CI_low # Or CI_high

#>  [1] 19.727518 19.727518 24.173366 19.727518 13.562810 19.721680 12.044237
#>  [8] 24.777260 24.118117 19.693139 19.693139 13.536280 13.536280 13.536280
#> [15] 13.222961 13.001249 12.570146 24.726448 24.872258 24.739874 24.061069
#> [22] 13.572504 13.572504 12.044237 13.562810 24.726448 24.226768 23.546572
#> [29] 11.286007 18.412227  8.058613 23.684073

as.data.frame(predicted)

#>    Predicted        SE   CI    CI_low  CI_high
#> 1   21.21678 0.7281647 0.95 19.727518 22.70605
#> 2   21.21678 0.7281647 0.95 19.727518 22.70605
#> 3   26.07124 0.9279509 0.95 24.173366 27.96911
#> 4   21.21678 0.7281647 0.95 19.727518 22.70605
#> 5   15.44448 0.9200310 0.95 13.562810 17.32616
#> 6   21.31239 0.7777664 0.95 19.721680 22.90310
#> 7   14.10597 1.0080670 0.95 12.044237 16.16769
#> 8   26.66401 0.9225132 0.95 24.777260 28.55076
#> 9   26.03299 0.9362657 0.95 24.118117 27.94787
#> 10  20.96820 0.6234320 0.95 19.693139 22.24326
#> 11  20.96820 0.6234320 0.95 19.693139 22.24326
#> 12  15.34888 0.8862558 0.95 13.536280 17.16147
#> 13  15.34888 0.8862558 0.95 13.536280 17.16147
#> 14  15.34888 0.8862558 0.95 13.536280 17.16147
#> 15  14.87083 0.8057154 0.95 13.222961 16.51871
#> 16  14.67962 0.8206255 0.95 13.001249 16.35798
#> 17  14.39279 0.8911693 0.95 12.570146 16.21544
#> 18  26.58752 0.9099596 0.95 24.726448 28.44860
#> 19  26.85523 0.9695585 0.95 24.872258 28.83820
#> 20  26.60665 0.9127445 0.95 24.739874 28.47342
#> 21  25.99475 0.9454598 0.95 24.061069 27.92843
#> 22  15.92253 1.1490264 0.95 13.572504 18.27255
#> 23  15.92253 1.1490264 0.95 13.572504 18.27255
#> 24  14.10597 1.0080670 0.95 12.044237 16.16769
#> 25  15.44448 0.9200310 0.95 13.562810 17.32616
#> 26  26.58752 0.9099596 0.95 24.726448 28.44860
#> 27  26.10948 0.9205392 0.95 24.226768 27.99220
#> 28  25.68880 1.0474287 0.95 23.546572 27.83104
#> 29  13.74265 1.2011595 0.95 11.286007 16.19930
#> 30  19.97387 0.7635547 0.95 18.412227 21.53552
#> 31  12.38501 2.1153615 0.95  8.058613 16.71141
#> 32  25.76529 1.0175965 0.95 23.684073 27.84651