Confidence and Prediction Interval for Model Predictions

Returns the Confidence (or Prediction) Interval (CI) associated with predictions made by a model.

get_predicted_ci(
  x,
  predictions = NULL,
  data = NULL,
  ci = 0.95,
  ci_type = "confidence",
  vcov_estimation = NULL,
  vcov_type = NULL,
  vcov_args = NULL,
  ...
)

Arguments

x
A statistical model (can also be a data.frame, in which case the second argument has to be a model).

predictions
A vector of predicted values (as obtained by stats::fitted(), stats::predict() or get_predicted).

data
An optional data frame in which to look for variables with which to predict. If omitted, the data used to fit the model is used.

ci
The interval level (default 0.95, i.e., 95% CI).

ci_type
Can be "prediction" or "confidence". Prediction intervals show the range that likely contains the value of a new observation (in what range it would fall), whereas confidence intervals reflect the uncertainty around the estimated parameters (and gives the range of the link; for instance of the regression line in a linear regressions). Prediction intervals account for both the uncertainty in the model's parameters, plus the random variation of the individual values. Thus, prediction intervals are always wider than confidence intervals. Moreover, prediction intervals will not necessarily become narrower as the sample size increases (as they do not reflect only the quality of the fit). This applies mostly for "simple" linear models (like lm), as for other models (e.g., glm), prediction intervals are somewhat useless (for instance, for a binomial model for which the dependent variable is a vector of 1s and 0s, the prediction interval is... [0, 1]).

vcov_estimation
String, indicating the suffix of the vcov*()-function from the sandwich or clubSandwich package, e.g. vcov_estimation = "CL" (which calls vcovCL to compute clustered covariance matrix estimators), or vcov_estimation = "HC" (which calls vcovHC() to compute heteroskedasticity-consistent covariance matrix estimators).

vcov_type
Character vector, specifying the estimation type for the robust covariance matrix estimation (see vcovHC() or clubSandwich::vcovCR() for details).

vcov_args
List of named vectors, used as additional arguments that are passed down to the sandwich-function specified in vcov_estimation.

...
Not used for now.

Value

The Confidence (or Prediction) Interval (CI).

Examples

data(mtcars)

# Linear model
x <- lm(mpg ~ cyl + hp, data = mtcars)
predictions <- predict(x)
get_predicted_ci(x, predictions, ci_type = "prediction")

#>                           SE    CI_low  CI_high
#> Mazda RX4           3.255505 14.558527 27.87504
#> Mazda RX4 Wag       3.255505 14.558527 27.87504
#> Datsun 710          3.305931 19.309850 32.83263
#> Hornet 4 Drive      3.255505 14.558527 27.87504
#> Hornet Sportabout   3.303717  8.687625 22.20134
#> Valiant             3.266957 14.630713 27.99407
#> Duster 360          3.329307  7.296768 20.91516
#> Merc 240D           3.304409 19.905736 33.42229
#> Merc 230            3.308275 19.266814 32.79918
#> Merc 280            3.233691 14.354560 27.58184
#> Merc 280C           3.233691 14.354560 27.58184
#> Merc 450SE          3.294471  8.610927 22.08683
#> Merc 450SL          3.294471  8.610927 22.08683
#> Merc 450SLC         3.294471  8.610927 22.08683
#> Cadillac Fleetwood  3.273723  8.175318 21.56635
#> Lincoln Continental 3.277425  7.976530 21.38270
#> Chrysler Imperial   3.295796  7.652132 21.13345
#> Fiat 128            3.300926 19.836372 33.33868
#> Honda Civic         3.317850 20.069462 33.64099
#> Toyota Corolla      3.301695 19.853921 33.35937
#> Toyota Corona       3.310888 19.223224 32.76628
#> Dodge Challenger    3.374663  9.020567 22.82449
#> AMC Javelin         3.374663  9.020567 22.82449
#> Camaro Z28          3.329307  7.296768 20.91516
#> Pontiac Firebird    3.303717  8.687625 22.20134
#> Fiat X1-9           3.300926 19.836372 33.33868
#> Porsche 914-2       3.303858 19.352333 32.86663
#> Lotus Europa        3.341436 18.854800 32.52281
#> Ford Pantera L      3.392768  6.803665 20.68164
#> Ferrari Dino        3.263603 13.299055 26.64869
#> Maserati Bora       3.813508  4.585514 20.18451
#> Volvo 142E          3.332205 18.950167 32.58042

get_predicted_ci(x, predictions, ci_type = "confidence")

#>                            SE    CI_low  CI_high
#> Mazda RX4           0.7281647 19.727518 22.70605
#> Mazda RX4 Wag       0.7281647 19.727518 22.70605
#> Datsun 710          0.9279509 24.173366 27.96911
#> Hornet 4 Drive      0.7281647 19.727518 22.70605
#> Hornet Sportabout   0.9200310 13.562810 17.32616
#> Valiant             0.7777664 19.721680 22.90310
#> Duster 360          1.0080670 12.044237 16.16769
#> Merc 240D           0.9225132 24.777260 28.55076
#> Merc 230            0.9362657 24.118117 27.94787
#> Merc 280            0.6234320 19.693139 22.24326
#> Merc 280C           0.6234320 19.693139 22.24326
#> Merc 450SE          0.8862558 13.536280 17.16147
#> Merc 450SL          0.8862558 13.536280 17.16147
#> Merc 450SLC         0.8862558 13.536280 17.16147
#> Cadillac Fleetwood  0.8057154 13.222961 16.51871
#> Lincoln Continental 0.8206255 13.001249 16.35798
#> Chrysler Imperial   0.8911693 12.570146 16.21544
#> Fiat 128            0.9099596 24.726448 28.44860
#> Honda Civic         0.9695585 24.872258 28.83820
#> Toyota Corolla      0.9127445 24.739874 28.47342
#> Toyota Corona       0.9454598 24.061069 27.92843
#> Dodge Challenger    1.1490264 13.572504 18.27255
#> AMC Javelin         1.1490264 13.572504 18.27255
#> Camaro Z28          1.0080670 12.044237 16.16769
#> Pontiac Firebird    0.9200310 13.562810 17.32616
#> Fiat X1-9           0.9099596 24.726448 28.44860
#> Porsche 914-2       0.9205392 24.226768 27.99220
#> Lotus Europa        1.0474287 23.546572 27.83104
#> Ford Pantera L      1.2011595 11.286007 16.19930
#> Ferrari Dino        0.7635547 18.412227 21.53552
#> Maserati Bora       2.1153615  8.058613 16.71141
#> Volvo 142E          1.0175965 23.684073 27.84651

predictions <- get_predicted(x, iterations = 500)
get_predicted_ci(x, predictions)

#>           SE    CI_low  CI_high
#> 1  0.7031546 19.934530 22.67952
#> 2  0.7031546 19.934530 22.67952
#> 3  1.1750487 23.801814 28.36394
#> 4  0.7031546 19.934530 22.67952
#> 5  0.7225206 13.974464 16.80481
#> 6  0.7459755 19.969942 22.94548
#> 7  0.9627085 11.492710 15.31728
#> 8  1.1523062 24.477470 28.97491
#> 9  1.1827347 23.764067 28.33274
#> 10 0.6206705 19.787201 22.23289
#> 11 0.6206705 19.787201 22.23289
#> 12 0.6917564 13.890161 16.67008
#> 13 0.6917564 13.890161 16.67008
#> 14 0.6917564 13.890161 16.67008
#> 15 0.6542993 13.407237 16.08918
#> 16 0.6970883 13.058719 15.89435
#> 17 0.8100814 12.263706 15.59144
#> 18 1.1448026 24.418809 28.93036
#> 19 1.1841021 24.624125 29.19990
#> 20 1.1463916 24.433474 28.95774
#> 21 1.1911197 23.690444 28.31186
#> 22 0.9522121 14.163479 17.95865
#> 23 0.9522121 14.163479 17.95865
#> 24 0.9627085 11.492710 15.31728
#> 25 0.7225206 13.974464 16.80481
#> 26 1.1448026 24.418809 28.93036
#> 27 1.1680755 23.865687 28.39933
#> 28 1.2814755 23.162466 28.14862
#> 29 1.1889021 10.317412 15.13272
#> 30 0.8376374 18.091193 21.44328
#> 31 2.1577888  6.167162 14.63238
#> 32 1.2552411 23.293597 28.18889

# Logistic model
x <- glm(vs ~ wt, data = mtcars, family = "binomial")
predictions <- predict(x, type = "link")
get_predicted_ci(x, predictions, ci_type = "prediction")

#>                           SE     CI_low    CI_high
#> Mazda RX4           1.166957 -1.5781474  2.9962405
#> Mazda RX4 Wag       1.124961 -1.9830256  2.4267393
#> Datsun 710          1.250300 -1.1683313  3.7327530
#> Hornet 4 Drive      1.115523 -2.6141137  1.7586550
#> Hornet Sportabout   1.139191 -3.0903759  1.3751707
#> Valiant             1.142408 -3.1348921  1.3432650
#> Duster 360          1.163204 -3.3858108  1.1738632
#> Merc 240D           1.114349 -2.5640491  1.8041177
#> Merc 230            1.113087 -2.4851538  1.8780663
#> Merc 280            1.139191 -3.0903759  1.3751707
#> Merc 280C           1.139191 -3.0903759  1.3751707
#> Merc 450SE          1.316278 -4.6411056  0.5186103
#> Merc 450SL          1.202361 -3.7682463  0.9449234
#> Merc 450SLC         1.216628 -3.8917355  0.8773579
#> Cadillac Fleetwood  1.917997 -8.0748998 -0.5564881
#> Lincoln Continental 2.022503 -8.6121623 -0.6840963
#> Chrysler Imperial   1.974734 -8.3676029 -0.6267891
#> Fiat 128            1.292493 -1.0217631  4.0447162
#> Honda Civic         1.553094 -0.4148607  5.6731547
#> Toyota Corolla      1.445916 -0.6251175  5.0427705
#> Toyota Corona       1.205806 -1.3581548  3.3685177
#> Dodge Challenger    1.153108 -3.2704974  1.2496046
#> AMC Javelin         1.138414 -3.0793011  1.3832014
#> Camaro Z28          1.234947 -4.0422724  0.7986290
#> Pontiac Firebird    1.236530 -4.0549296  0.7921808
#> Fiat X1-9           1.400545 -0.7272466  4.7627900
#> Porsche 914-2       1.315259 -0.9517499  4.2039688
#> Lotus Europa        1.605779 -0.3232470  5.9712927
#> Ford Pantera L      1.113623 -2.5244153  1.8409058
#> Ferrari Dino        1.138782 -1.8095074  2.6544362
#> Maserati Bora       1.163204 -3.3858108  1.1738632
#> Volvo 142E          1.137252 -1.8256134  2.6323312

get_predicted_ci(x, predictions, ci_type = "confidence")

#>                            SE       CI_low     CI_high
#> Mazda RX4           0.5623444 -0.393128219  1.81122128
#> Mazda RX4 Wag       0.4690190 -0.697403437  1.14111716
#> Datsun 710          0.7195076 -0.127998221  2.69241991
#> Hornet 4 Drive      0.4459072 -1.301691317  0.44623260
#> Hornet Sportabout   0.5021936 -1.841883910  0.12667872
#> Valiant             0.5094490 -1.894315229  0.10268813
#> Duster 360          0.5545130 -2.192799377 -0.01914822
#> Merc 240D           0.4429620 -1.248155280  0.48822395
#> Merc 230            0.4397777 -1.165492290  0.55840477
#> Merc 280            0.5021936 -1.841883910  0.12667872
#> Merc 280C           0.5021936 -1.841883910  0.12667872
#> Merc 450SE          0.8288732 -3.685809195 -0.43668612
#> Merc 450SL          0.6325464 -2.651429636 -0.17189323
#> Merc 450SLC         0.6592613 -2.799317246 -0.21506039
#> Cadillac Fleetwood  1.6227003 -7.496128128 -1.13525980
#> Lincoln Continental 1.7449815 -8.068230227 -1.22802830
#> Chrysler Imperial   1.6893831 -7.808326079 -1.18606591
#> Fiat 128            0.7905569 -0.037986554  3.06093970
#> Honda Civic         1.1689920  0.337964764  4.92032921
#> Toyota Corolla      1.0223094  0.205136978  4.21251600
#> Toyota Corona       0.6390698 -0.247372450  2.25773530
#> Dodge Challenger    0.5330115 -2.055129814  0.03423699
#> AMC Javelin         0.5004295 -1.828873572  0.13277386
#> Camaro Z28          0.6924848 -2.979066901 -0.26457646
#> Pontiac Firebird    0.6953056 -2.994148306 -0.26860054
#> Fiat X1-9           0.9570628  0.141963003  3.89358043
#> Porsche 914-2       0.8272529  0.004723634  3.24749524
#> Lotus Europa        1.2381316  0.397329560  5.25071615
#> Ford Pantera L      0.4411326 -1.206358704  0.52284928
#> Ferrari Dino        0.5012652 -0.559997424  1.40492617
#> Maserati Bora       0.5545130 -2.192799377 -0.01914822
#> Volvo 142E          0.4977787 -0.572269459  1.37898725