get_response {insight}	R Documentation

Get the values from the response variable

Description

Returns the values the response variable(s) from a model object. If the model is a multivariate response model, a data frame with values from all response variables is returned.

Usage

get_response(x, select = NULL, verbose = TRUE)

Arguments

x
A fitted model.

select
Optional name(s) of response variables for which to extract values. Can be used in case of regression models with multiple response variables.

verbose
Toggle warnings.

Value

The values of the response variable, as vector, or a data frame if x has more than one defined response variable.

Examples

if (require("lme4")) {
  data(cbpp)
  cbpp$trials <- cbpp$size - cbpp$incidence

  m <- glm(cbind(incidence, trials) ~ period, data = cbpp, family = binomial)
  head(get_response(m))
  get_response(m, select = "incidence")
}

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_response(m)

[Package insight version 0.14.0 Index]