Get statistic associated with estimates

Returns the statistic (t, z, ...) for model estimates. In most cases, this is the related column from coef(summary()).

get_statistic(x, ...)

# S3 method for default
get_statistic(x, column_index = 3, ...)

# S3 method for glmmTMB
get_statistic(
  x,
  component = c("all", "conditional", "zi", "zero_inflated"),
  ...
)

# S3 method for clm2
get_statistic(x, component = c("all", "conditional", "scale"), ...)

# S3 method for gee
get_statistic(x, robust = FALSE, ...)

# S3 method for betareg
get_statistic(x, component = c("all", "conditional", "precision"), ...)

Arguments

x
A model.

...
Currently not used.

column_index
For model objects that have no defined get_statistic() method yet, the default method is called. This method tries to extract the statistic column from coef(summary()), where the index of the column that is being pulled is column_index. Defaults to 3, which is the default statistic column for most models' summary-output.

component
Should all parameters, parameters for the conditional model, or for the zero-inflated part of the model be returned? Applies to models with zero-inflated component. component may be one of "conditional", "zi", "zero-inflated" or "all" (default). For models with smooth terms, component = "smooth_terms" is also possible. May be abbreviated. Note that the conditional component is also called count or mean component, depending on the model.

robust
Logical, if TRUE, test statistic based on robust standard errors is returned.

Value

A data frame with the model's parameter names and the related test statistic.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)
get_statistic(m)

#>     Parameter  Statistic
#> 1 (Intercept) 11.3994647
#> 2          wt -4.1204121
#> 3         cyl -2.2234114
#> 4          vs  0.3221477