Get the values from model weights

Returns weighting variable of a model.

get_weights(x, na_rm = FALSE, null_as_ones = FALSE, ...)

Arguments

x
A fitted model.

na_rm
Logical, if TRUE, removes possible missing values.

null_as_ones
Logical, if TRUE, will return a vector of 1 if no weights were specified in the model (as if the weights were all set to 1).

...
Currently not used.

Value

The weighting variable, or NULL if no weights were specified (or if all weights were 1).

Examples

data(mtcars)
mtcars$weight <- rnorm(nrow(mtcars), 1, .3)

# LMs
m <- lm(mpg ~ wt + cyl + vs, data = mtcars, weights = weight)
get_weights(m)

#>  [1] 1.1856588 1.0305780 0.8983399 0.9594662 0.8835298 0.4229393 0.5809255
#>  [8] 1.2154194 1.1912607 1.8176345 0.5527630 0.8832922 0.9884347 1.1691851
#> [15] 0.6833751 1.0654327 0.7619022 1.0007476 0.5238288 0.7285948 0.7711802
#> [22] 1.1196169 1.3485833 0.7331707 0.7279675 0.9834736 0.8918233 1.0123316
#> [29] 0.8219752 0.8506929 0.8465789 0.9516581

get_weights(lm(mpg ~ wt, data = mtcars), null_as_ones = TRUE)

#>  [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1

# GLMs
m <- glm(vs ~ disp + mpg, data = mtcars, weights = weight, family=quasibinomial)
get_weights(m)

#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>           1.1856588           1.0305780           0.8983399           0.9594662 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>           0.8835298           0.4229393           0.5809255           1.2154194 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>           1.1912607           1.8176345           0.5527630           0.8832922 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>           0.9884347           1.1691851           0.6833751           1.0654327 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>           0.7619022           1.0007476           0.5238288           0.7285948 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>           0.7711802           1.1196169           1.3485833           0.7331707 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>           0.7279675           0.9834736           0.8918233           1.0123316 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>           0.8219752           0.8506929           0.8465789           0.9516581 

m <- glm(cbind(cyl, gear) ~ mpg, data = mtcars, weights = weight, family = binomial)
get_weights(m)

#>           Mazda RX4       Mazda RX4 Wag          Datsun 710      Hornet 4 Drive 
#>           11.856588           10.305780            7.186719            8.635196 
#>   Hornet Sportabout             Valiant          Duster 360           Merc 240D 
#>            9.718828            3.806454            6.390181            9.723355 
#>            Merc 230            Merc 280           Merc 280C          Merc 450SE 
#>            9.530085           18.176345            5.527630            9.716214 
#>          Merc 450SL         Merc 450SLC  Cadillac Fleetwood Lincoln Continental 
#>           10.872782           12.861036            7.517126           11.719760 
#>   Chrysler Imperial            Fiat 128         Honda Civic      Toyota Corolla 
#>            8.380924            8.005981            4.190630            5.828759 
#>       Toyota Corona    Dodge Challenger         AMC Javelin          Camaro Z28 
#>            5.398261           12.315786           14.834416            8.064877 
#>    Pontiac Firebird           Fiat X1-9       Porsche 914-2        Lotus Europa 
#>            8.007643            7.867789            8.026410            9.110985 
#>      Ford Pantera L        Ferrari Dino       Maserati Bora          Volvo 142E 
#>           10.685678            9.357622           11.005525            7.613265