Checks if a regression model object is supported in insight package

Small helper that checks if a model is a supported (regression) model object. supported_models() prints a list of currently supported model classes.

Usage

is_model_supported(x)

supported_models()

Arguments

x

An object.

Value

A logical, TRUE if x is a (supported) model object.

Details

This function returns TRUE if x is a model object that works with the package's functions. A list of supported models can also be found here: https://github.com/easystats/insight.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)

is_model_supported(m)
#> [1] TRUE
is_model_supported(mtcars)
#> [1] FALSE

# to see all supported models
supported_models()
#>   [1] "AKP"                     "Anova.mlm"              
#>   [3] "Arima"                   "BBmm"                   
#>   [5] "BBreg"                   "BFBayesFactor"          
#>   [7] "BGGM"                    "DirichletRegModel"      
#>   [9] "Gam"                     "Glm"                    
#>  [11] "HLfit"                   "LORgee"                 
#>  [13] "MANOVA"                  "MCMCglmm"               
#>  [15] "MixMod"                  "PMCMR"                  
#>  [17] "RM"                      "Rchoice"                
#>  [19] "Sarlm"                   "SemiParBIV"             
#>  [21] "aareg"                   "afex_aov"               
#>  [23] "aov"                     "aovlist"                
#>  [25] "averaging"               "bamlss"                 
#>  [27] "bamlss.frame"            "bayesQR"                
#>  [29] "bayesx"                  "bcplm"                  
#>  [31] "betamfx"                 "betaor"                 
#>  [33] "betareg"                 "bfsl"                   
#>  [35] "bife"                    "bifeAPEs"               
#>  [37] "bigglm"                  "biglm"                  
#>  [39] "blavaan"                 "blrm"                   
#>  [41] "bracl"                   "brglm"                  
#>  [43] "brmsfit"                 "brmultinom"             
#>  [45] "btergm"                  "censReg"                
#>  [47] "cgam"                    "cgamm"                  
#>  [49] "cglm"                    "clm"                    
#>  [51] "clm2"                    "clmm"                   
#>  [53] "clmm2"                   "clogit"                 
#>  [55] "coeftest"                "complmrob"              
#>  [57] "confusionMatrix"         "coxme"                  
#>  [59] "coxph"                   "coxph.penal"            
#>  [61] "coxr"                    "cpglm"                  
#>  [63] "cpglmm"                  "crch"                   
#>  [65] "crq"                     "crqs"                   
#>  [67] "crr"                     "dep.effect"             
#>  [69] "drc"                     "eglm"                   
#>  [71] "elm"                     "epi.2by2"               
#>  [73] "ergm"                    "feglm"                  
#>  [75] "feis"                    "felm"                   
#>  [77] "fitdistr"                "fixest"                 
#>  [79] "flexsurvreg"             "gam"                    
#>  [81] "gamlss"                  "gamm"                   
#>  [83] "gamm4"                   "garch"                  
#>  [85] "gbm"                     "gee"                    
#>  [87] "geeglm"                  "glht"                   
#>  [89] "glimML"                  "glm"                    
#>  [91] "glmRob"                  "glmm"                   
#>  [93] "glmmPQL"                 "glmmTMB"                
#>  [95] "glmmadmb"                "glmrob"                 
#>  [97] "glmx"                    "gls"                    
#>  [99] "gmnl"                    "htest"                  
#> [101] "hurdle"                  "ivFixed"                
#> [103] "iv_robust"               "ivprobit"               
#> [105] "ivreg"                   "lavaan"                 
#> [107] "lm"                      "lmRob"                  
#> [109] "lm_robust"               "lme"                    
#> [111] "lmerMod"                 "lmerModLmerTest"        
#> [113] "lmodel2"                 "lmrob"                  
#> [115] "logistf"                 "logitmfx"               
#> [117] "logitor"                 "lqm"                    
#> [119] "lqmm"                    "lrm"                    
#> [121] "manova"                  "marginaleffects"        
#> [123] "marginaleffects.summary" "margins"                
#> [125] "maxLik"                  "mclogit"                
#> [127] "mcmc"                    "mcmc.list"              
#> [129] "mcp1"                    "mcp12"                  
#> [131] "mcp2"                    "med1way"                
#> [133] "mediate"                 "merMod"                 
#> [135] "merModList"              "meta_bma"               
#> [137] "meta_fixed"              "meta_random"            
#> [139] "metaplus"                "mhurdle"                
#> [141] "mipo"                    "mira"                   
#> [143] "mixed"                   "mixor"                  
#> [145] "mjoint"                  "mle"                    
#> [147] "mle2"                    "mlm"                    
#> [149] "mlogit"                  "mmlogit"                
#> [151] "model_fit"               "multinom"               
#> [153] "mvord"                   "negbinirr"              
#> [155] "negbinmfx"               "ols"                    
#> [157] "onesampb"                "orm"                    
#> [159] "pgmm"                    "plm"                    
#> [161] "poissonirr"              "poissonmfx"             
#> [163] "polr"                    "probitmfx"              
#> [165] "psm"                     "ridgelm"                
#> [167] "riskRegression"          "rjags"                  
#> [169] "rlm"                     "rlmerMod"               
#> [171] "rma"                     "rma.uni"                
#> [173] "robmixglm"               "robtab"                 
#> [175] "rq"                      "rqs"                    
#> [177] "rqss"                    "scam"                   
#> [179] "selection"               "sem"                    
#> [181] "semLm"                   "semLme"                 
#> [183] "slm"                     "speedglm"               
#> [185] "speedlm"                 "stanfit"                
#> [187] "stanmvreg"               "stanreg"                
#> [189] "summary.lm"              "survfit"                
#> [191] "survreg"                 "svy_vglm"               
#> [193] "svychisq"                "svyglm"                 
#> [195] "svyolr"                  "t1way"                  
#> [197] "tobit"                   "trimcibt"               
#> [199] "truncreg"                "vgam"                   
#> [201] "vglm"                    "wbgee"                  
#> [203] "wblm"                    "wbm"                    
#> [205] "wmcpAKP"                 "yuen"                   
#> [207] "yuend"                   "zcpglm"                 
#> [209] "zeroinfl"                "zerotrunc"