Checks if a regression model object is supported in insight package

Small helper that checks if a model is a supported (regression) model object. supported_models() prints a list of currently supported model classes.

Usage

is_model_supported(x)

supported_models()

Arguments

x

An object.

Value

A logical, TRUE if x is a (supported) model object.

Details

This function returns TRUE if x is a model object that works with the package's functions. A list of supported models can also be found here: https://github.com/easystats/insight.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)

is_model_supported(m)
#> [1] TRUE
is_model_supported(mtcars)
#> [1] FALSE

# to see all supported models
supported_models()
#>   [1] "AKP"                     "Anova.mlm"              
#>   [3] "Arima"                   "BBmm"                   
#>   [5] "BBreg"                   "BFBayesFactor"          
#>   [7] "BGGM"                    "DirichletRegModel"      
#>   [9] "Gam"                     "Glm"                    
#>  [11] "HLfit"                   "LORgee"                 
#>  [13] "MANOVA"                  "MCMCglmm"               
#>  [15] "MixMod"                  "PMCMR"                  
#>  [17] "RM"                      "Rchoice"                
#>  [19] "Sarlm"                   "SemiParBIV"             
#>  [21] "aareg"                   "afex_aov"               
#>  [23] "anova.rms"               "aov"                    
#>  [25] "aovlist"                 "averaging"              
#>  [27] "bamlss"                  "bamlss.frame"           
#>  [29] "bayesQR"                 "bayesx"                 
#>  [31] "bcplm"                   "betamfx"                
#>  [33] "betaor"                  "betareg"                
#>  [35] "bfsl"                    "bife"                   
#>  [37] "bifeAPEs"                "bigglm"                 
#>  [39] "biglm"                   "blavaan"                
#>  [41] "blrm"                    "bracl"                  
#>  [43] "brglm"                   "brmsfit"                
#>  [45] "brmultinom"              "btergm"                 
#>  [47] "censReg"                 "cgam"                   
#>  [49] "cgamm"                   "cglm"                   
#>  [51] "clm"                     "clm2"                   
#>  [53] "clmm"                    "clmm2"                  
#>  [55] "clogit"                  "coeftest"               
#>  [57] "complmrob"               "confusionMatrix"        
#>  [59] "coxme"                   "coxph"                  
#>  [61] "coxph.penal"             "coxr"                   
#>  [63] "cpglm"                   "cpglmm"                 
#>  [65] "crch"                    "crq"                    
#>  [67] "crqs"                    "crr"                    
#>  [69] "dep.effect"              "draws"                  
#>  [71] "drc"                     "eglm"                   
#>  [73] "elm"                     "epi.2by2"               
#>  [75] "ergm"                    "feglm"                  
#>  [77] "feis"                    "felm"                   
#>  [79] "fitdistr"                "fixest"                 
#>  [81] "flexsurvreg"             "gam"                    
#>  [83] "gamlss"                  "gamm"                   
#>  [85] "gamm4"                   "garch"                  
#>  [87] "gbm"                     "gee"                    
#>  [89] "geeglm"                  "glht"                   
#>  [91] "glimML"                  "glm"                    
#>  [93] "glmRob"                  "glmm"                   
#>  [95] "glmmPQL"                 "glmmTMB"                
#>  [97] "glmmadmb"                "glmrob"                 
#>  [99] "glmx"                    "gls"                    
#> [101] "gmnl"                    "htest"                  
#> [103] "hurdle"                  "ivFixed"                
#> [105] "iv_robust"               "ivprobit"               
#> [107] "ivreg"                   "lavaan"                 
#> [109] "lm"                      "lmRob"                  
#> [111] "lm_robust"               "lme"                    
#> [113] "lmerMod"                 "lmerModLmerTest"        
#> [115] "lmodel2"                 "lmrob"                  
#> [117] "logistf"                 "logitmfx"               
#> [119] "logitor"                 "lqm"                    
#> [121] "lqmm"                    "lrm"                    
#> [123] "manova"                  "marginaleffects"        
#> [125] "marginaleffects.summary" "margins"                
#> [127] "maxLik"                  "mblogit"                
#> [129] "mclogit"                 "mcmc"                   
#> [131] "mcmc.list"               "mcp1"                   
#> [133] "mcp12"                   "mcp2"                   
#> [135] "med1way"                 "mediate"                
#> [137] "merMod"                  "merModList"             
#> [139] "meta_bma"                "meta_fixed"             
#> [141] "meta_random"             "metaplus"               
#> [143] "mhurdle"                 "mipo"                   
#> [145] "mira"                    "mixed"                  
#> [147] "mixor"                   "mjoint"                 
#> [149] "mle"                     "mle2"                   
#> [151] "mlm"                     "mlogit"                 
#> [153] "mmclogit"                "mmlogit"                
#> [155] "model_fit"               "multinom"               
#> [157] "mvord"                   "negbinirr"              
#> [159] "negbinmfx"               "ols"                    
#> [161] "onesampb"                "orm"                    
#> [163] "pgmm"                    "plm"                    
#> [165] "poissonirr"              "poissonmfx"             
#> [167] "polr"                    "probitmfx"              
#> [169] "psm"                     "ridgelm"                
#> [171] "riskRegression"          "rjags"                  
#> [173] "rlm"                     "rlmerMod"               
#> [175] "rma"                     "rma.uni"                
#> [177] "robmixglm"               "robtab"                 
#> [179] "rq"                      "rqs"                    
#> [181] "rqss"                    "rvar"                   
#> [183] "scam"                    "selection"              
#> [185] "sem"                     "semLm"                  
#> [187] "semLme"                  "slm"                    
#> [189] "speedglm"                "speedlm"                
#> [191] "stanfit"                 "stanmvreg"              
#> [193] "stanreg"                 "summary.lm"             
#> [195] "survfit"                 "survreg"                
#> [197] "svy_vglm"                "svychisq"               
#> [199] "svyglm"                  "svyolr"                 
#> [201] "t1way"                   "tobit"                  
#> [203] "trimcibt"                "truncreg"               
#> [205] "vgam"                    "vglm"                   
#> [207] "wbgee"                   "wblm"                   
#> [209] "wbm"                     "wmcpAKP"                
#> [211] "yuen"                    "yuend"                  
#> [213] "zcpglm"                  "zeroinfl"               
#> [215] "zerotrunc"