Checks if a regression model object is supported in insight package

Small helper that checks if a model is a supported (regression) model object. supported_models() prints a list of currently supported model classes.

Usage

is_model_supported(x)

supported_models()

Arguments

x

An object.

Value

A logical, TRUE if x is a (supported) model object.

Details

This function returns TRUE if x is a model object that works with the package's functions. A list of supported models can also be found here: https://github.com/easystats/insight.

Examples

data(mtcars)
m <- lm(mpg ~ wt + cyl + vs, data = mtcars)

is_model_supported(m)
#> [1] TRUE
is_model_supported(mtcars)
#> [1] FALSE

# to see all supported models
supported_models()
#>   [1] "AKP"               "Anova.mlm"         "Arima"            
#>   [4] "BBmm"              "BBreg"             "BFBayesFactor"    
#>   [7] "BGGM"              "DirichletRegModel" "Gam"              
#>  [10] "Glm"               "HLfit"             "LORgee"           
#>  [13] "MANOVA"            "MCMCglmm"          "MixMod"           
#>  [16] "PMCMR"             "RM"                "Rchoice"          
#>  [19] "Sarlm"             "SemiParBIV"        "aareg"            
#>  [22] "afex_aov"          "aov"               "aovlist"          
#>  [25] "averaging"         "bamlss"            "bamlss.frame"     
#>  [28] "bayesQR"           "bayesx"            "bcplm"            
#>  [31] "betamfx"           "betaor"            "betareg"          
#>  [34] "bfsl"              "bife"              "bifeAPEs"         
#>  [37] "bigglm"            "biglm"             "blavaan"          
#>  [40] "blrm"              "bracl"             "brglm"            
#>  [43] "brmsfit"           "brmultinom"        "btergm"           
#>  [46] "censReg"           "cgam"              "cgamm"            
#>  [49] "cglm"              "clm"               "clm2"             
#>  [52] "clmm"              "clmm2"             "clogit"           
#>  [55] "coeftest"          "complmrob"         "confusionMatrix"  
#>  [58] "coxme"             "coxph"             "coxph.penal"      
#>  [61] "coxr"              "cpglm"             "cpglmm"           
#>  [64] "crch"              "crq"               "crqs"             
#>  [67] "crr"               "dep.effect"        "drc"              
#>  [70] "eglm"              "elm"               "epi.2by2"         
#>  [73] "ergm"              "feglm"             "feis"             
#>  [76] "felm"              "fitdistr"          "fixest"           
#>  [79] "flexsurvreg"       "gam"               "gamlss"           
#>  [82] "gamm"              "gamm4"             "garch"            
#>  [85] "gbm"               "gee"               "geeglm"           
#>  [88] "glht"              "glimML"            "glm"              
#>  [91] "glmRob"            "glmm"              "glmmPQL"          
#>  [94] "glmmTMB"           "glmmadmb"          "glmrob"           
#>  [97] "glmx"              "gls"               "gmnl"             
#> [100] "htest"             "hurdle"            "ivFixed"          
#> [103] "iv_robust"         "ivprobit"          "ivreg"            
#> [106] "lavaan"            "lm"                "lmRob"            
#> [109] "lm_robust"         "lme"               "lmerMod"          
#> [112] "lmerModLmerTest"   "lmodel2"           "lmrob"            
#> [115] "logistf"           "logitmfx"          "logitor"          
#> [118] "lqm"               "lqmm"              "lrm"              
#> [121] "manova"            "margins"           "maxLik"           
#> [124] "mclogit"           "mcmc"              "mcmc.list"        
#> [127] "mcp1"              "mcp12"             "mcp2"             
#> [130] "med1way"           "mediate"           "merMod"           
#> [133] "merModList"        "meta_bma"          "meta_fixed"       
#> [136] "meta_random"       "metaplus"          "mhurdle"          
#> [139] "mipo"              "mira"              "mixed"            
#> [142] "mixor"             "mjoint"            "mle"              
#> [145] "mle2"              "mlm"               "mlogit"           
#> [148] "mmlogit"           "model_fit"         "multinom"         
#> [151] "mvord"             "negbinirr"         "negbinmfx"        
#> [154] "ols"               "onesampb"          "orm"              
#> [157] "pgmm"              "plm"               "poissonirr"       
#> [160] "poissonmfx"        "polr"              "probitmfx"        
#> [163] "psm"               "ridgelm"           "riskRegression"   
#> [166] "rjags"             "rlm"               "rlmerMod"         
#> [169] "rma"               "rma.uni"           "robmixglm"        
#> [172] "robtab"            "rq"                "rqs"              
#> [175] "rqss"              "scam"              "selection"        
#> [178] "sem"               "semLm"             "semLme"           
#> [181] "slm"               "speedglm"          "speedlm"          
#> [184] "stanfit"           "stanmvreg"         "stanreg"          
#> [187] "summary.lm"        "survfit"           "survreg"          
#> [190] "svy_vglm"          "svychisq"          "svyglm"           
#> [193] "svyolr"            "t1way"             "tobit"            
#> [196] "trimcibt"          "truncreg"          "vgam"             
#> [199] "vglm"              "wbgee"             "wblm"             
#> [202] "wbm"               "wmcpAKP"           "yuen"             
#> [205] "yuend"             "zcpglm"            "zeroinfl"         
#> [208] "zerotrunc"