Access information from model objects

Retrieve information from model objects.

model_info(x, ...)

Arguments

x
A fitted model.

...
Currently not used.

Value

A list with information about the model, like family, link-function etc. (see 'Details').

Details

model_info() returns a list with information about the model for many different model objects. Following information is returned, where all values starting with is_ are logicals.

is_binomial: family is binomial (but not negative binomial)

is_poisson: family is poisson

is_negbin: family is negative binomial

is_count: model is a count model (i.e. family is either poisson or negative binomial)

is_beta: family is beta

is_logit: model has logit link

is_linear: family is gaussian

is_tweedie: family is tweedie

is_ordinal: family is ordinal or cumulative link

is_categorical: family is categorical link

is_zeroinf: model has zero-inflation component

is_zero_inflated: alias for is_zeroinf

is_mixed: model is a mixed effects model (with random effects)

is_multivariate: model is a multivariate response model (currently only works for brmsfit objects)

is_trial: model response contains additional information about the trials

is_bayesian: model is a Bayesian model

is_anova: model is an Anova object

link_function: the link-function

family: the family-object

n_obs: number of observations

model_terms: a list with all model terms, including terms such as random effects or from zero-inflated model parts.

Examples

library(glmmTMB)

#> Warning: Package version inconsistency detected.
#> TMB was built with Matrix version 1.2.16
#> Current Matrix version is 1.2.17
#> Please re-install 'TMB' from source using install.packages('TMB', type = 'source') or ask CRAN for a binary version of 'TMB' matching CRAN's 'Matrix' package

data("Salamanders")
m <- glmmTMB(
  count ~ spp + cover + mined + (1 | site),
  ziformula = ~ spp + mined,
  dispformula = ~DOY,
  data = Salamanders,
  family = nbinom2
)

model_info(m)

#> $is_binomial
#> [1] FALSE
#> 
#> $is_count
#> [1] TRUE
#> 
#> $is_poisson
#> [1] FALSE
#> 
#> $is_negbin
#> [1] TRUE
#> 
#> $is_beta
#> [1] FALSE
#> 
#> $is_logit
#> [1] FALSE
#> 
#> $is_probit
#> [1] FALSE
#> 
#> $is_linear
#> [1] FALSE
#> 
#> $is_tweedie
#> [1] FALSE
#> 
#> $is_zeroinf
#> [1] TRUE
#> 
#> $is_zero_inflated
#> [1] TRUE
#> 
#> $is_ordinal
#> [1] FALSE
#> 
#> $is_categorical
#> [1] FALSE
#> 
#> $is_mixed
#> [1] TRUE
#> 
#> $is_multivariate
#> [1] FALSE
#> 
#> $is_trial
#> [1] FALSE
#> 
#> $is_bayesian
#> [1] FALSE
#> 
#> $is_anova
#> [1] FALSE
#> 
#> $is_ttest
#> [1] FALSE
#> 
#> $is_correlation
#> [1] FALSE
#> 
#> $link_function
#> [1] "log"
#> 
#> $family
#> [1] "nbinom2"
#> 
#> $n_obs
#> [1] 644
#> 
#> $model_terms
#> $model_terms$response
#> [1] "count"
#> 
#> $model_terms$conditional
#> [1] "spp"   "cover" "mined"
#> 
#> $model_terms$random
#> [1] "site"
#> 
#> $model_terms$zero_inflated
#> [1] "spp"   "mined"
#> 
#> $model_terms$dispersion
#> [1] "DOY"
#> 
#>