trim_ws {insight}	R Documentation

Small helper functions

Description

Collection of small helper functions. trim_ws() is an efficient function to trim leading and trailing whitespaces from character vectors or strings. n_unique() returns the number of unique values in a vector. safe_deparse() is comparable to deparse1(), i.e. it can safely deparse very long expressions into a single string.

Usage

trim_ws(x, ...)

## S3 method for class 'data.frame'
trim_ws(x, character_only = TRUE, ...)

n_unique(x, ...)

## Default S3 method:
n_unique(x, na.rm = TRUE, ...)

safe_deparse(x, ...)

Arguments

x
A (character) vector, or for some functions may also be a data frame.

...
Currently not used.

character_only
Logical, if TRUE and x is a data frame or list, only processes character vectors.

na.rm
Logical, if missing values should be removed from the input.

Value

For a vector, n_unique always returns an integer value, even if the input is NULL (the return value will be 0 then). For data frames or lists, n_unique() returns a named numeric vector, with the number of unique values for each element.

Examples

trim_ws("  no space!  ")
n_unique(iris$Species)

[Package insight version 0.18.0 Index]